% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_stics_versions.R
\name{add_stics_version}
\alias{add_stics_version}
\title{Adding a new version in the SticsRFiles library or package}
\usage{
add_stics_version(
  version_name,
  url,
  file_name = "all",
  location = "install",
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{version_name}{name (i.e. "VX.Y") of the version to set
in the csv file containing informations about versions}

\item{url}{Subversion repository address of the branch, tag to get
information from}

\item{file_name}{File name(s)
or keyword "all" for both files (default)}

\item{location}{The destination where to write information
"install" for writing
things in the installed SticsRFiles library (default), "package" for writing
them in the package project (in RStudio)}

\item{overwrite}{A logical, TRUE to overwrite csv files,
FALSE otherwise (default)}

\item{verbose}{Logical, TRUE for displaying warnings (default),
FALSE otherwise}
}
\value{
An invisible data.frame containing versions data
}
\description{
Adding a new version in the SticsRFiles library or package
}
\examples{
\dontrun{
SticsRFiles:::add_stics_version(
  version_name = "V10.0",
  url = "https://w3.avignon.inra.fr/svn/modulostics/branches/branch10"
)

SticsRFiles:::add_stics_version(
  version_name = "V10.0",
  url = "https://w3.avignon.inra.fr/svn/modulostics/branches/branch10",
  location = "package"
)
}
}
\keyword{internal}
