% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_out_var.R
\name{all_out_var}
\alias{all_out_var}
\title{Return all possible STICS outputs for var.mod}
\usage{
all_out_var(stics_version = "latest")
}
\arguments{
\item{stics_version}{Name of the Stics version. Optional, can be used to search
parameters information relative to a specific Stics version. By default the
latest version returned by \code{get_stics_versions_compat()} is used.}
}
\description{
Helper function to print the list of all possible variables to set as output
from the STICS model.
}
\examples{
\dontrun{
all_out_var()
}
}
\seealso{
\code{get_var_info()}, \code{gen_varmod()}, and \code{get_stics_versions_compat()}
}
\keyword{internal}
