% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes_list2matrix.R
\name{attributes_list2matrix}
\alias{attributes_list2matrix}
\title{Converting attributes list to matrix}
\usage{
attributes_list2matrix(attr_list)
}
\arguments{
\item{attr_list}{XML attributes list}
}
\value{
A matrix with attributes names as column names
}
\description{
Converting attributes list to matrix
}
\keyword{internal}
