% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_txt.R
\name{character_to_numeric_list}
\alias{character_to_numeric_list}
\title{Character list to numeric}
\usage{
character_to_numeric_list(x)
}
\arguments{
\item{z}{A list with potential numeric values written a characters}
}
\value{
A list with numeric values when possible
}
\description{
Tries to convert the values in a list into numeric values,
and if it fails, return as character.
}
\examples{
\dontrun{
test <- list(a = "2", b = "toto")
character_to_numeric_list(test)
}

}
\keyword{internal}
