% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_choice_param.R
\name{check_choice_param}
\alias{check_choice_param}
\title{Checking if any parameter appear in 2 option choices}
\usage{
check_choice_param(xml_doc, param_name, stop = FALSE)
}
\arguments{
\item{xml_doc}{an xmldocument object}

\item{param_name}{parameter names vector, i.e.: parameter name or option code}

\item{stop}{TRUE for rising an error, FALSE for just warning}
}
\value{
invisible NULL
}
\description{
Checking if any parameter appear in 2 option choices
}
\examples{
\dontrun{
  check_choice_param( xml_doc = xml_doc, param_name = param_name)
}

}
\keyword{internal}
