% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javastics_cmd_util.R
\name{check_java_version}
\alias{check_java_version}
\title{Checking if the java virtual machine is compatible with the given
JavaStics command executable}
\usage{
check_java_version(javastics_cmd = "JavaSticsCmd.exe", java_cmd = "java")
}
\arguments{
\item{javastics_cmd}{JavaSTICS command line executable name}

\item{java_cmd}{Name or path of the java virtual machine executable}
}
\description{
For Unix like systems, the system java virtual machine is
detected and consistency for JavaStics executable used is checked
}
\details{
Errors are raised if the JavaStics executable cannot be run with the
given java virtual machine.
}
\examples{
\dontrun{
 check_java_version(javastics_cmd = "JavaStics.exe")
}

}
\keyword{internal}
