% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_java_workspace.R
\name{check_java_workspace}
\alias{check_java_workspace}
\title{Checking and getting JavaStics workspace path}
\usage{
check_java_workspace(javastics, workspace = NULL)
}
\arguments{
\item{javastics}{Path of JavaSTICS installation directory}

\item{workspace}{An absolute or relative path (to JavaSTICS path)
of a workspace (Optional)}
}
\value{
An absolute JavaSTICS workspace path
}
\description{
Looking in JavaSTICS for a workspace path in JavaSTICS
preferences or producing a full path to a workspace located in JavaSTICS root
directory or validating an external absolute path. And also checking if the
path is a valid JavaStics workspace
}
\keyword{internal}
