% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javastics_cmd_util.R
\name{check_javastics_cmd}
\alias{check_javastics_cmd}
\title{Checking if the java virtual machine is compatible with the given
JavaSTICS command executable, and command line options}
\usage{
check_javastics_cmd(
  javastics_cmd = "JavaSticsCmd.exe",
  java_cmd = "java",
  verbose = TRUE
)
}
\arguments{
\item{javastics_cmd}{path to JavaSTICS command line executable}

\item{java_cmd}{Name or path of the java virtual machine executable}
}
\value{
A named list with the JavaSTICS command line path,
the java path and the verbose string option (according to JavaSTICS command
line version, and the "verbose" argument value)
}
\description{
For Unix like systems, the system java virtual machine is
detected and consistency for JavaSTICS executable used is checked.
Command line verbosity option is detected for JavaSTICS command line for all
systems.
}
\details{
Errors are raised if the JavaSTICS executable cannot be run with the
given java virtual machine for unix like systems.
}
\examples{
\dontrun{
check_javastics_cmd(javastics_cmd = "/path/to/JavaSticsCmd.exe")
check_javastics_cmd(
  javastics_cmd = "/path/to/JavaSticsCmd.exe",
  java_cmd = "/path/to/java"
)
}

}
\keyword{internal}
