% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_usms_files.R
\name{check_usms_files}
\alias{check_usms_files}
\title{Checking xml files existence got from usms.xml}
\usage{
check_usms_files(
  workspace_path,
  javastics_path,
  usms_list = NULL,
  file_name = "usms.xml"
)
}
\arguments{
\item{workspace_path}{The path to a JavaStics-compatible workspace}

\item{javastics_path}{JavaStics installation path (Optional, needed if the
plant files are not in the \code{workspace_path}
but rather in the JavaStics default plant folder)}

\item{usms_list}{Vector of usms names (Optional)}

\item{file_name}{Usms XML file name (Optional)}
}
\value{
A data.frame with files names usm row and a status (all_exist)
indicating weither or not all listed files exist in the \code{workspace_path}
}
\description{
Checking xml files existence got from usms.xml
}
\examples{
\dontrun{

check_usms_files(
  workspace_path = "/path/to/workspace",
  javastics_path = "/path/to/JavaSTICS/folder"
)

check_usms_files(
  workspace_path = "/path/to/workspace",
  javastics_path = "/path/to/JavaSTICS/folder", usm_list = c("usm1", "usm3")
)
}

}
\keyword{internal}
