% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stics_versions_compat.R
\name{checkversion}
\alias{checkversion}
\title{Checking the validity of a given version code}
\usage{
checkversion(version_name = "last")
}
\arguments{
\item{version}{An optional version name as listed in get_stics_versions_compat() return}
}
\value{
A valid version string
}
\description{
Checking the validity of a given version code
}
\examples{
\dontrun{

checkversion()
[1] "V9.1"

checkversion(version = "V8.5")
[1] "V8.5"

}
}
