% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_names_to_var.R
\name{col_names_to_var}
\alias{col_names_to_var}
\title{Convert columns names list to Stics variables names}
\usage{
col_names_to_var(var_list = c())
}
\arguments{
\item{var_list}{Columns names list}
}
\value{
List of Stics variables names
}
\description{
Change the \code{varname.n.} or \code{varname_n} by \code{varname(n)}
other names are unchanged (i.e.: varname, varname(n), ...)
}
