% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_climate.R
\name{complete_climate_paths}
\alias{complete_climate_paths}
\title{Generating the complete wheather files paths list}
\usage{
complete_climate_paths(files_path)
}
\arguments{
\item{files_path}{Vector of 1 or 2 weather data files}
}
\value{
a character vector of files paths
}
\description{
Generating the complete wheather files paths list
}
\examples{
\dontrun{
SticsRFiles:::complete_climate_paths(
            c("path/to/weather.year1", "path/to/weather.year2"))
}

}
\keyword{internal}
