% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_date_from_day.R
\name{compute_date_from_day}
\alias{compute_date_from_day}
\title{Convert day number into date}
\usage{
compute_date_from_day(day, start_year)
}
\arguments{
\item{day}{day number(s) to be converted}

\item{start_year}{year to be used as time reference (simulation start year).}
}
\value{
Date vector
}
\description{
Computes the date corresponding to a given day number (or vector of)
with reference to a start year. Typically, the start year should
be the year of a STICS simulation start, but not necessarily.
}
\examples{

compute_date_from_day(day = 290, start_year = 1994)

compute_date_from_day(day = 700, start_year = 1994)

}
\author{
Timothee Flutre
}
