% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_xml2txt.R
\name{convert_xml2txt}
\alias{convert_xml2txt}
\title{Transforming a Stics xml file into a text file}
\usage{
convert_xml2txt(
  xml_file,
  java_dir,
  plt_num = 1,
  out_dir = NULL,
  out_file = NULL
)
}
\arguments{
\item{xml_file}{File path of the input xml file}

\item{java_dir}{Root directory of the JavaStics installation}

\item{plt_num}{The plant identifier (main crop: 1 ; associated crop: 2)}

\item{out_dir}{Output directory of the generated file (optional, default:
path of the input xml file)}

\item{out_file}{Name of the output file (optional, default: fixed name for Stics)}
}
\description{
The input file according to his type (ini,plant,tec,station,soil,par)
is converted to a text file readable by the Stics model (ficini.txt, ficplt1.txt,...)
}
\examples{
\dontrun{
xml_path <- "/path/to/corn_plt.xml"
javastics_path <- "/path/to/javastics/root"
convert_xml2txt(xml_file = xml_path, java_dir = javastics_path)
}

}
