% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_xml2txt_int.R
\name{convert_xml2txt_int}
\alias{convert_xml2txt_int}
\title{Transforming an xml file with the help of an xsl style file}
\usage{
convert_xml2txt_int(xml_file, style_file, out_file = NULL)
}
\arguments{
\item{xml_file}{File path of the input xml file}

\item{style_file}{File path of the xsl file}

\item{out_file}{File path of the generated file}
}
\value{
Path of the generated file
}
\description{
Using an xslt tranformation, an xml input file is converted to
an other file format (text, xml). The output file name may be given or calculated.
In the latest case, the extension is defined after getting xsl method from the xsl file (xml, txt).
}
\examples{
\dontrun{
xml_plt <- file.path(get_examples_path(file_type = "xml"), "file_plt.xml")
xsl_file <- "/path/to/JavaSTICS/folder/bin/resources/xml/stylesheet/xml2txt.xsl"

SticsRFiles:::convert_xml2txt_int(xml_file = xml_plt, style_file = xsl_file)
}

}
\keyword{internal}
