% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_param_xml.R
\name{exist_param_csv}
\alias{exist_param_csv}
\title{Checking existing parameters names in csv input files}
\usage{
exist_param_csv(param, javastics)
}
\arguments{
\item{param}{Parameter name or a vector of}

\item{javastics}{Path of JavaSTICS}
}
\value{
A named logical vector of existing parameters, with
real parameters names
}
\description{
Checking existing parameters names in csv input files
}
\examples{
\dontrun{

exist_param_csv(
  param = "albedo",
  javastics = "/path/to/JavaSTICS/folder"
)

exist_param_csv(
  param = c("albedo", "latitude"),
  javastics = "/path/to/JavaSTICS/folder"
)
}

}
\keyword{internal}
