% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exist_param_xml.R
\name{exist_param_xml}
\alias{exist_param_xml}
\title{Checking existing parameters names in XML files}
\usage{
exist_param_xml(param, stics_version = NULL)
}
\arguments{
\item{param}{Parameter name or a vector of}

\item{stics_version}{Optional, Stics version.
Only the 2 latest are referenced: V9.0, V9.1 (default value)}
}
\value{
A logical vector of existing parameters
}
\description{
Checking existing parameters names in XML files
}
\examples{
\dontrun{

exist_param_xml(param = "albedo")

exist_param_xml(param = "albedo", stics_version = "V9.0")

exist_param_xml(param = c("albedo", "latitude", "humcapil"))

exist_param_xml(param = c("albedo", "latitude", "humcapi"))
}

}
\keyword{internal}
