% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javastics_path.R
\name{exists_javastics_path}
\alias{exists_javastics_path}
\title{Evaluates if a \code{javastics_path} environment variable exists}
\usage{
exists_javastics_path()
}
\value{
boolean, TRUE if the variable exists, FALSE otherwise.
A \code{path} attribute is added to the object \code{path} containing
the variable content.
}
\description{
Evaluates if a \code{javastics_path} environment variable exists
}
\keyword{internal}
