% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists_param.R
\name{exists_param}
\alias{exists_param}
\title{Test if a parameter exists}
\usage{
exists_param(xml_doc_object, param_name, ...)
}
\arguments{
\item{xml_doc_object}{XML document}

\item{param_name}{The parameter name}

\item{...}{Other parameters to pass to \code{\link[=get_param_number]{get_param_number()}}.}
}
\description{
Test if a parameter exists in an XML file
}
