% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_stics_names.R
\name{expand_stics_names}
\alias{expand_stics_names}
\title{Expand stics names from parameters/variables simple names}
\usage{
expand_stics_names(in_csv_file, out_csv_file, header_vec = NULL)
}
\arguments{
\item{in_csv_file}{Input csv file path}

\item{out_csv_file}{Output csv file path}

\item{header_vec}{An optional header vector}
}
\description{
Expand stics names from parameters/variables simple names
}
\keyword{internal}
