% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_workspace.R
\name{extract_workspace}
\alias{extract_workspace}
\title{Extract an XML JavaSTICS workspace from a source workspace}
\usage{
extract_workspace(
  from_workspace,
  to_workspace,
  javastics,
  usm = NULL,
  use_mod_files = FALSE
)
}
\arguments{
\item{from_workspace}{Source workspace path}

\item{to_workspace}{Target workspace path}

\item{javastics}{JavaStics directory path, needed for getting
plant files which are stored in JavaSTICS \code{plant} directory, or general
parameters files and *.mod files from the \code{config} directory}

\item{usm}{vector of situation names (i.e. USMs to extract.)}

\item{use_mod_files}{Logical: if TRUE, *.mod files are searched in the
from_workspace and copyed to the to_workspace if found. Default is FALSE.}
}
\value{
None
}
\description{
Extract an XML JavaSTICS workspace from a source workspace
}
\examples{
\dontrun{
# Extract all usms from a source workspace to a target workspace
extract_workspace("/path/to/source/workspace",
                  "/path/to/target/workspace",
                  "/path/to/javastics/dir")

# Extract only selected usms from a source workspace to a target workspace
extract_workspace("/path/to/source/workspace",
                  "/path/to/target/workspace",
                  "/path/to/javastics/dir",
                  usm = c("SugarCane", "potato"))

}
}
