% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_names.R
\name{find_names}
\alias{find_names}
\title{Finding strings in a character vector using exact or partial search}
\usage{
find_names(names, name, exact = FALSE)
}
\arguments{
\item{names}{Character vector}

\item{name}{Character vector to use as partial or exact search in names}

\item{exact}{logical, TRUE to find exact names, FALSE otherwise (default)}
}
\value{
Character vector of founds strings
}
\description{
Finding strings in a character vector using exact or partial search
}
\examples{

names <- c("abscission", "alphaphot", "codelaitr", "dlaimax", "lai")

SticsRFiles:::find_names(names, "lai")

#> [1] "codelaitr" "dlaimax"

SticsRFiles:::find_names(names, "lai", exact = TRUE)

#> [1] "lai"
}
\keyword{internal}
