% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_param.R
\name{find_param}
\alias{find_param}
\title{Finding parameters names or formalism using partial search words}
\usage{
find_param(name, version = NULL, kind = "parameter", fixed = FALSE)
}
\arguments{
\item{name}{Name or partial name or a vector of}

\item{version}{Optional, Stics version.
Only the 2 last are referenced: V9.0, V9.1 (default value)}

\item{kind}{Kind of information to be retrieved for parameters
among "paramete", "formalism"}

\item{fixed}{Logical, if TRUE, the exact name is searched}
}
\value{
A data.frame containing parameters names,
their origin (file name) and their bounds or a named list
of files containing formalisms and attached parameters names.
}
\description{
Finding parameters names or formalism using partial search words
}
\examples{
\dontrun{

find_param(name = "albedo")

find_param(name = "albedo", kind = "formalism)

find_param(name = "albedo", version = "V9.0")

find_param(name = c("albedo", "latitude", "humcapil"))

find_param(name = c("albedo", "latitude", "humcapil"),
kind = "formalism)

}


}
\keyword{internal}
