% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_param_formalisms.R
\name{find_param_formalisms}
\alias{find_param_formalisms}
\title{Finding parameters names using partial search words}
\usage{
find_param_formalisms(name, version = NULL)
}
\arguments{
\item{name}{Name or partial name or a vector of}

\item{version}{Optional, Stics version.
Only the 2 last are referenced: V9.0, V9.1 (default value)}
}
\value{
A named list of files containing formalisms
and attached parameters names
}
\description{
Finding parameters names using partial search words
}
\examples{
\dontrun{

find_param_formalisms(name = "albedo")

find_param_formalisms(name = "albedo", version = "V9.0")

find_param_formalisms(name = c("albedo", "latitude", "humcapil"))

find_param_formalisms(name = c("al", "hum"))

}


}
