% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_param_names.R
\name{find_param_names}
\alias{find_param_names}
\title{Finding parameters names using partial search words}
\usage{
find_param_names(name, version = NULL, exact = FALSE)
}
\arguments{
\item{name}{Name or partial name or a vector of}

\item{version}{Optional, Stics version.
Only the 2 last are referenced: V9.0, V9.1 (default value)}

\item{exact}{Logical, if TRUE, the exact name is searched}
}
\value{
A data.frame containing parameters names,
their origin (file name) and their bounds.
}
\description{
Finding parameters names using partial search words
}
\examples{
\dontrun{

find_param_names(name = "albedo")

find_param_names(name = "albedo", version = "V9.0")

find_param_names(name = c("albedo", "latitude", "humcapil"))

find_param_names(name = c("al", "hum"))

find_param_names(name = "albedo", exact = TRUE)

}


}
