% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_bounds.R
\name{fix_missing_bounds}
\alias{fix_missing_bounds}
\title{Setting bounds values (min, max) for missing bounds and/or
naming bounds values vector with bounds names}
\usage{
fix_missing_bounds(values, bounds_name)
}
\arguments{
\item{values}{Vector of bounds values (may be NULL or empty value)}

\item{bounds_name}{Character vector of bounds names (in "min", "max"
one of them or both)}
}
\value{
A named vector for bound(s)
}
\description{
Setting bounds values (min, max) for missing bounds and/or
naming bounds values vector with bounds names
}
\keyword{internal}
