% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_climate.R
\name{gen_climate}
\alias{gen_climate}
\title{Generating the STICS climat.txt file}
\usage{
gen_climate(files_path, out_dir)
}
\arguments{
\item{files_path}{Vector of 1 or 2 weather data files}

\item{out_dir}{output directory path}
}
\value{
a generation success status (TRUE/FALSE)
}
\description{
Generating the STICS climat.txt file
}
\details{
If 2 files are given, the file list is completed if the years
are not consecutive using the complete_climate_paths function
(for a multi-years data set)
}
\examples{
\dontrun{
SticsRFiles:::gen_climate(
                  c("path/to/weather.year1", "path/to/weather.year2"),
                 "/path/to/out/dir" )
}

}
\keyword{internal}
