% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ini_doc.R
\name{gen_ini_doc}
\alias{gen_ini_doc}
\title{Generates from a template a STICS ini xml_document}
\usage{
gen_ini_doc(
  xml_doc = NULL,
  param_table = NULL,
  crop_tag = "Crop",
  params_desc = NULL,
  stics_version = "latest",
  check_names = TRUE
)
}
\arguments{
\item{xml_doc}{an optional xml_document object (created from an ini file)}

\item{param_table}{a table (df, tibble) containing parameters to use
(optional)}

\item{crop_tag}{a crop identifier for crop parameters
(for example like "crop" used for in a parameter suffix :
param_crop1, param_crop2)}

\item{params_desc}{a list describing crop parameters and others}

\item{stics_version}{the STICS files version to use (optional,
default to latest). Only used if xml_doc = NULL.}

\item{check_names}{logical for checking names of param_table columns or not}
}
\value{
an invisible xml_document object or a list of
}
\description{
Generates from a template a STICS ini xml_document
}
\examples{
\dontrun{
library(readxl)

xl_path <- "inputs_stics_example.xlsx"
download_usm_xl(file = xl_path)
ini_param_df <- read_excel(xl_path, sheet = "Ini")
ini_doc <- SticsRFiles:::gen_ini_doc(param_table = ini_param_df)
}

}
\keyword{internal}
