% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ini_xml.R
\name{gen_ini_xml}
\alias{gen_ini_xml}
\title{Generate from a template or an input file a Stics ini xml file}
\usage{
gen_ini_xml(
  ini_out_file = NULL,
  param_table = NULL,
  ini_in_file = NULL,
  out_path = getwd(),
  crop_tag = "crop",
  stics_version = "last"
)
}
\arguments{
\item{ini_out_file}{file name of the output ini xml file}

\item{param_table}{a table (df, tibble) containing parameters to use}

\item{ini_in_file}{file path for an input ini xml file}

\item{out_path}{path for the ini output file (optional)}

\item{crop_tag}{identifier for the crop parameters names}

\item{stics_version}{the stics files version to use}
}
\value{
an invisible xmlDocument object or a list of
}
\description{
Generate from a template or an input file a Stics ini xml file
}
\examples{
\dontrun{
copy_mailing_example(xl_name = "inputs_stics_example.xlsx", dest_dir = "path/to/dest/dir")
xl_path <- file.path("path/to/dest/dir","inputs_stics_example.xlsx")
ini_param_df <- read_excel(xl_path, sheet = "Ini")
gen_ini_xml(param_table = ini_param_df)

}

}
