% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_new_travail.R
\name{gen_new_travail}
\alias{gen_new_travail}
\title{Generates a \code{new_travail.usm} file from usm XML files content and some
switch parameters for lai forcing, parameters forcing and usm chaining.}
\usage{
gen_new_travail(
  workspace,
  usm,
  lai_forcing = 0,
  codesuite = 0,
  codoptim = 0,
  out_dir = getwd()
)
}
\arguments{
\item{workspace}{Path of the directory containing the usm XML files}

\item{usm}{Usm name}

\item{lai_forcing}{1, if \code{lai} is to be read from a daily lai, 0 otherwise
input file.}

\item{codesuite}{1, if the usm is to be chained with the previous
simulated (for getting system state variables), 0 otherwise
input file.}

\item{codoptim}{1, if parameters are to be read from a \code{param.sti} file
(containing parameters values to be forced), 0 otherwise}

\item{out_dir}{Directory path where to store the \code{new_travail.usm} file.}
}
\description{
Generates a \code{new_travail.usm} file from usm XML files content and some
switch parameters for lai forcing, parameters forcing and usm chaining.
}
\keyword{internal}
