% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_obs.R
\name{gen_obs}
\alias{gen_obs}
\title{Write STICS obs data from data.frame}
\usage{
gen_obs(obs_table, path)
}
\arguments{
\item{obs_table}{data.frame to write}

\item{path}{Path to the file to write to}
}
\description{
Write STICS obs data from data.frame
}
\examples{
\dontrun{
xl_path <- file.path(system.file(package="SticsRFiles","extdata/xl/inputs_stics_example.xlsx")
obs_df <- read_excel(xl_path, sheet = "Obs")
gen_obs(obs_df, "USM_2017_T1_CI.obs")
}

}
