% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_obs.R
\name{gen_obs_}
\alias{gen_obs_}
\title{Write STICS obs data from data.frame}
\usage{
gen_obs_(obs_table, file_path)
}
\arguments{
\item{obs_table}{data.frame to write}

\item{file_path}{Path to the file to write to}
}
\value{
A logical value if the file generation succeeded (TRUE) or not (FALSE)
}
\description{
Write STICS obs data from data.frame/tibble
}
\examples{
\dontrun{
# Getting observations data
xl_path <- download_usm_xl(file = "inputs_stics_example.xlsx")

# Loading and filtering data for usm "USM_2017_T1_CI"
obs_df <- read_params_table(file = xl_path, sheet_name = "Obs") \%>\%
  dplyr::filter(usm_name \%in\% "USM_2017_T1_CI")

# Generating the csv file
gen_obs_(obs_df, "USM_2017_T1_CI.obs")
}

}
\keyword{internal}
