% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_param_sti.R
\name{gen_param_sti}
\alias{gen_param_sti}
\title{Generating a param.sti type file}
\usage{
gen_param_sti(workspace, par_names, par_values, file_name = "param.sti")
}
\arguments{
\item{workspace}{Stics or JavaStics workspace path}

\item{par_names}{vector of parameters names}

\item{par_values}{vector of parameters values}

\item{file_name}{file name to generate (default value: param.sti)}
}
\description{
Generating a parameters forcing file from parameters names
and parameters values vectors
}
\examples{
\dontrun{
gen_param_sti(".", c("par1","par2"), c(1,2))
gen_param_sti("/path/to/stics/workspace", c("par1","par2"), c(1,2))
}
}
