% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_paramsti.R
\name{gen_paramsti}
\alias{gen_paramsti}
\title{Generating a param.sti type file}
\usage{
gen_paramsti(workspace, par_names, par_values, file_name = "param.sti")
}
\arguments{
\item{workspace}{Stics or JavaStics workspace path}

\item{par_names}{vector of parameters names}

\item{par_values}{vector of parameters values}

\item{file_name}{file name to generate (default value: param.sti)}
}
\value{
A logical status TRUE if successfull, FALSE otherwise
}
\description{
Generating a parameters forcing file from parameters names
and parameters values vectors
}
\examples{
\dontrun{
gen_paramsti(".", c("par1","par2"), c(1,2))
gen_paramsti("/path/to/stics/workspace", c("par1","par2"), c(1,2))
}

}
