% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_xml2txt.R
\name{gen_sol_xsl_file}
\alias{gen_sol_xsl_file}
\title{Generating the soil xsl stylesheet for a soil name}
\usage{
gen_sol_xsl_file(workspace, usm, stics_version)
}
\arguments{
\item{workspace}{Path of a JavaSTICS workspace
(i.e. containing the STICS XML input files)}

\item{usm}{an usm name}

\item{stics_version}{the STICS files version to use}
}
\value{
conversion success status (TRUE/FALSE)
}
\description{
Generating the soil xsl stylesheet for a soil name
}
\examples{
\dontrun{
SticsRFiles:::gen_sol_xsl_file("path/to/workspace", "usm_name", "V10" )
}

}
\keyword{internal}
