% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sols_doc.R
\name{gen_sols_doc}
\alias{gen_sols_doc}
\title{Generate from a template a Stics sols xmlDocument}
\usage{
gen_sols_doc(
  xml_doc_object = NULL,
  sols_nb = NULL,
  sols_param = NULL,
  stics_version = "last"
)
}
\arguments{
\item{xml_doc_object}{an xmlDocument object (created from a sols.xml file)}

\item{sols_nb}{number of soils to create}

\item{sols_param}{a table (df, tibble) containing parameters to use}

\item{stics_version}{the stics files version to use}
}
\value{
an xmlDocument object
}
\description{
Generate from a template a Stics sols xmlDocument
}
\keyword{internal}
