% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sols_xml.R
\name{gen_sols_xml}
\alias{gen_sols_xml}
\title{Generate from a template a Stics sols xml file}
\usage{
gen_sols_xml(
  sols_out_file,
  sols_nb = NULL,
  sols_param = NULL,
  sols_in_file = NULL,
  stics_version = "last"
)
}
\arguments{
\item{sols_out_file}{file path of the output sols xml file}

\item{sols_nb}{number of soils to create}

\item{sols_param}{a table (df, tibble) containing parameters to use}

\item{sols_in_file}{file path for an input sols xml file}

\item{stics_version}{the stics files version to use}
}
\value{
an invisible xmlDocument object
}
\description{
Generate from a template a Stics sols xml file
}
\examples{
\dontrun{
copy_mailing_example(xl_name = "inputs_stics_example.xlsx", dest_dir = "/path/to/dest/dir")
xl_path <- file.path("/path/to/dest/dir","inputs_stics_example.xlsx")
sols_param_df <- read_excel(xl_path, sheet = "Soils")
gen_sols_xml(sols_out_file = file.path("/path/to/dest/dir","sols.xml"),
sols_param = sols_param_df)
}

}
