% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sta_doc.R
\name{gen_sta_doc}
\alias{gen_sta_doc}
\title{Generates from a template a STICS sta xml_document}
\usage{
gen_sta_doc(
  xml_doc = NULL,
  param_table = NULL,
  stics_version = "latest",
  check_names = TRUE
)
}
\arguments{
\item{xml_doc}{an optional xml_document object (created from an sta file)}

\item{param_table}{a table (df, tibble) containing parameters to use
(optional)}

\item{stics_version}{the STICS files version to use
(optional, default to latest). Only used if xml_doc = NULL.}

\item{check_names}{logical for checking names of param_table columns or not}
}
\value{
an invisible xml_document object or a list of
}
\description{
Generates from a template a STICS sta xml_document
}
\examples{
\dontrun{
library(readxl)

xl_path <- "inputs_stics_example.xlsx"
download_usm_xl(file = xl_path)
sta_param_df <- readxl::read_excel(xl_path, sheet = "Station")
sta_doc <- SticsRFiles:::gen_sta_doc(param_table = sta_param_df)
}

}
\keyword{internal}
