% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sta_file_xml.R
\name{gen_sta_file_xml}
\alias{gen_sta_file_xml}
\title{Generate from a template  or an input file a Stics sta xml file}
\usage{
gen_sta_file_xml(
  sta_out_file = NULL,
  param_table = NULL,
  sta_in_file = NULL,
  out_path = getwd(),
  stics_version = "last"
)
}
\arguments{
\item{sta_out_file}{file name of the output sta xml file}

\item{param_table}{a table (df, tibble) containing parameters to use}

\item{sta_in_file}{file path for an input sta xml file}

\item{out_path}{path for the sta output file}

\item{stics_version}{the stics files version to use}
}
\value{
an invisible xmlDocument object or a list of
}
\description{
Generate from a template  or an input file a Stics sta xml file
}
