% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tec_doc.R
\name{gen_tec_doc}
\alias{gen_tec_doc}
\title{Generate  from a template or modify a Stics tec xmlDocument}
\usage{
gen_tec_doc(
  xml_doc_object = NULL,
  param_table = NULL,
  stics_version = "last",
  dict = NULL
)
}
\arguments{
\item{xml_doc_object}{an xmlDocument object (created from an ini file)}

\item{param_table}{a table (df, tibble) containing parameters to use}

\item{stics_version}{the stics files version to use}

\item{dict}{List of correspondance between given parameter names and internal names.}
}
\value{
an xmlDocument object or a list of
}
\description{
Generate  from a template or modify a Stics tec xmlDocument
}
\keyword{internal}
