% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tec_file.R
\name{gen_tec_file}
\alias{gen_tec_file}
\title{Generate from a template  or an input file a Stics tec xml file}
\usage{
gen_tec_file(tec_out_file = NULL, param_table = NULL,
  tec_in_file = NULL, out_path = getwd(), stics_version = "last",
  dict = NULL)
}
\arguments{
\item{tec_out_file}{file name of the output tec xml file}

\item{param_table}{a table (df, tibble) containing parameters to use}

\item{tec_in_file}{file path for an input tec xml file}

\item{out_path}{path for the ini output file}

\item{stics_version}{the stics files version to use}
}
\value{
an invisible xmlDocument object or a list of
}
\description{
Generate from a template  or an input file a Stics tec xml file
}
