% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_usms_doc.R
\name{gen_usms_doc}
\alias{gen_usms_doc}
\title{Generate from a template a Stics usms xmlDocument}
\usage{
gen_usms_doc(
  xml_doc = NULL,
  usms_nb = NULL,
  usms_param = NULL,
  stics_version = "last"
)
}
\arguments{
\item{xml_doc}{an optional xmlDocument object (created from an usms.xml file)}

\item{usms_nb}{number of usms (usm nodes) to create in the xmlDocument (optional)}

\item{usms_param}{a table (df, tibble) containing parameters to use (optional)}

\item{stics_version}{the stics files version to use (optional, default to last). Only used if xml_doc = NULL.}
}
\value{
an invisible xmlDocument object
}
\description{
Generate from a template a Stics usms xmlDocument
}
\examples{
\dontrun{
library(readxl)

xl_path <- "inputs_stics_example.xlsx"
download_usm_xl(xl_name = xl_path)
usms_param_df <- read_excel(xl_path, sheet = "USMs")
usms_doc <- SticsRFiles:::gen_usms_doc(usms_param = usms_param_df)
}

}
\keyword{internal}
