% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_usms_obs.R
\name{gen_usms_obs}
\alias{gen_usms_obs}
\title{Generating observation data files from a data.frame}
\usage{
gen_usms_obs(obs_table, out_path = getwd())
}
\arguments{
\item{obs_table}{The input onservations data.frame}

\item{out_path}{The output directory path for storing .obs files}
}
\value{
A return logical status indicating if any error when writing files (FALSE),
TRUE when no errors.
}
\description{
Generating observation data files from a data.frame
}
\examples{
\dontrun{
copy_mailing_example(xl_name = "inputs_stics_example.xlsx", dest_dir = "/path/to/dest/dir")
xl_path <- file.path("/path/to/dest/dir","inputs_stics_example.xlsx")
obs_df <- read_excel(xl_path, sheet = "Obs")
gen_usms_obs(obs_table = obs_df, out_path = file.path("/path/to/dest/dir","obs"))
}

}
