% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_usms_sols_doc.R
\name{gen_usms_sols_doc}
\alias{gen_usms_sols_doc}
\title{Generate from a template or modify a Stics sols or usms xmlDocument}
\usage{
gen_usms_sols_doc(
  doc_type,
  xml_doc = NULL,
  nodes_nb = NULL,
  nodes_param = NULL,
  stics_version = "last"
)
}
\arguments{
\item{doc_type}{Document type}

\item{xml_doc}{The xml document}

\item{nodes_nb}{The number of nodes}

\item{nodes_param}{Node parameter}

\item{stics_version}{Version of the STICS model}
}
\value{
An xmlDocument object
}
\description{
Generate from a template or modify a Stics sols or usms xmlDocument
}
\examples{
\dontrun{
# A newly created one, with one or more usms and
# fake parameters values
xml_doc <- SticsRFiles:::gen_xml_doc(doc_type = "usms")
xml_doc <- SticsRFiles:::gen_xml_doc(doc_type = "usms", node_nb = 3)

# With changing parameters values, from an existing document
# and a data.frame contaning usms parameters values
existing_doc <- SticsRFiles:::xmldocument("/path/to/usms.xml")

out_xml_doc <- SticsRFiles:::gen_xml_doc(doc_type = "usms",
xml_doc = existing_doc, nodes_param = param_data_frame)


}

}
\keyword{internal}
