% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_var_mod.R
\name{gen_var_mod}
\alias{gen_var_mod}
\title{Generating a var.mod type file}
\usage{
gen_var_mod(workspace, var_names, file_name = "var.mod")
}
\arguments{
\item{workspace}{Stics or JavaStics workspace path}

\item{var_names}{vector of variables names}

\item{file_name}{file name to generate (default value: var.mod)}
}
\description{
Generating a daily variable list file from variables names
}
\examples{
\dontrun{
 gen_var_mod(".", c("var1","var2"))
 gen_var_mod("/path/to/stics/workspace", c("var1","var2"))
}

}
