% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_xml_doc.R
\name{gen_xml_doc}
\alias{gen_xml_doc}
\title{Generate from a template or modify a Stics sols or usms xmlDocument}
\usage{
gen_xml_doc(doc_type, xml_doc_object = NULL, nodes_nb = NULL,
  nodes_param = NULL, stics_version = "last", overwrite = F)
}
\description{
Generate from a template or modify a Stics sols or usms xmlDocument
}
