% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_stics_versions.R
\name{get_csv_files}
\alias{get_csv_files}
\title{Dowloading inputs and/or outputs csv files from a repository branch/tag}
\usage{
get_csv_files(
  branch_url,
  dest_dir,
  file = "all",
  ids = get_svn_identifiers(),
  overwrite = FALSE
)
}
\arguments{
\item{branch_url}{Address of the branch or tag}

\item{dest_dir}{Directory path where to store files}

\item{file}{File name (inputs.csv or ouputs.csv)
or keyword "all" for both files (default)}

\item{ids}{Connexion identifiers to the subversion server}

\item{overwrite}{Logical, TRUE for overwtiting files, FALSE otherwise}
}
\description{
Dowloading inputs and/or outputs csv files from a repository branch/tag
}
\examples{
\dontrun{
SticsRFiles:::get_csv_files(
  branch_url = "https://w3.avignon.inra.fr/svn/modulostics/branches/branch10",
  dest_dir = system.file("extdata", package = "SticsRFiles"))
}
}
\keyword{internal}
