% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cultivars_list.R
\name{get_cultivars_list}
\alias{get_cultivars_list}
\title{Get the cultivar names for an xml plant file (*_plt.xml)}
\usage{
get_cultivars_list(file)
}
\arguments{
\item{file}{The path of a plant file.}
}
\value{
A vector of cultivar names
}
\description{
Extracts the cultivar names from a plant file
}
\examples{
path <- get_examples_path(file_type = "xml")

# Read from a plant file (all cultivars available in a plant file)
cv_list <- get_cultivars_list(file = file.path(path, "file_plt.xml"))

}
