% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cultivars_param.R
\name{get_cultivars_param}
\alias{get_cultivars_param}
\title{Get the values of cultivar-specific parameters from an xml
plant file (*_plt.xml)}
\usage{
get_cultivars_param(file)
}
\arguments{
\item{file}{The path of a plant file.}
}
\value{
A data.frame with one row per cultivar and one column per parameter
}
\description{
Extracts the values of cultivar-specific parameters from a
plant file
}
\examples{
path <- get_examples_path(file_type = "xml")

# Read from a plant file (all cultivars available in a plant file)
cv_param_df <- get_cultivars_param(file = file.path(path, "file_plt.xml"))

}
