% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_results.R
\name{get_daily_result}
\alias{get_daily_result}
\title{Load and format Stics daily output file}
\usage{
get_daily_result(
  workspace,
  usm_name,
  var_list = NULL,
  doy_list = NULL,
  dates_list = NULL,
  mixed = NULL,
  usms_filename = "usms.xml",
  javastics_path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{workspace}{Stics or JavaStics workspace path containing the \code{mod_s*.sti} files (see details)}

\item{usm_name}{usm name}

\item{var_list}{vector of output variables names (optional, see \code{get_var_info()} to get the names of the variables)}

\item{doy_list}{vector of cumulative DOYs (optional)}

\item{dates_list}{list of dates (optional)}

\item{mixed}{\code{TRUE}: intercrop, \code{FALSE}: sole crops (default), \code{NULL}: guess from XML file.}

\item{usms_filename}{usms file path (e.g. "usms.xml") if \code{mixed= NULL}.}

\item{javastics_path}{JavaStics installation path (Optional, needed if the plant files are not in the \code{workspace}
but rather in the JavaStics default workspace)}

\item{verbose}{Logical value (optional), TRUE to display infos on error, FALSE otherwise (default)}
}
\value{
A tibble
}
\description{
Reads and format daily output file (mod_s*.sti) from STICS as a tibble with
possible selection on variable names, cumulative DOY, dates list.
}
\details{
The function can guess if the usm is mixed or not by reading the XML
input file. To do so, set the \code{mixed} argument to \code{NULL}.
}
\examples{
\dontrun{
path <- get_examples_path( file_type = "sti")
get_daily_result(path,"banana")
}
}
\seealso{
\code{get_daily_results()}, the function exported by the package that calls \code{get_daily_result()} on all usms
}
\keyword{internal}
