% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_results.R
\name{get_daily_results}
\alias{get_daily_results}
\title{Loading Stics daily output file(s)}
\usage{
get_daily_results(
  workspace,
  usm_name,
  var_list = NULL,
  doy_list = NULL,
  dates_list = NULL
)
}
\arguments{
\item{workspace}{Stics or JavaStics workspace path containing the \code{mod_s*.sti} files (see details)}

\item{usm_name}{vector of usm(s) names}

\item{var_list}{vector of output variables names (optional)}

\item{doy_list}{vector of cumulative DOYs (optional)}

\item{dates_list}{list of dates (optional)}
}
\value{
A tibble or a list of
}
\description{
Reading daily file(s) (mod_s*.sti) as a tibble with possible selection
on variable names, cumulative DOY, dates list.
}
\details{
For intercrops, please add a "p" or a "a" before the USM name in \code{usm_name} to
get the results for the principal or associated crop respectively.
}
\examples{
path <- system.file(file.path("extdata","sti","V9.0"), package = "SticsRFiles")
get_daily_results(path,"banana")

}
