% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_stics_versions.R
\name{get_data_dir}
\alias{get_data_dir}
\title{Getting package or library data path}
\usage{
get_data_dir(location = "install")
}
\arguments{
\item{location}{The destination where to remove information and data
"install" for removing things from the installed SticsRFiles library
(default), "package" for removing them from the package project (in RStudio)}
}
\value{
A directory path
}
\description{
Getting package or library data path
}
\examples{
\dontrun{
SticsRFiles:::get_data_dir()

SticsRFiles:::get_data_dir(location = "package")
}
}
\keyword{internal}
