% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_day_number.R
\name{get_day_of_year}
\alias{get_day_of_year}
\title{Day of year}
\usage{
get_day_of_year(date)
}
\arguments{
\item{date}{date you wish to convert, in the \code{\link[base]{Date}} format}
}
\value{
integer
}
\description{
Returns the day number of the year corresponding to a given date.
}
\author{
Timothee Flutre
}
