% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dev_stages.R
\name{get_dev_stages}
\alias{get_dev_stages}
\title{Get developmental stages from simulations}
\usage{
get_dev_stages(
  sim,
  dev_vars = c("iplts", "imbs", "ilets", "igers", "idebdorms", "ifindorms", "ilevs",
    "iamfs", "ilaxs", "ilats", "iflos", "idrps", "inous", "idebdess", "imats", "irecs"),
  usm = NULL
)
}
\arguments{
\item{sim}{A list, where each element is a \code{data.frame} of simulation results
for the given usm. The list is named after the USM name.}

\item{dev_vars}{Vector of output variables corresponding to developmental stages.}

\item{usm}{Vector of USM names. Optional, if not provided, the function
returns the results for all USMs.}
}
\value{
List of data.frame, one per USM, with one row per developmental stage.
}
\description{
Extracts the values of developmental stages from simulations
}
\seealso{
\code{\link{get_sim}}
}
\author{
Timothee Flutre
}
