% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_int.R
\name{get_file_int}
\alias{get_file_int}
\title{Read STICS observation or simulation file (\emph{.obs or mod_s})}
\usage{
get_file_int(workspace, filename, plant_name = NULL, verbose = TRUE)
}
\arguments{
\item{workspace}{A directory where to read the files}

\item{filename}{A vector of file names}

\item{plant_name}{A vector of plant names. Length of filename, or recycled if one (optional).}

\item{verbose}{Logical value (optional), TRUE to display read file name, FALSE otherwise (default)}
}
\value{
A \code{data.frame} of all observation or simulation files concatenated by row.
}
\description{
Read STICS observation or simulation file(s). This function is not exported because
it only reads and pre-formats the file without any clever tests for intercrops or usms.
}
\examples{
\dontrun{
path <- file.path(get_examples_path(file_type = "obs"))
get_file_int(path, "banana.obs")
}
}
\keyword{internal}
