% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files_list.R
\name{get_files_list}
\alias{get_files_list}
\title{Getting existing xml files path list per usm from an usms.xml file}
\usage{
get_files_list(
  workspace,
  usm = NULL,
  usms_file = "usms.xml",
  file_type = NULL,
  only_existing = TRUE,
  javastics = NULL,
  use_mod_files = FALSE
)
}
\arguments{
\item{workspace}{Path of a JavaSTICS workspace (i.e. containing the STICS
XML input files)}

\item{usm}{Vector of usms names (optional)}

\item{usms_file}{Path (including name) of a USM XML file (default: usms.xml).}

\item{file_type}{Vector of file(s) type to get (optional; if not given,
all types are returned, see details)}

\item{only_existing}{logical, optional; if TRUE (default) only existing files
list is returned, FALSE otherwise for all files}

\item{javastics}{Path of JavaSTICS (optional), only needed when the plant
files, general parameter files or .mod files are not in the workspace
but included in the JavaSTICS distribution}

\item{use_mod_files}{logical, optional; if TRUE adding the .mod files to the list FALSE
otherwise (default)}
}
\value{
A named list (usms names) with files path in \code{paths} field and
a logical vector in \code{exist} field indicating if the file exists or not.
}
\description{
Getting existing xml files path list per usm from an usms.xml file
}
\details{
The possible values for file_type are: "fplt", "finit", "fclim1",
"fclim2", "fstation", "ftec", "sols", "pargen", "parnew", "usms", "mod"
}
\examples{
\dontrun{

get_files_list(
  workspace = "/path/to/workspace",
  javastics = "/path/to/JavaSTICS/folder"
)

get_files_list(
  workspace = "/path/to/workspace",
  javastics = "/path/to/JavaSTICS/folder",
  usm = c("usm1", "usm3")
)

get_files_list(
  workspace = "/path/to/workspace",
  file_type = c("finit", "ftec")
)

get_files_list(
  workspace = "/path/to/workspace",
  use_mod_files = TRUE,
  javastics = "/path/to/JavaSTICS/folder"
)
}

}
\seealso{
See \code{get_soils_list()} to get all soils in a usm file,
and \code{get_usms_list()} to get the list of usms.
}
