% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files_param_bounds.R
\name{get_files_param_bounds}
\alias{get_files_param_bounds}
\title{Getting parameters bounds from xml files}
\usage{
get_files_param_bounds(xml_file, param_name, bounds_name = NULL)
}
\arguments{
\item{xml_file}{an xml file path or a vector of paths}

\item{param_name}{a parameter name of a vector of parameters names}

\item{bounds_name}{bounds name "min" or "max" (optional, default value c("min","max ))}
}
\value{
A list of parameters bounds values
}
\description{
Extracting parameters min and/or max bounds for a parameter
or a vector of parameters from a file or a vector of xml files.
}
\examples{
library(SticsRFiles)
xml_file <- "path/to/xmlfile"
xml_files_list <- c("path/to/xmlfile1", "path/to/xmlfile2")
param_bounds <- get_files_param_bounds(xml_file, "param", "min")

param_bounds <- get_files_param_bounds(xml_file, c("param1", "param2"), "max")

param_bounds <- get_files_param_bounds(xml_file, c("param1", "param2"))

param_bounds <- get_files_param_bounds(xml_files_list, "param", "min")

}
