% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files_param_names.R
\name{get_files_params_names}
\alias{get_files_params_names}
\title{Getting parameters names from xml files}
\usage{
get_files_params_names(xml_file, mult = FALSE)
}
\arguments{
\item{xml_file}{an xml file path or a vector of paths}

\item{mult}{extra argument for detecting a single file use}
}
\value{
A list of parameters names list
}
\description{
Extracting parameters names for an xml files or a vector of files.
}
\examples{
\dontrun{
library(SticsRFiles)
xml_file <- "path/to/xmlfile"
xml_files_list <- c("path/to/xmlfile1", "path/to/xmlfile2")
param_names <- get_files_params_names(xml_file)

param_names <- get_files_params_names(xml_file)

param_names <- get_files_params_names(xml_file)

param_names <- get_files_params_names(xml_files_list)
}

}
