% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formalisms_xml.R
\name{get_formalisms_xml}
\alias{get_formalisms_xml}
\title{Get the formalism(s) parameters from a file or a list of}
\usage{
get_formalisms_xml(xml_file, par_name = NULL, by_form = TRUE)
}
\arguments{
\item{xml_file}{an xml file path or a vector of paths}

\item{par_name}{Optional. A parameter name or a vector of. If not provided,
all the parameters are extracted.}

\item{by_form}{Logical. Structuring parameters by formalism (TRUE),
or not (structuring output by parameter) for each file}
}
\value{
A list (by file) of list of parameters names by formalism name
}
\description{
Get the formalism(s) parameters from a file or a list of
}
\examples{
\dontrun{

xml_sta <- system.file(paste0("extdata/xml/examples/V9.1/file_sta.xml"),
package = "SticsRFiles")

par_form <- get_formalisms_xml(xml_sta, "zr")

par_form_list <- get_formalisms_xml(xml_sta, c("zr","altistation"))

par_form_list <- get_formalisms_xml(xml_sta)

}

}
