% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_txt.R
\name{get_ini_txt}
\alias{get_ini_txt}
\alias{get_general_txt}
\alias{get_tmp_txt}
\alias{get_plant_txt}
\alias{get_tec_txt}
\alias{get_soil_txt}
\alias{get_station_txt}
\alias{get_usm_txt}
\title{Read STICS input parameters files}
\usage{
get_ini_txt(filepath = "ficini.txt")

get_general_txt(filepath = "tempopar.sti")

get_tmp_txt(filepath = "tempoparv6.sti")

get_plant_txt(filepath = "ficplt1.txt", variety = NULL)

get_tec_txt(
  filepath = "fictec1.txt",
  several_fert = T,
  several_thin = T,
  is_pasture = F
)

get_soil_txt(filepath = "param.sol")

get_station_txt(filepath = "station.txt")

get_usm_txt(filepath = "new_travail.usm")
}
\arguments{
\item{filepath}{File path}

\item{variety}{Integer. The plant variety to get the parameter from.}

\item{several_fert}{Is there several fertilization in the USM ?}

\item{several_thin}{Is there several thinning in the USM ?}

\item{is_pasture}{Is the plant a pasture ?}
}
\value{
A list of parameters, depending on the file/function:
\item{ini}{Initialization parameters}
\item{general}{General parameters}
\item{tec}{Technical parameters}
\item{plant}{Plant parameters}
\item{soil}{Soil parameters}
\item{station}{Station parameters}
}
\description{
Read a specific STICS model input parameter file.
Users would generally use the wrapper \code{get_param_txt()} instead.
}
\details{
he functions are compatible with intercrops.
}
\note{
Users generally only use \code{get_param_txt()}, which is a wrapper for all
these functions.
}
\examples{
\dontrun{
# Read the initialisation file (ficini.txt):
library(SticsRFiles)
path = system.file("extdata/txt/V9.1/ficini.txt", package = "SticsRFiles")
get_ini_txt(path)
}

}
\seealso{
\code{get_param_txt()}.
}
