% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inputs_example.R
\name{get_inputs_example}
\alias{get_inputs_example}
\title{Get an example of an XL parameters file into a directory (from package data)}
\usage{
get_inputs_example(xl_name = NULL, dest_dir = getwd())
}
\arguments{
\item{xl_name}{name of the XL file or partial match}

\item{dest_dir}{directory path where to copy the XL file}
}
\description{
Get an example of an XL parameters file into a directory (from package data)
}
