% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javastics_cmd_util.R
\name{get_java_version}
\alias{get_java_version}
\title{Getting the java virtual machine version}
\usage{
get_java_version(java_cmd = "java")
}
\arguments{
\item{java_cmd}{Name or path of the java virtual machine executable}
}
\value{
A numerical version, with a string version as "version" attribute
}
\description{
For Unix like systems, the system java virtual machine needs to
be detected for compatibility checking
}
\examples{
\dontrun{
get_java_version(java_cmd = "java")
get_java_version(java_cmd = "/path/to/java")
}

}
\keyword{internal}
