% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_java_workspace.R
\name{get_java_workspace}
\alias{get_java_workspace}
\title{Getting current workspace in JavaSTICS preferences configuration}
\usage{
get_java_workspace(javastics)
}
\arguments{
\item{javastics}{JavaSTICS installation root folder}
}
\description{
Getting current JavaSTICS working directory,
if not any setting to \code{example} directory
}
\examples{
\dontrun{
get_java_wd("/path/to/JavaSTICS/folder")
}

}
\keyword{internal}
