% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javastics_cmd_util.R
\name{get_javastics_cmd}
\alias{get_javastics_cmd}
\title{Get JavaSTICS command line strings}
\usage{
get_javastics_cmd(
  javastics,
  java_cmd = "java",
  type = c("generate", "run"),
  workspace = NULL,
  verbose = TRUE
)
}
\arguments{
\item{javastics}{JavaSTICS installation root folder}

\item{java_cmd}{Name or path of the java virtual machine executable
(optional)}

\item{type}{keywords for filtering in the returned list what kind of command
to be generated (default both: "generate" and "run")}

\item{workspace}{A JavaSTICS workspace path (optional)}

\item{verbose}{Logical value (optional), TRUE to display run information,
FALSE otherwise (default)}
}
\value{
A named list with fields : command (java command/path), cmd_generate
and/or cmd_run that are to be used as args in the system2 function call.
}
\description{
From a JavaSTICS path, constructing command lines for
files generation from XML and/or running simulation through
a system2 command execution
}
\details{
According to a JavaSTICS version (i.e. javastics path content) and
the OS type, the function builds command line strings used from system2
calls. Commands may be filtered using type argument.
}
\examples{
\dontrun{
get_javastics_cmd(
  javastics = "/path/to/JavaSTICS/folder", ,
  workspace = "path/to/workspace"
)
}

}
\keyword{internal}
