% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javastics_path.R
\name{get_javastics_path}
\alias{get_javastics_path}
\title{Getting JavaSTICS path from javastics_path environment variable}
\usage{
get_javastics_path()
}
\value{
Path of the JavaSTICS root directory if javastics_path
environment variable exists
}
\description{
Getting JavaSTICS path from javastics_path environment variable
}
\keyword{internal}
