% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javastics_cmd_util.R
\name{get_javasticscmd_exe}
\alias{get_javasticscmd_exe}
\title{Get JavaSTICS command line executable name}
\usage{
get_javasticscmd_exe(javastics, verbose = TRUE)
}
\arguments{
\item{javastics}{JavaSTICS installation root folder}

\item{verbose}{Logical value (optional), TRUE to display run infos, FALSE otherwise (default)}
}
\value{
An executable name, with attached command option string as attribute
}
\description{
From a JavaSTICS path, searching for the JavaSTICS command
}
\details{
If JavaStics version < 1.5, returns \code{JavaSticsCmd.exe}
otherwise, returns \code{JavaStics.exe}
}
\examples{
\dontrun{
 get_javasticscmd_exe("path/to/JavaSTICS")
}

}
\keyword{internal}
