% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_to_julian_dates.R
\name{get_julian_days}
\alias{get_julian_days}
\title{Julian days}
\usage{
get_julian_days(date)
}
\arguments{
\item{date}{date you wish to convert, in the \code{\link[base]{Date}} format}
}
\value{
integer
}
\description{
Returns the Julian day equivalent to a given date.
}
\author{
Timothee Flutre
}
