% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lai_forcing.R
\name{get_lai_forcing_xml}
\alias{get_lai_forcing_xml}
\title{Is LAI forced ? plants number per usm, all or selected from
a given list}
\usage{
get_lai_forcing_xml(usm_xml_path, usms_list = c())
}
\arguments{
\item{usm_xml_path}{Path to usms.xml file}

\item{usms_list}{Usms to filter (optional, see details)}
}
\value{
A names numeric vector of a boolean value for each usm
}
\description{
Is the lai forced in usms from a usms.xml file
}
\details{
Use \code{get_usms_list()} to get the list of the usm names.
}
\keyword{internal}
