% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs_txt.R
\name{get_obs_txt}
\alias{get_obs_txt}
\title{Read STICS observation file (*.obs) in the context of txt files}
\usage{
get_obs_txt(dirpath = getwd(), filename = NULL, mixed = NULL)
}
\arguments{
\item{dirpath}{Directory path}

\item{filename}{A vector of observation file name(s). Optional, see details.}

\item{mixed}{(optional) Is the simulation on mixed species (boolean)}
}
\value{
A data.frame of the STICS-formatted observations.
Return is \code{NULL} if no files were found,
or more files than usable. If mixed crops (two \code{filename} provided),
the function binds them and add a new column called "Plant"
which corresponds to the name of each observation file.
}
\description{
Read STICS observation file(s) for sole and/or mixed crops.
}
\details{
If \code{filename} is not specified (or equal to \code{NULL}),
the function tries first to match the \code{mod_s*} names for
the same \*.obs names, and then to use the \code{.obs} file if there
is one only (used for sole crops or for both mixed crops).
If there are no .obs files, or two but not specified for reading,
the function returns \code{NULL}
If \code{mixed} is not specified (or equal to \code{NULL}),
the function tries to read the number of species from the
input files.
}
\examples{
\dontrun{

obs_table <- SticsRFiles:::get_obs_txt()
}

}
\keyword{internal}
