% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_option_choice_param_values.R
\name{get_option_choice_param_values}
\alias{get_option_choice_param_values}
\title{Getting parameter values from an xml parameter file for an option choice}
\usage{
get_option_choice_param_values(
  xml_file_path,
  option_param_name,
  choice_name_or_code
)
}
\arguments{
\item{xml_file_path}{path of xml parameter file}

\item{option_param_name}{name of the option}

\item{choice_name_or_code}{name or code of the option choice desired}
}
\value{
A list containing a vector of strings of parameters values ($values)
and a vector of strings of parameters names ($names)
}
\description{
Extracting parameter values from an xml file data, correponding to parameters
defined in a option choice node
}
\examples{
\dontrun{
get_option_choice_param_values(path/to/xml/file,"codetemp","yes")
}
}
