% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_options_choices.R
\name{get_options_choices}
\alias{get_options_choices}
\title{Getting options parameter choices from an xml parameter file}
\usage{
get_options_choices(xml_file_path, options_names = NULL)
}
\arguments{
\item{xml_file_path}{path of xml parameter file}

\item{options_names}{options names to take into account}
}
\value{
A list of strings of options choice values
}
\description{
Extracting options choices values from an xml file data
}
\examples{
\dontrun{
get_option_choices(path/to/xml/file)
}


}
