% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_options_names.R
\name{get_options_names}
\alias{get_options_names}
\title{Getting options parameter names from an xml parameter file}
\usage{
get_options_names(xml_file_path)
}
\arguments{
\item{xml_file_path}{path of xml parameter file}
}
\value{
A vector of strings of options parameter names
}
\description{
Extracting options parameter name from an xml file data
}
\examples{
\dontrun{
get_options_names(path/to/xml/file)
}

}
\keyword{internal}
