% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_txt.R
\name{get_out_var_txt}
\alias{get_out_var_txt}
\title{Get desired STICS outputs}
\usage{
get_out_var_txt(filepath = "var.mod")
}
\arguments{
\item{filepath}{The path to the "var.mod" file.}
}
\value{
The variables that will be returned by STICS
}
\description{
Get the variables STICS will return ("var.mod" file)
}
\examples{
get_out_var_txt(file.path(system.file("extdata/txt/V8.5", package = "SticsRFiles"),"var.mod"))

}
\seealso{
\code{gen_varmod}
}
