% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_bounds.R
\name{get_param_bounds}
\alias{get_param_bounds}
\title{Getting parameters bounds from an xmlDocument object}
\usage{
get_param_bounds(xml_doc, param_name, bounds_name = NULL)
}
\arguments{
\item{xml_doc}{an xmlDocument class object}

\item{param_name}{a parameter name of a vector of parameters names}

\item{bounds_name}{bounds name "min" or "max" (optional, default value c("min","max ))}
}
\value{
A list of parameters bounds values
}
\description{
Extracting parameters min and/or max bounds for a parameter
or a vector of parameters from an xmlDocument class object.
}
\examples{
\dontrun{
xml_sta <- system.file(paste0("extdata/xml/examples/V9.0/file_sta.xml"),
package = "SticsRFiles")

sta_doc <- SticsRFiles:::xmldocument(xml_sta)

par_bounds <- SticsRFiles:::get_param_bounds(sta_doc,"zr")

par_bounds_list <- SticsRFiles:::get_param_bounds(sta_doc,c("zr","altistation"))


SticsRFiles:::get_param_bounds(sta_doc,c("zr","altistation"),"min")

}


}
