% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_desc.R
\name{get_param_desc}
\alias{get_param_desc}
\title{Title}
\usage{
get_param_desc(file_path = NULL, version = "last", name = NULL, kind = FALSE)
}
\arguments{
\item{file_path}{csv file path}

\item{version}{The stics version. See \code{get_stics_versions_compat()} to get all compatible versions. Default
to "last", a special code to get the last version.}

\item{name}{a name vector for selecting loaded content using name column matching}

\item{kind}{a name vector for selecting loaded content using kind column matching}
}
\description{
Title
}
