% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_formalism.R
\name{get_param_formalism}
\alias{get_param_formalism}
\title{Get the "formalisme" node "@nom" attribute value of a
(or a list of) parameter(s) in an xmlDocument, of a vector of}
\usage{
get_param_formalism(xml_doc, name = NULL, form_only = FALSE)
}
\arguments{
\item{xml_doc}{an xmlDocument object (created from any xml file),
or a vector of}

\item{name}{Optional. A parameter name or a vector of. If not given,
all parameters names are extracted.}
}
\value{
a list parameters formalism name
}
\description{
Get the "formalisme" node "@nom" attribute value of a
(or a list of) parameter(s) in an xmlDocument, of a vector of
}
\examples{
\dontrun{

xml_sta <- system.file(paste0("extdata/xml/examples/V9.1/file_sta.xml"),
package = "SticsRFiles")

sta_doc <- SticsRFiles:::xmldocument(xml_sta)

par_form <- SticsRFiles:::get_param_formalism(sta_doc,"zr")

par_form_list <- SticsRFiles:::get_param_formalism(sta_doc,c("zr","altistation"))

}

}
\keyword{internal}
