% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_formalisms.R
\name{get_param_formalisms}
\alias{get_param_formalisms}
\title{Get the "formalisme" node "@nom" attribute value of a
(or a list of) parameter(s) in an xml_document, of a vector of}
\usage{
get_param_formalisms(xml_doc, name = NULL, form_only = FALSE)
}
\arguments{
\item{xml_doc}{an xml_document object (created from any xml file),
or a vector of}

\item{name}{Optional. A parameter name or a vector of. If not given,
all parameters names are extracted.}
}
\value{
a list parameters formalism name
}
\description{
Get the "formalisme" node "@nom" attribute value of a
(or a list of) parameter(s) in an xml_document, of a vector of
}
\examples{
\dontrun{

xml_sta <- file.path(get_examples_path(file_type = "xml"), "file_sta.xml")

sta_doc <- SticsRFiles:::xmldocument(xml_sta)

par_form <- SticsRFiles:::get_param_formalisms(sta_doc, "zr")

par_form_list <- SticsRFiles:::get_param_formalisms(sta_doc,
                                                   c("zr", "altistation"))
}

}
\keyword{internal}
