% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_infos_txt.R
\name{get_param_infos_txt}
\alias{get_param_infos_txt}
\title{Title}
\usage{
get_param_infos_txt(
  file_path = NULL,
  version = "V9.0",
  name = NULL,
  kind = FALSE
)
}
\arguments{
\item{version}{}
}
\value{

}
\description{
Title
}
