% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_names.R
\name{get_param_names}
\alias{get_param_names}
\title{Get a list of STICS xml parameters names an xml_document or  XML node}
\usage{
get_param_names(
  xml_object,
  param_list = c(),
  parent_name = NULL,
  parent_sel_attr = NULL,
  full_list = FALSE,
  root_name = NULL
)
}
\arguments{
\item{xml_object}{an xml XML::XMLInternalElementNode
or SticsRFiles::xml_document object}

\item{param_list}{param names vector, only used for recursive calls}

\item{full_list}{TRUE for getting all names, FALSE otherwise (default)}

\item{root_name}{Only for getting the root node name (file type),
useful for filtering unwanted names to be includes in parameters names list

for unique names list}
}
\value{
a character vector of parameters names
}
\description{
Get a list of STICS xml parameters names an xml_document or  XML node
}
\examples{
\dontrun{
xml_path <- file.path(get_examples_path(file_type = "xml"), "sols.xml")
sols_doc <- SticsRFiles:::xmldocument(xml_path)
SticsRFiles:::get_param_names(sols_doc)
}

}
\keyword{internal}
