% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_names_xml.R
\name{get_param_names_xml}
\alias{get_param_names_xml}
\title{Getting parameters names from xml files}
\usage{
get_param_names_xml(
  xml_file,
  bounds = TRUE,
  output = "data.frame",
  combine = TRUE
)
}
\arguments{
\item{xml_file}{an xml file path or a vector of paths}

\item{output}{Output data format either "list" or "data.frame" (default)}

\item{combine}{Logical, usefull only for data.frame.
TRUE, to transform a data.frame list to a unique data.frame,
FALSE otherwise.}
}
\value{
A list of parameters names data.frames or list, or a unique data.frame
for multiple files.
}
\description{
Extracting parameters names for an xml files or a vector of files.
}
\examples{
\dontrun{
library(SticsRFiles)
xml_file <- "path/to/xmlfile"
xml_files_list <- c("path/to/xmlfile1", "path/to/xmlfile2")

param_names <- get_param_names_xml(xml_file)

param_names <- get_param_names_xml(xml_files_list)
}

}
