% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_number.R
\name{get_param_number}
\alias{get_param_number}
\title{Get the number of a (or alist of) Stics xml parameter(s) occurence from a request in an xmlDocument}
\usage{
get_param_number(xml_doc_object, param_name, ...)
}
\arguments{
\item{xml_doc_object}{an xmlDocument object (created from an xml file)}

\item{param_name}{parameter name or a vector of names}

\item{...}{other arguments that can be transmitted to get_param_value function (see doc)}
}
\value{
a numeric vector
}
\description{
Get the number of a (or alist of) Stics xml parameter(s) occurence from a request in an xmlDocument
}
\keyword{internal}
