% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_txt.R
\name{get_param_txt}
\alias{get_param_txt}
\title{Read STICS input parameters from text files}
\usage{
get_param_txt(dirpath = getwd(), param = NULL, ...)
}
\arguments{
\item{dirpath}{USM directory path}

\item{param}{Parameter name. Optional, if not provided, the function
return an object with all parameters}

\item{...}{Helper to pass arguments from \code{\link{get_param_txt}} to the
other functions}

\item{variety}{Either the variety name or index for plant parameters (optional, see details).}
}
\value{
A parameter value, or if \code{param= NULL} a list of all parameters:
\item{ini}{Initialization parameters}
\item{general}{General parameters}
\item{tec}{Technical parameters}
\item{plant}{Plant parameters}
\item{soil}{Soil parameters}
\item{station}{Station parameters}
}
\description{
Read STICS model input parameters from a usm in text format (STICS input)
Generally used after calling building a usm with \code{JavaStics}.
}
\note{
Users would generally use \code{get_param_txt} to identify parameters names and values and pass
them to other functions.
}
\examples{
\dontrun{
# Read the interrow distance parameter:

library(SticsRFiles)
path = system.file("extdata/txt/V9.1", package = "SticsRFiles")
get_param_txt(path, param='interrang')

}

}
\seealso{
\code{gen_varmod()},
}
