% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_value.R
\name{get_param_value}
\alias{get_param_value}
\title{Get a set of a (or a list of) Stics xml parameter(s) values from a request}
\usage{
get_param_value(
  xml_doc_object,
  param_name,
  parent_name = NULL,
  parent_sel_attr = NULL,
  ids = NULL,
  show_xpath = FALSE
)
}
\arguments{
\item{xml_doc_object}{an xmlDocument object}

\item{param_name}{parameter name or a vector of names}

\item{parent_name}{parent node name or attribute name (optional)}

\item{parent_sel_attr}{parent attribute value (optional)}

\item{ids}{elements indices (optional)}

\item{show_xpath}{Print the xpath}
}
\value{
a numeric vector values of parameter or a list of
}
\description{
Extracting parameter value from an xml document object
}
\keyword{internal}
