% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params_dict.R
\name{get_params_dict}
\alias{get_params_dict}
\title{Getting an intern dictionnary or merging it with a given dictionnary
given as argument}
\usage{
get_params_dict(in_dict = NULL)
}
\arguments{
\item{in_dict}{a named list with parameters names as values}
}
\value{
A named list with XML parameters names
}
\description{
Getting an intern dictionnary or merging it with a given dictionnary
given as argument
}
\examples{
\dontrun{
# Getting internal dictionnary default content
# get_param_dict()

# Giving a new dictionnary
in_dict <- list(name1 = "param_name1", name2 = "param_name2")

get_params_dict(in_dict)

# Giving a new dictionnary with common values with the internal one
in_dict <- list(name1 = "amount", name2 = "julapI_or_sum_upvt")

get_params_dict(in_dict)

}
}
\keyword{internal}
