% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params_from_doc_attr.R
\name{get_params_from_doc_attr}
\alias{get_params_from_doc_attr}
\title{Get a list of Stics xml parameters names from nodes attributes in
an xml_document}
\usage{
get_params_from_doc_attr(xml_doc, type_name = NULL, unique_val = TRUE)
}
\arguments{
\item{xml_doc}{an xml_document object (created from an xml file)}

\item{type_name}{type name, one of "option", "param", "colonne"}

\item{unique_val}{logical, TRUE to get unique names list, FALSE otherwise}
}
\value{
a named list of parameter names
}
\description{
Get a list of Stics xml parameters names from nodes attributes in
an xml_document
}
\keyword{internal}
