% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params_from_doc_node.R,
%   R/get_params_names.R
\name{get_params_from_doc_node}
\alias{get_params_from_doc_node}
\title{Get a list of Stics xml parameters names from nodes names in an xmlDocument}
\usage{
get_params_from_doc_node(xml_node, param_list = c(), unique_val = TRUE)

get_params_from_doc_node(xml_node, param_list = c(), unique_val = TRUE)
}
\arguments{
\item{xml_node}{an xml XMLInternalElementNode}

\item{param_list}{param names vector, used for recursive calls}

\item{unique_val}{logical, TRUE to get unique names list, FALSE otherwise}

\item{xml_node}{an xml XMLInternalElementNode}

\item{param_list}{param names vector, used for recursive calls}

\item{unique_val}{logical, TRUE to get unique names list, FALSE otherwise}
}
\value{
a character vector of parameters names

a character vector of parameters names
}
\description{
Get a list of Stics xml parameters names from nodes names in an xmlDocument

Get a list of Stics xml parameters names from nodes names in an xmlDocument
}
