% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params_from_table.R
\name{get_params_from_table}
\alias{get_params_from_table}
\title{Get a list of STICS xml parameters names and values from a table
(data.frame, tibble)}
\usage{
get_params_from_table(
  params_table,
  param_names = NULL,
  xml_doc,
  lines_id = NULL,
  stopping = FALSE,
  dict = NULL,
  na_values = NA
)
}
\arguments{
\item{params_table}{a table (df, tibble) containing parameters to use}

\item{param_names}{parameters names to select}

\item{xml_doc}{an xml_document object (of any xml type)}

\item{lines_id}{table lines identifiers to select}

\item{stopping}{logical value for stopping if any unknown parameters
(if TRUE)}

\item{dict}{List of names correspondence between short names (tags)
and real parameters names}

\item{na_values}{value to use as missing value in param_table
(optional, default : NA)}
}
\value{
a named list (with parameters names as list names)
of data.frame/tibble
}
\description{
Get a list of STICS xml parameters names and values from a table
(data.frame, tibble)
}
\examples{
\dontrun{

download_usm_xl(file = "inputs_stics_example.xlsx",
                dest_dir = "/path/to/dest/dir")
xl_path <- file.path("/path/to/dest/dir", "inputs_stics_example.xlsx")
ini_param_df <- read_excel(xl_path, sheet = "Ini")
xml_path <- "path/to/ini/xml"
ini_doc <- SticsRFiles:::xmldocument(xml_path)
get_params_from_table(ini_param_df, ini_doc)
}

}
\keyword{internal}
