% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params_types.R
\name{get_params_types}
\alias{get_params_types}
\title{Get a list of Stics xml parameters types and xpath from a request in an xmlDocument}
\usage{
get_params_types(xml_doc, param_names)
}
\arguments{
\item{xml_doc}{an xmlDocument object (created from an ini file)}

\item{param_names}{parameters names}
}
\value{
a list of list of parameter type, xpath and values number
}
\description{
Get a list of Stics xml parameters types and xpath from a request in an xmlDocument
}
\keyword{internal}
