% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plant_name.R
\name{get_plant_code}
\alias{get_plant_code}
\title{Getting plants code from plant files list by usm}
\usage{
get_plant_code(plant_dir_path, plant_xml_names)
}
\arguments{
\item{plant_dir_path}{vector of plant files directories}

\item{plant_xml_names}{named list of plant files by usm}
}
\value{
a named list of plant codes by usm
}
\description{
Getting plants code from plant files list by usm
}
\keyword{internal}
