% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plant_name.R
\name{get_plant_id}
\alias{get_plant_id}
\title{Getting plant tag(s) from a plant files list by usm}
\usage{
get_plant_id(usms_plant)
}
\arguments{
\item{usms_plant}{a list of plants files names by usm}
}
\value{
a list of plant tags by usm
}
\description{
Getting plant tag(s) from a plant files list by usm
}
\keyword{internal}
