% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plant_name.R
\name{get_plant_name}
\alias{get_plant_name}
\title{Get plant names}
\usage{
get_plant_name(
  workspace,
  usm_name = NULL,
  usms_filename = "usms.xml",
  javastics_path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{workspace}{Path of a JavaStics workspace, or a vector of (recursive call).}

\item{usm_name}{Vector of usms to read (optional, used to filter usms)}

\item{usms_filename}{Name of the usm file (optional, default to the standard \code{usms.xml})}

\item{javastics_path}{JavaStics installation path (Optional, needed if the plant files are not in the \code{workspace}
but rather in the JavaStics default workspace). Only used to get the plants names.}

\item{verbose}{Logical value (optional), TRUE to display infos on error, FALSE otherwise (default)}
}
\value{
A list of plant names for each usm with one value for sole crop and two values for intercrop (principal, associated).
}
\description{
Get the plant name (and file name) for each usm in a workspace
}
\examples{
\dontrun{
path <- system.file(file.path("extdata","obs","V9.0"), package = "SticsRFiles")

# Get plant names for all usms (no javastics path, so only the file name is returned):
get_plant_name(path)
# Get plant names only for banana:
get_plant_name(path, "banana")

# Get plant names for banana with a javastics path, so the plant name is returned:
get_plant_name(path)
# NB: if the plant folder is in the workspace, no need to provide javastics_path

}

}
\keyword{internal}
