% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_v10_v11.R
\name{get_plt_IC_param}
\alias{get_plt_IC_param}
\title{Get crop new parameters for STICS V11}
\usage{
get_plt_IC_param(crop, warning = TRUE)
}
\arguments{
\item{crop}{Stics crop code among code list
(to get crops codes get_plt_IC_param() )}

\item{warning}{Logical for rising warnings, FALSE otherwise}
}
\value{
A named list of V11 new plant parameters values
}
\description{
Get crop new parameters for STICS V11
}
\examples{
# get plant codes
get_plt_IC_param()
# get default parameters values
get_plt_IC_param(NULL)
# get parameters values for poi
get_plt_IC_param("poi")

}
