% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get_referenced_dirs}
\alias{get_referenced_dirs}
\title{Getting valid directories string for download from SticsRpacks data repository}
\usage{
get_referenced_dirs(dirs = NULL, stics_version = NULL)
}
\arguments{
\item{dirs}{Directories names of the referenced use cases (optional), starting with "study_case_"}

\item{stics_version}{An optional version string
within those given by get_stics_versions_compat()$versions_list}
}
\value{
Vector of referenced directories string (as "study_case_1/V9.0")
}
\description{
Getting valid directories string for download from SticsRpacks data repository
}
\examples{
\dontrun{
# Getting all available dirs from the data repos
SticsRFiles:::get_referenced_dirs()

# Getting dirs for a use case
SticsRFiles:::get_referenced_dirs("study_case_1")

# Getting dirs for a use case and a version
SticsRFiles:::get_referenced_dirs("study_case_1", "V9.0" )

SticsRFiles:::get_referenced_dirs(c("study_case_1", "study_case_2"), "V9.0" )

}

}
\keyword{internal}
