% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_stics_versions.R
\name{get_svn_identifiers}
\alias{get_svn_identifiers}
\title{Getting read only account identifiers for the STICS subversion repository}
\usage{
get_svn_identifiers()
}
\value{
A list with fields \code{username} and \code{password}
}
\description{
Getting read only account identifiers for the STICS subversion repository
}
\examples{
\dontrun{
SticsRFiles:::get_svn_identifiers()
}
}
\keyword{internal}
