% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_txt_param_value.R
\name{get_txt_param_value}
\alias{get_txt_param_value}
\title{Getting parameters values from a Stics text file}
\usage{
get_txt_param_value(file_path, param_names, file_type, names_dict = NULL)
}
\arguments{
\item{file_path}{Stics text file path}

\item{param_names}{vector of parameters names}

\item{file_type}{type of file tag from "name_value", "soil", "ini", "tec"}

\item{names_dict}{List of correspondance between parameter names given and files intern names}
}
\value{
A vector of parameters values
}
\description{
Getting parameters values from a Stics text file
}
\keyword{internal}
