% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usms_list.R
\name{get_usms_list}
\alias{get_usms_list}
\title{Getting usms names list for an usms.xml file}
\usage{
get_usms_list(usm_path, name = NULL)
}
\arguments{
\item{usm_path}{usm file path (e.g. "usms.xml")}

\item{name}{A vector of usm name(s) or partial name(s), optional}
}
\value{
A vector of usm names
}
\description{
Extracting a usm names list from an usms.xml file
}
\examples{
path = get_examples_path( file_type = "xml")

usms_list <- get_usms_list(usm_path = file.path(path, "usms.xml"))

usms_list <- get_usms_list(usm_path = file.path(path, "usms.xml"), name = c("usm1", "usm2"))

}
