% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values_by_param.R
\name{get_values_by_param}
\alias{get_values_by_param}
\title{Extract a named list of parameters values or a list of
from data.frame row(s), for all parameters columns or a selection}
\usage{
get_values_by_param(
  params_table,
  param_name = NULL,
  lines_id = NULL,
  na_values = NA
)
}
\arguments{
\item{params_table}{a table (df, tibble) with parameters names
indexed with suffixes _1, _2 ... i.e. paramname_1, paramname_2
or not for scalar parameters}

\item{param_name}{Optional arameter(s) name(s) vector}

\item{lines_id}{Optional data.frame lines numerical identifiers}

\item{na_values}{value to use as missing value in param_table (optional, default : NA)}
}
\value{
a named list of parameters values, with sorted values for multiple
parameter values (indexed with suffixes _1, _2, ...)
}
\description{
Extract a named list of parameters values or a list of
from data.frame row(s), for all parameters columns or a selection
}
\examples{
\dontrun{
download_usm_xl(file = "inputs_stics_example.xlsx", dest_dir = "/path/to/dest/dir")
xl_path <- file.path("/path/to/dest/dir", "inputs_stics_example.xlsx")
tec_param_df <- read_excel(xl_path, sheet = "Tec")
SticsRFiles:::get_values_by_param(params_table = tec_param_df)
}

}
\keyword{internal}
