% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values_from_table.R
\name{get_values_from_table}
\alias{get_values_from_table}
\title{Get a list of Stics xml parameters values from a table (data.frame, tibble)}
\usage{
get_values_from_table(params_table, param_names = NULL, lines_id = NULL)
}
\arguments{
\item{params_table}{a table (df, tibble) containing parameters to use}

\item{param_names}{parameters names}

\item{lines_id}{lines ids for selecting}
}
\value{
a named list (with param names as list names) of data.frame/tibble
}
\description{
Get a list of Stics xml parameters values from a table (data.frame, tibble)
}
\examples{
\dontrun{
copy_mailing_example(xl_name = "inputs_stics_example.xlsx", dest_dir = "/path/to/dest/dir")
xl_path <- file.path("/path/to/dest/dir","inputs_stics_example.xlsx")
sols_param_df <- read_excel(xl_path, sheet = "Soils")

SticsRFiles:::get_values_from_table(sols_param_df)

SticsRFiles:::get_values_from_table(sols_param_df, param_names = c("DAF","HMINF"))

SticsRFiles:::get_values_from_table(sols_param_df, lines_id = c(1,10))

SticsRFiles:::get_values_from_table(sols_param_df,
param_names = c("DAF","HMINF"), lines_id = c(1,10))

}

}
\keyword{internal}
