% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_mod.R
\name{get_var_mod}
\alias{get_var_mod}
\title{Get desired STICS outputs}
\usage{
get_var_mod(...)
}
\arguments{
\item{workspace}{Path of the directory containing the Stics var.mod file to modify}

\item{usms_file}{Path (including name) of a USM XML file.}

\item{file_name}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{file_name} is no
longer supported, use \code{usms_file} instead.}
}
\value{
The variables that will be returned by STICS
}
\description{
Get the variables STICS will return ("var.mod" file)

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated to homogenised writing process
}
\examples{
get_varmod(get_examples_path( file_type = "txt" ))


 \dontrun{
get_var_mod(workspace,usms_file)
# ->
get_varmod(workspace,usms_file)
}

}
\seealso{
\code{gen_varmod}
}
\keyword{internal}
