% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_varmod.R
\name{get_varmod}
\alias{get_varmod}
\title{Get desired STICS outputs}
\usage{
get_varmod(workspace, file_name = "var.mod")
}
\arguments{
\item{workspace}{Path of the directory containing the STICS var.mod file}

\item{file_name}{file name to read (without path, default value: "var.mod")}
}
\value{
The variables that will be returned by STICS
}
\description{
Get the STICS output variables (from var.mod file)
}
\examples{
get_varmod(get_examples_path(file_type = "txt"))
}
\seealso{
\code{gen_varmod}
}
