% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_stics_versions.R
\name{get_version_info_tmpl}
\alias{get_version_info_tmpl}
\title{Get a basic data.frame for a version}
\usage{
get_version_info_tmpl(version_name)
}
\arguments{
\item{version_name}{name (i.e. "VX.Y") of the version to set
in the output data.frame}
}
\value{
a data.frame, with version set in \code{versions} and \code{csv} columns
(for existing inputs.csv and outputs.csv files in a versions VX.Y
sub-directory)
}
\description{
Get a basic data.frame for a version
}
\examples{
\dontrun{
SticsRFiles:::get_versions_info(version_name = "V10.0")

>
>  versions   csv obs sti txt xml xml_tmpl xl study_case_1
  1    V10.0 V10.0

}
}
\keyword{internal}
