% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stics_versions_compat.R
\name{get_versions_file_name}
\alias{get_versions_file_name}
\title{Getting the csv file name storing versions information}
\usage{
get_versions_file_name()
}
\value{
file name
}
\description{
Getting the csv file name storing versions information
}
\examples{
\dontrun{
SticsRFiles:::get_versions_file_name()
}
}
\keyword{internal}
