% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_stics_versions.R
\name{get_versions_file_path}
\alias{get_versions_file_path}
\title{Getting versions information csv file path according to its location}
\usage{
get_versions_file_path(location = "install")
}
\arguments{
\item{location}{The destination where to remove information and data
"install" for removing things from the installed SticsRFiles library
(default), "package" for removing them from the package project (in RStudio)}
}
\value{
A file path
}
\description{
Getting versions information csv file path according to its location
}
\examples{
\dontrun{
SticsRFiles:::get_versions_file_path()

SticsRFiles:::get_versions_file_path(location = "package")
}
}
\keyword{internal}
