% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stics_versions_compat.R
\name{get_versions_info}
\alias{get_versions_info}
\title{Getting versions data (versions strings and examples files directories list)}
\usage{
get_versions_info(stics_version = NULL, location = "install")
}
\arguments{
\item{stics_version}{Optional version string (i.e. "VX.Y")}

\item{versions_dir}{Optional, either an \code{extdata} directory path
of the installed SticsRFiles library (default) or of the package project}
}
\value{
A data.frame with versions data
}
\description{
Getting versions data (versions strings and examples files directories list)
}
\examples{
\dontrun{

SticsRFiles:::get_versions_info()

get_versions_info(stics_version = "V8.5")


get_versions_info(
  stics_version = "V8.5",
  versions_dir = "path/to/SticsRFilesproject/inst/extdata"
)
}
}
\keyword{internal}
