% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xml_base_doc.R
\name{get_xml_base_doc}
\alias{get_xml_base_doc}
\title{Get an xmlDocument from a Stics xml file template}
\usage{
get_xml_base_doc(xml_type = NULL, stics_version = "last")
}
\arguments{
\item{xml_type}{xml file type (see types returned when calling get_xml_base_doc())}

\item{stics_version}{the stics files version to use}
}
\value{
an xmlDocument object
}
\description{
Get an xmlDocument from a Stics xml file template
}
