% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xml_base_node.R
\name{get_xml_base_node}
\alias{get_xml_base_node}
\title{Get an xmlDocument node (set of parameters) from a Stics xml node file template}
\usage{
get_xml_base_node(file_tag, form_name = NULL, stics_version = "last")
}
\arguments{
\item{file_tag}{file tag among "usms","sols", "tec"}

\item{form_name}{formalism name}

\item{stics_version}{the stics files version to use}
}
\description{
Get an xmlDocument node (set of parameters) from a Stics xml node file template
}
