% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xml_files_param_df.R
\name{get_xml_files_param_df}
\alias{get_xml_files_param_df}
\title{Extracting a data.frame of parameters values for an xml file or a set of (whatever the xml file kind)}
\usage{
get_xml_files_param_df(
  file_path,
  select = NULL,
  name = NULL,
  param_names = NULL
)
}
\arguments{
\item{file_path}{A file path or a vector of}

\item{select}{node name or attribute name to use for selection (optional, default to no selection)}

\item{name}{value used for select (optional)}

\item{param_names}{vector of parameters names (optional)}
}
\value{
A data.frame with name, type, param, id and value columns
}
\description{
Extracting a data.frame of parameters values for an xml file or a set of (whatever the xml file kind)
}
\details{
An extract of an example of data.frame\tabular{lllrl}{
   name \tab type \tab param \tab id \tab value \cr
   SugarCane \tab usms \tab datedebut \tab NA \tab 286 \cr
   SugarCane \tab usms \tab datefin \tab NA \tab 650 \cr
   SugarCane \tab usms \tab finit \tab NA \tab canne_ini.xml \cr
   SugarCane \tab usms \tab nomsol \tab NA \tab solcanne \cr
   SugarCane \tab usms \tab fstation \tab NA \tab climcanj_sta.xml \cr
   ... \tab ... \tab ... \tab ... \tab ... \cr
   ... \tab ... \tab ... \tab ... \tab ... \cr
   param_gen.xml \tab fichierpar \tab Qmulchdec \tab 17 \tab 0 \cr
   param_gen.xml \tab fichierpar \tab Qmulchdec \tab 18 \tab 0 \cr
   param_gen.xml \tab fichierpar \tab Qmulchdec \tab 19 \tab 0 \cr
   param_gen.xml \tab fichierpar \tab Qmulchdec \tab 20 \tab 0 \cr
   param_gen.xml \tab fichierpar \tab Qmulchdec \tab 21 \tab 0 \cr
}


name: a file name or an usm or soil name
type: type of file
param: param name
id: NA for scalar, integer id for vectors parameters
value: parameter value
}
\examples{

\dontrun{
dir_path <- get_examples_path("xml")
get_xml_files_param_df(file_path = file.path(dir_path,"sols.xml"))

get_xml_files_param_df(file_path = file.path(dir_path,"sols.xml"),
select = "sol", c("solcanne", "solble"))

files_list <- file.path(dir_path, c("sols.xml", "usms.xml", "param_gen.xml" ))
get_xml_files_param_df(file_path = files_list)

}


}
\keyword{internal}
