% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xml_stics_version.R
\name{get_xml_stics_version}
\alias{get_xml_stics_version}
\title{Get a version/or the laste version string of available Stics files templates in the package}
\usage{
get_xml_stics_version(stics_version = "last", xml_doc = NULL)
}
\arguments{
\item{stics_version}{A version key of a Stics version (i.e. V9.0)}

\item{xml_doc}{an xmlDocument of a Stics xml file (Unused for the moment, no version String included in xml files)}
}
\value{
a Stics version string
}
\description{
Get a version/or the laste version string of available Stics files templates in the package
}
\keyword{internal}
