% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_climate.R
\name{get_year_from_file_name}
\alias{get_year_from_file_name}
\title{Getting years from file(s) names(s)}
\usage{
get_year_from_file_name(file_path)
}
\arguments{
\item{file_path}{a vector of weather file path(s)}
}
\value{
a vector of numeric years
}
\description{
Getting years from file(s) names(s)
}
\examples{
#'
\dontrun{
get_year_from_file_name(
  c("path/to/weather.year1", "path/to/weather.year2")
)
# should return c(year1, year2)

get_year_from_file_name(
  c("path/to/weather.year1", "path/to/weather.year3")
)
# should return c(year1, year2, year3)
}

}
\keyword{internal}
