% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_years.R
\name{get_years}
\alias{get_years}
\title{Extract USM years}
\usage{
get_years(workspace, file_name = "usms.xml", usm)
}
\arguments{
\item{workspace}{Path of the workspace containing the STICS files to read}

\item{file_name}{Name of the XML file listing the USMs (default to \code{usms.xml}).}

\item{usm}{Character vector of USM name(s) for which the years will be extracted (default to all USMs).}
}
\value{
A data.frame of both years for each requested USM.
}
\description{
Extract USM years ("usms.xml")
}
