% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_day_number.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Leap year}
\usage{
is_leap_year(year)
}
\arguments{
\item{year}{numeric}
}
\value{
logical
}
\description{
Returns TRUE if the given year is a leap year, FALSE otherwise.
}
\author{
Timothee Flutre
}
