% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_stics_doc.R
\name{is_stics_doc}
\alias{is_stics_doc}
\alias{is_stics_ini}
\alias{is_stics_usms}
\alias{is_stics_sols}
\alias{is_stics_tec}
\alias{is_stics_sta}
\alias{is_stics_plt}
\alias{is_stics_par}
\alias{is_stics_newpar}
\title{Is it a STICS document}
\usage{
is_stics_doc(xml_doc, doc_type = NULL, doc_types = NULL)

is_stics_ini(xml_doc)

is_stics_usms(xml_doc)

is_stics_sols(xml_doc)

is_stics_tec(xml_doc)

is_stics_sta(xml_doc)

is_stics_plt(xml_doc)

is_stics_par(xml_doc)

is_stics_newpar(xml_doc)
}
\arguments{
\item{xml_doc}{An xml document}

\item{doc_type}{The type of xml document}

\item{doc_types}{The different possible document types (optionnal)}
}
\description{
Test if an XML document is a STICS document.
}
