% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_in_par.R
\name{is_stics_param}
\alias{is_stics_param}
\title{Search if a STICS parameter exist}
\usage{
is_stics_param(param, stics_version = "latest")
}
\arguments{
\item{param}{A vector of parameter names}

\item{stics_version}{Name of the STICS version. Optional, can be used
to search parameters information relative to a specific STICS version.
By default the latest version returned by \code{get_stics_versions_compat()}
is used.}
}
\value{
A boolean vector: \code{TRUE} if the parameter exist, \code{FALSE} otherwise
}
\description{
Tells if one or more parameter names are valid STICS
input parameters.
}
\examples{
is_stics_param(c("adil", "adilmax", "unknown"))

}
\seealso{
\code{get_param_info()} for interactive use.
}
