% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_out_var.R
\name{is_stics_var}
\alias{is_stics_var}
\title{Search if a STICS variable exist}
\usage{
is_stics_var(var, stics_version = "latest")
}
\arguments{
\item{var}{A vector of variable names}

\item{stics_version}{Name of the STICS version. Optional, can be used
to search parameters information relative to a specific STICS version.
By default the latest version returned by \code{get_stics_versions_compat()}
is used.}
}
\value{
A boolean vector: \code{TRUE} if the variable exist, \code{FALSE} otherwise
}
\description{
Tells if one or more variable names are valid STICS
output variables.
}
\examples{
is_stics_var(c("lai(n)", "masec(n)", "unknown"))

}
\seealso{
\code{get_var_info()} for interactive use.
}
