% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_stics_xml.R
\name{is_stics_xml}
\alias{is_stics_xml}
\alias{is_ini_xml}
\alias{is_usms_xml}
\alias{is_sols_xml}
\alias{is_tec_xml}
\alias{is_sta_xml}
\alias{is_plt_xml}
\alias{is_par_xml}
\alias{is_newpar_xml}
\title{Is it a STICS xml file}
\usage{
is_stics_xml(xml_path, file_type = NULL)

is_ini_xml(xml_path)

is_usms_xml(xml_path)

is_sols_xml(xml_path)

is_tec_xml(xml_path)

is_sta_xml(xml_path)

is_plt_xml(xml_path)

is_par_xml(xml_path)

is_newpar_xml(xml_path)
}
\arguments{
\item{xml_path}{An xml file path}

\item{file_type}{The type of the xml file}
}
\value{
A logical value giving if input xml file is a STICS one
(TRUE), or not (FALSE)
}
\description{
Test if an XML file is a STICS one.
}
\examples{
\dontrun{
xml_path <- file.path(get_examples_path(file_type = "xml"), "sols.xml")
SticsRFiles:::is_stics_xml(xml_path)
}

}
\keyword{internal}
