% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_os_name.R
\name{is_windows}
\alias{is_windows}
\title{Evaluating if the OS is a windows type}
\usage{
is_windows()
}
\value{
TRUE/FALSE
}
\description{
Evaluating if the OS is a windows type
}
\examples{
is_windows()
}
