% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_to_julian_dates.R
\name{isLeapYear}
\alias{isLeapYear}
\title{Leap year}
\usage{
isLeapYear(year)
}
\arguments{
\item{year}{numeric}
}
\value{
logical
}
\description{
Returns TRUE if the given year is a leap year, FALSE otherwise
}
\author{
Timothee Flutre
}
