% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_to_julian_dates.R
\name{julian_date_for_stics}
\alias{julian_date_for_stics}
\title{Convert dates}
\usage{
julian_date_for_stics(date, start_date)
}
\arguments{
\item{date}{date you wish to convert, in the \code{\link[base]{Date}} format}

\item{start_date}{date used as the reference point, in the \code{\link[base]{Date}} format, typically the simulation start year}
}
\value{
numeric
}
\description{
Converts a calendar date into a Julian date using a start date, as used for STICS, that is, Julian day counted from January 1 of the year of the start date, typically the simulation start year.
Leap years are properly handled.
}
\examples{
start_date <- as.Date("2014-08-01")
date <- as.Date("2015-02-10")
julian_date_for_stics(date=date, start_date=start_date)

start_date <- as.Date("2008-08-01")
date <- as.Date("2009-02-10")
julian_date_for_stics(date=date, start_date=start_date)
}
\author{
Timothee Flutre
}
