% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_nodesets.R
\name{merge_nodesets}
\alias{merge_nodesets}
\title{Merging 2 XML node sets}
\usage{
merge_nodesets(ns1, ns2)
}
\arguments{
\item{ns1}{First XMLNodeSet}

\item{ns2}{Second XMLNodeSet}
}
\value{
An XMLNodeSet
}
\description{
Merging 2 XML node sets
}
\keyword{internal}
