% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_out_var.R
\docType{data}
\name{out_data}
\alias{out_data}
\title{Output variables data}
\format{A data frame with seven columns:
\describe{
\item{variable}{Variable name}
\item{details}{a brief definition of the variable}
\item{unit}{unit of the variable}
\item{file}{the file where the variable comes from}
\item{type}{the data type}
\item{V1}{ }
\item{V2}{ }
}}
\usage{
out_data
}
\description{
All output variables from STICS, with their units and brief description
}
\keyword{datasets}
