% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{parse_mixed_file}
\alias{parse_mixed_file}
\title{Get mixed file names}
\usage{
parse_mixed_file(file_names, type = c("sim", "obs"))
}
\value{
A list of observation or simulation files associated to their usm name. The mixed
crops are always returned as c("Principal","Associated").
}
\description{
Get mixed observation or simulation files by name
}
\note{
The function use the obs/sim files names to retrieve the usm name.
So each obs file should be named with the usm name, followed by a or p
at the end in the case of associated crops.
}
\examples{
\dontrun{
parse_mixed_file(list("banana.obs", "IC_banana_sorghuma.obs", "IC_banana_sorghump.obs"),
                  type = "obs")

# Simulations with usm names starting with "a", with or without intercropping:
file_names = list("mod_sauzevilleSC_Wheat_Wheat_2005-2006_N0.sti",
"mod_saIC_Wheat_Wheat_2005-2006_N0.sti", "mod_spIC_Wheat_Wheat_2005-2006_N0.sti")

parse_mixed_file(file_names, type = "sim")

}

}
\keyword{internal}
