% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_obs_to_list.R
\name{read_obs_to_list}
\alias{read_obs_to_list}
\title{Read STICS observation files (*.obs) and store data into a list}
\usage{
read_obs_to_list(dirpath = getwd(), obs_filenames = NULL,
  usms = NULL, usms_filename = "usms.xml")
}
\arguments{
\item{dirpath}{Directory path}

\item{obs_filenames}{A vector of observation file name(s). Optional, see details.}

\item{usms}{A vector of usm name(s) corresponding the the obs files. Optional, see details.}

\item{usms_filename}{name of the usm file. Optional, see details.}
}
\value{
A list of STICS-formated observations. Return \code{NULL} if no files were found.
}
\description{
Read STICS observation files for sole crops and store data into a list
}
\details{
If \code{filename} is not specified (or equal to \code{NULL}), the
         function read all obs files in the \code{dirpath} folder. If there are no .obs files,
         the function returns \code{NULL}
         If usms names are not provided they are supposed to have the same name as the obs file.
         If usms_filename is not provided, existence of usms are checked in the usms.xml file.
}
\examples{
\dontrun{
library(sticRFiles)
Meas= read_obs_to_list()
}

}
\seealso{
\code{\link{read_obs}}
}
