% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_parent_from_doc.R
\name{remove_parent_from_doc}
\alias{remove_parent_from_doc}
\title{Remove parent node}
\usage{
remove_parent_from_doc(
  xml_doc,
  param_name,
  parent_path = NULL,
  remove_parent = FALSE,
  nodes_ids = NULL
)
}
\arguments{
\item{xml_doc}{The XML document}

\item{param_name}{The parent name}

\item{parent_path}{The parent path}

\item{remove_parent}{Boolean. Do we remove the parent node?}

\item{nodes_ids}{The requested node IDs.}
}
\description{
Remove a parent node from an XML file.
}
\keyword{internal}
