% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_string_in_file.R
\name{replace_string_in_file}
\alias{replace_string_in_file}
\title{Replacing a string in a file}
\usage{
replace_string_in_file(file_path, target_string, replace_string,
  new_file_path = NULL)
}
\description{
Replacing occurrences of a string by another one, in place or creating a new file
}
