% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_java_workspace.R
\name{set_java_workspace}
\alias{set_java_workspace}
\title{Setting JavaSTICS workspace}
\usage{
set_java_workspace(javastics, workspace)
}
\arguments{
\item{javastics}{JavaSTICS installation root folder}

\item{workspace}{JavaSTICS working directory
(absolute,relative to javastics path)}
}
\description{
Setting a new JavaSTICS working directory, a relative directory
to JavaSTICS path or an absolute one
}
\details{
Checking if the directory is a JavaSTICS workspace
(any usms.xml file), and if it's already registered

before setting new one
}
\examples{
\dontrun{
set_java_workspace("/path/to/JavaSTICS/folder", "my_wd")
set_java_workspace("/path/to/JavaSTICS/folder", "/path/to/my_wd")
}

}
\keyword{internal}
