% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javastics_path.R
\name{set_javastics_path}
\alias{set_javastics_path}
\title{Set JavaSTICS path in \code{javastics_path} environment variable}
\usage{
set_javastics_path(javastics_path, write = FALSE)
}
\arguments{
\item{javastics_path}{Absolute path of the root directory of JavaSTICS}

\item{write}{boolean, TRUE if \code{javastics_path} has to be written in
a \code{.Renviron} file, FALSE otherwise}
}
\description{
Set JavaSTICS path in \code{javastics_path} environment variable
}
\keyword{internal}
