% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sols_param_xml.R
\name{set_sols_param_xml}
\alias{set_sols_param_xml}
\title{Setting a sol param value(s) in a sols xmlDocument}
\usage{
set_sols_param_xml(xml_doc_object, sols_param, overwrite = FALSE)
}
\arguments{
\item{xml_doc_object}{an xmlDocument object (created from an xml sols file)}

\item{sols_param}{soils parameters}

\item{overwrite}{replace existing soil (TRUE) or not, updating existing ones (FALSE)}
}
\description{
Setting a sol param value(s) in a sols xmlDocument
}
\keyword{internal}
