% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_usms_param.R
\name{set_usms_param}
\alias{set_usms_param}
\title{Setting a usm param value(s) in an usms xmlDocument}
\usage{
set_usms_param(xml_doc_object, usms_param, overwrite = FALSE)
}
\arguments{
\item{xml_doc_object}{an xmlDocument object (created from an usms file)}

\item{usms_param}{usms parameters}

\item{overwrite}{replace existing soil (TRUE) or not, updating existing ones (FALSE)}
}
\description{
Setting a usm param value(s) in an usms xmlDocument
}
