% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_stics_versions.R
\name{set_versions_info}
\alias{set_versions_info}
\title{Writing information about Stics versions and related example files
directories in the SticsRFiles library or package}
\usage{
set_versions_info(
  version_name,
  versions_dir = system.file("extdata", package = "SticsRFiles"),
  verbose = TRUE
)
}
\arguments{
\item{version_name}{name (i.e. "VX.Y") of the version to add in versions
information}

\item{versions_dir}{Optional, , either an \code{extdata} directory path
of the installed SticsRFiles library (default) or of the package project}

\item{verbose}{Logical, TRUE for displaying warnings, FALSE otherwise}
}
\description{
Writing information about Stics versions and related example files
directories in the SticsRFiles library or package
}
\keyword{internal}
