% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_params_table.R
\name{sort_params_table}
\alias{sort_params_table}
\title{Sort a table against column names indexes with a suffix}
\usage{
sort_params_table(params_table)
}
\arguments{
\item{params_table}{a table (df, tibble) with parameters names
indexed with suffixes _1, _2 ... i.e. paramname_1, paramname_2}
}
\description{
Sort a table against column names indexes with a suffix
}
