% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_help.R
\name{static_help}
\alias{static_help}
\title{Generate html help from functions Rd help files}
\usage{
static_help(
  pkg,
  links_level = 0,
  topic = NULL,
  out_dir = getwd(),
  overwrite = TRUE
)
}
\arguments{
\item{pkg}{Package name}

\item{links_level}{integer, 0: inside the given package, 1: with base and
recommended packages, 2: other packages specified by .libPaths()}

\item{topic}{Optional, selecting specific topics with their names vector (functions names)}

\item{out_dir}{Optional, where to store html functions help files}

\item{overwrite}{Optional, logical TRUE for overwriting hmtl files (default), FALSE otherwise}
}
\value{
An invisible character vector of file paths
}
\description{
Generate html help from functions Rd help files
}
\examples{
\dontrun{
static_help("SticsRFiles")
static_help(pkg = "SticsRFiles", out_dir = "/path/to/out/dir")
}
}
\keyword{internal}
