% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_environment.R
\name{stics_class}
\alias{stics_class}
\title{Get the class of a STICS object}
\usage{
stics_class(name, env_name = sticsenv_name())
}
\value{
A vector of classes
}
\description{
Get the class of a STICS object
}
\keyword{internal}
