% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_environment.R
\name{stics_clean}
\alias{stics_clean}
\title{Clean STICS environment}
\usage{
stics_clean()
}
\value{
Nothing. Just cleans the STICS environment.
}
\description{
Clean STICS environment
}
\examples{
\dontrun{
stics_clean()
}
}
\keyword{internal}
