% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_environment.R
\name{stics_env}
\alias{stics_env}
\title{Don't know what purpose this function serve}
\usage{
stics_env(name = NULL, env_name = globalenv(), create = TRUE)
}
\arguments{
\item{name}{Name of an object in an env OR env name ??? (optional)}

\item{env_name}{The environment name}

\item{create}{Create the environment name}
}
\value{

}
\description{
Too complicated, does so many things I don't know what to tell
about this function. Patrice, can you write the doc, or simplify
the interface ?
}
\keyword{internal}
