% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_environment.R
\name{stics_remove}
\alias{stics_remove}
\title{Remove objects from STICS environment}
\usage{
stics_remove(name = NULL, env_name = sticsenv_name())
}
\value{
Nothing. Just removes an object from the STICS environment.
}
\description{
Remove objects listed in a STICS environment.
}
\examples{
\dontrun{
stics_remove()
}
}
\keyword{internal}
