% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_environment.R
\name{sticsenv_create}
\alias{sticsenv_create}
\title{Create a new stics environment}
\usage{
sticsenv_create(name, env_name = ".GlobalEnv")
}
\arguments{
\item{name}{The name of the stics environment}

\item{env_name}{The name of the parent environment}
}
\value{
The new environment
}
\description{
Create a new stics environment
}
\keyword{internal}
