% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_environment.R
\name{sticsenv_get_name}
\alias{sticsenv_get_name}
\title{Get stics env name from object}
\usage{
sticsenv_get_name(name = NULL, env_name = sticsenv_name())
}
\arguments{
\item{name}{The name of an object potentially found in a stics environment}

\item{env_name}{The name of the stics environment}
}
\value{
A vector of environment names
}
\description{
Get the name of the environment where an object is found
}
\keyword{internal}
