% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_environment.R
\name{sticsenv_name}
\alias{sticsenv_name}
\title{stics env name}
\usage{
sticsenv_name()
}
\value{
The default stics env name
}
\description{
Get the default stics environment name
}
\keyword{internal}
