% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{[.cropr_simulation}
\alias{[.cropr_simulation}
\title{\code{[} method for \code{cropr_simulation}}
\usage{
\method{[}{cropr_simulation}(x, ...)
}
\arguments{
\item{x}{A \code{cropr_simulation} list}

\item{...}{An index}
}
\value{
A subset of a \code{cropr_simulation}, keeping its attribute
}
\description{
This method ensure keeping the \code{cropr_simulation} attribute when subsetting a
\code{cropr_simulation} list.
}
\examples{
path <- file.path(get_examples_path("sti"), "workspace1")
sim <- SticsRFiles::get_sim(workspace = path)
# sim returns a `cropr_simulation` list

}
