% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_examples_utils.R
\name{unzip_examples}
\alias{unzip_examples}
\title{Unzip files archive if needed and return examples files path
in extdata directory}
\usage{
unzip_examples(files_type, version_dir)
}
\arguments{
\item{version_dir}{version directory names of the example files}

\item{examples_type_path}{library path for examples files set}
}
\value{
library examples files path
}
\description{
Unzip files archive if needed and return examples files path
in extdata directory
}
\keyword{internal}
