% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_stics_versions.R
\name{update_stics_version}
\alias{update_stics_version}
\title{Updating csv files for a given version}
\usage{
update_stics_version(
  version_name,
  url,
  file_name = "all",
  dest = "install",
  verbose = FALSE
)
}
\arguments{
\item{version_name}{name (i.e. "VX.Y") of the version}

\item{url}{Subversion repository address of the branch, tag to get
information from}

\item{file_name}{File name(s)}

\item{dest}{The destination where to write information "install" for writing
things in the installed SticsRFiles library (default), "package" for writing
them in the package project (in RStudio)}

\item{verbose}{Logical, TRUE for displaying warnings, FALSE otherwise}
}
\value{
An invisible data.frame containing versions data in the
SticsRFiles library (or package)
}
\description{
Updating csv files for a given version
}
\examples{
\dontrun{
SticsRFiles::update_stics_version(version_name = "V10.0",
 url = "https://w3.avignon.inra.fr/svn/modulostics/branches/branch10")

}
}
