% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_xml_files.R
\name{upgrade_param_gen_xml}
\alias{upgrade_param_gen_xml}
\title{Upgrade a param_gen.xml file to a next major version}
\usage{
upgrade_param_gen_xml(file, out_dir, ..., overwrite = FALSE)
}
\arguments{
\item{file}{xml param_gen file path or a vector of}

\item{out_dir}{Output directory path}

\item{...}{additional argument(s) to pass to called functions}

\item{overwrite}{Logical TRUE for overwriting files,
FALSE otherwise (default)}
}
\value{
None
}
\description{
Upgrade a param_gen.xml file to a next major version
}
\examples{
dir_path <- get_examples_path(file_type = "xml", stics_version = "V10.0")

upgrade_param_gen_xml(
  file = file.path(dir_path, "param_gen.xml"),
  out_dir = tempdir()
)

}
