% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_v10_v11.R
\name{upgrade_param_gen_xml_10_11}
\alias{upgrade_param_gen_xml_10_11}
\title{Upgrade a param_gen.xml file from STICS version 10 to 11}
\usage{
upgrade_param_gen_xml_10_11(file, out_dir, ..., overwrite = FALSE)
}
\arguments{
\item{file}{xml param_gen file path or a vector of}

\item{out_dir}{Output directory path}

\item{...}{additional argument(s) to pass}

\item{overwrite}{Logical TRUE for overwriting files,
FALSE otherwise (default)}
}
\value{
None
}
\description{
Upgrade a param_gen.xml file from STICS version 10 to 11
}
\examples{
\dontrun{
dir_path <- get_examples_path(file_type = "xml", stics_version = "V10.0")

upgrade_param_gen_xml_10_11(
  file = file.path(dir_path, "param_gen.xml"),
  out_dir = tempdir()
)
}
}
\keyword{internal}
