% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_v9_v10.R
\name{upgrade_param_gen_xml_9_10}
\alias{upgrade_param_gen_xml_9_10}
\title{Upgrade a param_gen.xml file from STICS version 9 to 10}
\usage{
upgrade_param_gen_xml_9_10(file, out_dir, overwrite = FALSE)
}
\arguments{
\item{file}{Path of a param_gen.xml file}

\item{out_dir}{Output directory path of the generated file}

\item{overwrite}{logical (optional),
TRUE for overwriting file if it exists, FALSE otherwise}
}
\value{
None
}
\description{
Upgrade a param_gen.xml file from STICS version 9 to 10
}
\details{
See get_stics_versions_compat() for listing versions
}
\examples{
dir_path <- get_examples_path(file_type = "xml", stics_version = "V9.2")

upgrade_param_gen_xml_9_10(
  file = file.path(dir_path, "param_gen.xml"),
  out_dir = tempdir()
)
}
