% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_v10_v11.R
\name{upgrade_param_newform_xml_10_11}
\alias{upgrade_param_newform_xml_10_11}
\title{Upgrade xml param_newform file from STICS version 10 to 11}
\usage{
upgrade_param_newform_xml_10_11(file, out_dir, overwrite = FALSE)
}
\arguments{
\item{file}{xml param_newform file path or a vector of}

\item{out_dir}{Output directory path}

\item{overwrite}{Logical TRUE for overwriting files,
FALSE otherwise (default)}
}
\value{
None
}
\description{
Upgrade xml param_newform file from STICS version 10 to 11
}
