% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_xml_files.R
\name{upgrade_plt_xml}
\alias{upgrade_plt_xml}
\title{Upgrade _plt.xml file(s) to a next major version}
\usage{
upgrade_plt_xml(
  file,
  out_dir,
  param_gen_file,
  param_newform_file,
  ...,
  overwrite = FALSE,
  warning = TRUE
)
}
\arguments{
\item{file}{xml plant file path or a vector of}

\item{out_dir}{Output directory path}

\item{param_gen_file}{Path of the param_gen.xml file corresponding
to the initial version}

\item{param_newform_file}{Path of the param_newform.xml file corresponding
to the file version}

\item{...}{additional argument(s) to pass to called functions}

\item{overwrite}{Logical TRUE for overwriting files,
FALSE otherwise (default)}

\item{warning}{Logical for rising warnings, FALSE otherwise}
}
\value{
None
}
\description{
Upgrade _plt.xml file(s) to a next major version
}
\examples{

dir_path <- get_examples_path(file_type = "xml", stics_version = "V10.0")

upgrade_plt_xml(
  file = file.path(dir_path, "file_plt.xml"),
  out_dir = tempdir(),
  param_gen_file = file.path(dir_path, "param_gen.xml"),
  param_newform_file = file.path(dir_path, "param_newform.xml")
)

}
