% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_v10_v11.R
\name{upgrade_plt_xml_10_11}
\alias{upgrade_plt_xml_10_11}
\title{Upgrade _plt.xml file(s) from version 10 to 11}
\usage{
upgrade_plt_xml_10_11(file, out_dir, ..., overwrite = FALSE, warning = TRUE)
}
\arguments{
\item{file}{xml plant file path or a vector of}

\item{out_dir}{Output directory path}

\item{...}{additional argument(s) to pass}

\item{overwrite}{Logical TRUE for overwriting files,
FALSE otherwise (default)}

\item{warning}{Logical for rising warnings, FALSE otherwise}
}
\value{
None
}
\description{
Upgrade _plt.xml file(s) from version 10 to 11
}
\examples{
\dontrun{
dir_path <- get_examples_path(file_type = "xml", stics_version = "V10.0")

upgrade_plt_xml_10_11(
  file = file.path(dir_path, "file_plt.xml"),
  out_dir = tempdir()
)
}
}
\keyword{internal}
