% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_v9_v10.R
\name{upgrade_plt_xml_9_10}
\alias{upgrade_plt_xml_9_10}
\title{Upgrade _plt.xml file(s) from STICS version 9 to 10}
\usage{
upgrade_plt_xml_9_10(
  file,
  out_dir,
  param_gen_file,
  param_newform_file,
  overwrite = FALSE
)
}
\arguments{
\item{file}{Path of an plant (*_plt.xml) file or a vector of}

\item{out_dir}{Output directory path of the generated files}

\item{param_gen_file}{Path of the param_gen.xml file corresponding
to the file version}

\item{param_newform_file}{Path of the param_newform.xml file corresponding
to the file version}

\item{overwrite}{logical (optional),
TRUE for overwriting file if it exists, FALSE otherwise}
}
\value{
None
}
\description{
Upgrade _plt.xml file(s) from STICS version 9 to 10
}
\details{
See get_stics_versions_compat() for listing versions
}
\examples{
\dontrun{
dir_path <- get_examples_path(file_type = "xml", stics_version = "V9.2")

upgrade_plt_xml_9_10(
  file = file.path(dir_path, "file_plt.xml"),
  out_dir = tempdir(),
  param_gen_file = file.path(dir_path, "param_gen.xml"),
  param_newform_file = file.path(dir_path, "param_newform.xml")
)
}
}
\keyword{internal}
