% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_xml_files.R
\name{upgrade_sols_xml}
\alias{upgrade_sols_xml}
\title{Upgrade a sols.xml file to a next major version}
\usage{
upgrade_sols_xml(file, out_dir, param_gen_file, overwrite = FALSE)
}
\arguments{
\item{file}{xml sols file path}

\item{out_dir}{Output directory path}

\item{param_gen_file}{Path of the param_gen.xml file corresponding
to the file version}

\item{overwrite}{Logical TRUE for overwriting files,
FALSE otherwise (default)}
}
\value{
None
}
\description{
Upgrade a sols.xml file to a next major version
}
\examples{

dir_path <- get_examples_path(file_type = "xml", stics_version = "V10.0")

upgrade_sols_xml(
  file = file.path(dir_path, "sols.xml"),
  out_dir = tempdir(),
  param_gen_file = file.path(dir_path, "param_gen.xml")
)

}
