% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_v9_v10.R
\name{upgrade_usms_xml_9_10}
\alias{upgrade_usms_xml_9_10}
\title{Upgrade a usms.xml file from STICS version 9 to 10}
\usage{
upgrade_usms_xml_9_10(file, out_dir, param_gen_file, overwrite = FALSE)
}
\arguments{
\item{file}{Path of a usms.xml file}

\item{out_dir}{Output directory path of the generated file}

\item{param_gen_file}{Path of the param_gen.xml file corresponding
to the file version}

\item{overwrite}{logical (optional),
TRUE for overwriting file if it exists, FALSE otherwise}
}
\value{
None
}
\description{
Upgrade a usms.xml file from STICS version 9 to 10
}
\details{
See get_stics_versions_compat() for listing versions
}
\examples{

dir_path <- get_examples_path(file_type = "xml", stics_version = "V9.2")

upgrade_usms_xml_9_10(
  file = file.path(dir_path, "usms.xml"),
  out_dir = tempdir(),
  param_gen_file = file.path(dir_path, "param_gen.xml")
)

}
\keyword{internal}
