% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_xml_files.R
\name{upgrade_workspace_xml}
\alias{upgrade_workspace_xml}
\title{Upgrade a Javastics STICS xml workspace directory to a next major version}
\usage{
upgrade_workspace_xml(
  workspace,
  out_dir,
  ...,
  javastics = NULL,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{workspace}{JavaStics xml workspace path}

\item{out_dir}{Output directory path}

\item{...}{additional argument(s) to pass to called functions}

\item{javastics}{Path of JavaSTICS containing the STICS version corresponding
to the version of the files to be converted (optional)}

\item{overwrite}{Logical TRUE for overwriting files,
FALSE otherwise (default)}

\item{verbose}{logical, TRUE for displaying a copy message
FALSE otherwise (default)}
}
\value{
None
}
\description{
Upgrade a Javastics STICS xml workspace directory to a next major version
}
\examples{
\dontrun{
upgrade_workspace_xml(
  workspace = "/path/to/JavaSTICS/workspace",
  out_dir = "/path/to/an/output/directory",
  javastics = "/path/to/JavaSTICS/folder"
)
}

}
