% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_v10_v11.R
\name{upgrade_workspace_xml_10_11}
\alias{upgrade_workspace_xml_10_11}
\title{Upgrade a Javastics STICS V10 xml workspace directory}
\usage{
upgrade_workspace_xml_10_11(
  workspace,
  javastics = NULL,
  out_dir,
  overwrite = FALSE,
  verbose = FALSE,
  use_patho = FALSE
)
}
\arguments{
\item{workspace}{JavaStics xml workspace path}

\item{javastics}{JavaStics (STICS V10) folder path (Optional)}

\item{out_dir}{Output directory path}

\item{overwrite}{Logical TRUE for overwriting files,
FALSE otherwise (default)}

\item{verbose}{logical, TRUE for displaying a copy message
FALSE otherwise (default)}

\item{use_patho}{Logical TRUE if code_patho is to be added to the
param_newform.xml file, FALSE otherwise}
}
\value{
None
}
\description{
Upgrade a Javastics STICS V10 xml workspace directory
}
\examples{
\dontrun{
upgrade_workspace_xml_10_11(
  workspace = "/path/to/JavaSTICS/workspace",
  javastics = "/path/to/JavaSTICS/folder",
  out_dir = "/path/to/an/output/directory"
)
}

}
