% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_v9_v10.R
\name{upgrade_workspace_xml_9_10}
\alias{upgrade_workspace_xml_9_10}
\title{Upgrade XML files of a JavaSTICS V9 workspace directory}
\usage{
upgrade_workspace_xml_9_10(
  workspace,
  out_dir,
  javastics = NULL,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{workspace}{Path of a JavaSTICS workspace}

\item{out_dir}{Output directory of the generated files}

\item{javastics}{Path of JavaSTICS containing the STICS version corresponding
to the version of the files to be converted (optional)}

\item{overwrite}{logical (optional),
TRUE for overwriting files if they exist, FALSE otherwise}

\item{verbose}{logical, TRUE for displaying a copy message
FALSE otherwise (default)}
}
\value{
None
}
\description{
Upgrade XML files of a JavaSTICS V9 workspace directory
}
\details{
\itemize{
\item See SticsRFiles::get_stics_versions_compat() for listing versions
\item If general parameters files exist in \code{workspace}, they are also upgraded.
In that case, residues parameters values are kept and might not be adapted
to the target model version.
\item Weather data and observations files are fully copied to \code{out_dir}
}
}
\examples{
\dontrun{
upgrade_workspace_xml_9_10(
  workspace = "/path/to/JavaSTICS/workspace",
  javastics = "/path/to/JavaSTICS/folder",
  out_dir = "/path/to/an/output/directory"
)
}
}
