% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_to_col_names.R
\name{var_to_col_names}
\alias{var_to_col_names}
\title{Convert Stics variables names, or generated column names to
valid conventional column names.}
\usage{
var_to_col_names(var_vec)
}
\arguments{
\item{var_vec}{Valid names vector}
}
\value{
Vector of formatted column names
}
\description{
Change Stics variables names as valid variable R name
Like \code{varname(n)} or \code{varname.n.} to  \code{varname_n}.
Other names are unchanged (i.e.: varname, varname_n, ...)
}
\examples{
var_names <- c("var1","var2(n)")
valid_names <- var_to_col_names(var_names)

}
