% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javastics_path.R
\name{write_javastics_path}
\alias{write_javastics_path}
\title{Write the \code{javastics_path} variable in a \code{.Renviron} file, or replace its
content if it already exists.}
\usage{
write_javastics_path(javastics_path)
}
\arguments{
\item{javastics_path}{Absolute path of the root directory of JavaSTICS}
}
\description{
Write the \code{javastics_path} variable in a \code{.Renviron} file, or replace its
content if it already exists.
}
\keyword{internal}
