% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_obs.R
\name{write_obs}
\alias{write_obs}
\title{Write STICS obs data from data.frame}
\usage{
write_obs(x, path)
}
\arguments{
\item{x}{Data.frame to write}

\item{path}{Path to the file to write to}
}
\description{
Write STICS obs data from data.frame
}
\examples{
\dontrun{
write_obs(obs, "wheat.obs")
}
}
