% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_node_to_doc.R
\name{add_node_to_doc}
\alias{add_node_to_doc}
\title{Add node to document}
\usage{
add_node_to_doc(xml_doc, new_node, nodes_nb = 1, parent_path)
}
\arguments{
\item{xml_doc}{XML document}

\item{new_node}{New node to add}

\item{nodes_nb}{Nodes number}

\item{parent_path}{Path to the parent node}
}
\description{
Add node to document
}
