% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stics_nodes.R
\name{add_stics_nodes}
\alias{add_stics_nodes}
\title{Add node to a Stics XML document}
\usage{
add_stics_nodes(
  xml_doc,
  file_tag,
  nodes_nb = 1,
  formalism_name = NULL,
  stics_version = "last"
)
}
\arguments{
\item{xml_doc}{XML document}

\item{file_tag}{File tag (i.e. usms, sols, or tec)}

\item{nodes_nb}{Nodes numbers}

\item{formalism_name}{Name of the formalism}

\item{stics_version}{The version of STICS (eg "V9.0")}
}
\description{
Add node to a Stics XML document
}
