% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_vapour_pressure.R
\name{approx_vapour_pressure}
\alias{approx_vapour_pressure}
\title{Approximate vapour pressure}
\usage{
approx_vapour_pressure(TM, RH)
}
\arguments{
\item{TM}{numeric vector of daily mean temperature}

\item{RH}{numeric vector of relative humidity}
}
\value{
numeric vector
}
\description{
Approximates vapour pressure as the saturated vapour pressure at the daily mean temperature using relative humidity.
See section 9.3 of the STICS book, as well as M. Launay's explanations on the STICS forum (\href{https://w3.avignon.inrae.fr/forge/boards/26/topics/2237?r=2243#message-2243}{link}).
}
\author{
Timothee Flutre
}
