% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_java_path.R
\name{check_java_path}
\alias{check_java_path}
\title{Checking JavaStics directory content}
\usage{
check_java_path(javastics)
}
\arguments{
\item{javastics}{JavaStics installation root folder}
}
\description{
Checking if directory exists and if it contains JavaStics jar files
}
\details{
Rising an exception for each checking step !
}
\examples{
\dontrun{
check_java_path("/path/to/JavaSTICS/folder")
}

}
\keyword{internal}
