% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_xml2txt.R
\name{convert_xml2txt}
\alias{convert_xml2txt}
\title{Transforming a STICS xml file into a text file}
\usage{
convert_xml2txt(
  file,
  plant_id = 1,
  soil_name = NULL,
  out_dir = NULL,
  save_as = NULL,
  stics_version = "latest"
)
}
\arguments{
\item{file}{Path (including name) of the xml file to convert}

\item{plant_id}{The plant identifier (main crop: 1 ; associated crop: 2)}

\item{soil_name}{Soil name (optional, required for soil file)}

\item{out_dir}{Path of the directory where to generate the file.
Optional, set to the path of the input xml file by default}

\item{save_as}{Name of the output file
(optional, default: fixed name for STICS)}

\item{stics_version}{the STICS files version to use (optional,
default to latest).}
}
\value{
None
}
\description{
The input file according to his type
(ini,plant,tec,station,soil,par)
is converted to a text file readable by the STICS model
(ficini.txt, ficplt1.txt,...)
}
\examples{
xml_path <- file.path(get_examples_path("xml"), "file_plt.xml")
convert_xml2txt(file = xml_path)
xml_path <- file.path(get_examples_path("xml"), "sols.xml")
convert_xml2txt(file = xml_path, soil_name = "soil_rice")

}
