% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download example USMs}
\usage{
download_data(
  out_dir = tempdir(),
  example_dirs = NULL,
  stics_version = "latest"
)
}
\arguments{
\item{out_dir}{Path of the directory where to download the data}

\item{example_dirs}{List of use case directories names (optional)}

\item{stics_version}{Name of the STICS version (optional)
The default value is the latest version returned by
get_stics_versions_compat().}
}
\value{
The path of the folder data have been downloaded into
}
\description{
Download locally the example data from the
\href{https://github.com/SticsRPacks/data}{data repository} in the SticsRPacks
organization.
}
\examples{

# Getting data for a given example : study_case_1 and a given STICS version
download_data(example_dirs = "study_case_1", stics_version = "V9.0")

}
