% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download example USMs}
\usage{
download_data(dir = tempdir(), example_dirs = NULL, version_name = NULL)
}
\arguments{
\item{dir}{The directory where to download the data}

\item{example_dirs}{List of use case directories names (optional)}

\item{version_name}{An optional version string
within those given by get_stics_versions_compat()$versions_list, or "latest"
for getting the latest version (default : NULL)}
}
\value{
The path to the folder where data have been downloaded
}
\description{
Download locally the example data from the \href{https://github.com/SticsRPacks/data}{data repository}
in the SticsRPacks organisation.
}
\examples{
\dontrun{

# Getting all data
download_data()

download_data(example_dirs = "study_case_1")

# Getting data for a given example : study_case_1 and a given version
download_data(example_dirs = "study_case_1", version_name = "V9.0")


}
}
