% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download example USMs}
\usage{
download_data(dir = NULL)
}
\value{
Download the data and return the path to the folder where the data was downloaded
}
\description{
Download locally the example data from the \href{https://github.com/SticsRPacks/data}{data repository}
in the SticsRPacks organisation.
}
\examples{
\dontrun{
download_data()
}
}
