% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_usm_xl.R
\name{download_usm_csv}
\alias{download_usm_csv}
\title{Downloading a CSV usms data file example into a directory}
\usage{
download_usm_csv(
  csv_name = NULL,
  dest_dir = getwd(),
  version_name = "last",
  overwrite = FALSE
)
}
\arguments{
\item{csv_name}{Name of a csv file (optional, not used for the moment)}

\item{dest_dir}{Directory path where to copy the CSV file
(optional, default: current directory)}

\item{version_name}{An optional version string (default: last version returned by get_stics_versions_compat())}

\item{overwrite}{Optional logical, TRUE for overwriting files, FALSE otherwise (default)}
}
\value{
A vector of copied files path.
}
\description{
The file is an example that can be used for generating JavaStics
usms.xml input file from parameters values stored in a CSV file using the function
\code{\link{gen_usms_xml}}
}
\examples{
\dontrun{
download_usm_csv(dest_dir = "/path/to/destination/dir")
}

}
