% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_usm_xl.R
\name{download_usm_xl}
\alias{download_usm_xl}
\title{Downloading an Excel usms data file example into a directory}
\usage{
download_usm_xl(xl_name = NULL, dest_dir = getwd())
}
\arguments{
\item{xl_name}{Name of an Excel file (optional, not used for the moment)}

\item{dest_dir}{Directory path where to copy the Excel file
(optional, default: current directory)}
}
\value{
A copy status, TRUE if successfull, FALSE otherwise
}
\description{
The file is an example that can be used for generating JavaStics input files
from parameters values stored in Excel spreadsheet format (USMs, Ini, Soils, Tec, Station, ...).
Each sheet contains parameters values to insert into XML files, with the help of
these functions: \code{\link{gen_usms_xml}}, \code{\link{gen_sols_xml}}, \code{\link{gen_tec_xml}},
\code{\link{gen_sta_xml}}, \code{\link{gen_usms_xml}}, \code{\link{gen_ini_xml}}
}
\examples{
\dontrun{
download_usm_xl()
download_usm_xl(dest_dir = "/path/to/destination/dir")
}

}
