% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_usm_xl.R
\name{download_usm_xl}
\alias{download_usm_xl}
\title{Downloading an Excel usms data file example into a directory}
\usage{
download_usm_xl(
  xl_name = NULL,
  dest_dir = getwd(),
  version_name = "last",
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{xl_name}{Name of an Excel file (optional, not used for the moment)}

\item{dest_dir}{Directory path where to copy the Excel file
(optional, default: current directory)}

\item{version_name}{An optional version string (default: last version returned by get_stics_versions_compat())}

\item{overwrite}{Optional logical, TRUE for overwriting files, FALSE otherwise (default)}

\item{...}{Additional arguments to be passed}
}
\value{
A vector of copied files path.
}
\description{
The file is an example that can be used for generating JavaStics input files
from parameters values stored in Excel spreadsheet format (USMs, Ini, Soils, Tec, Station, ...).
Each sheet contains parameters values to insert into XML files, with the help of
these functions: \code{\link{gen_usms_xml}}, \code{\link{gen_sols_xml}}, \code{\link{gen_tec_xml}},
\code{\link{gen_sta_xml}}, \code{\link{gen_usms_xml}}, \code{\link{gen_ini_xml}}
}
\examples{
\dontrun{
download_usm_xl()
download_usm_xl(dest_dir = "/path/to/destination/dir")
}

}
