% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists_param.R
\name{exists_param}
\alias{exists_param}
\title{Test if a parameter exists}
\usage{
exists_param(xml_doc_object, param_name, ...)
}
\arguments{
\item{xml_doc_object}{XML document}

\item{param_name}{The parameter name or a vector of}

\item{...}{Other parameters to pass to \code{\link[=get_param_number]{get_param_number()}}.}
}
\value{
A vector of logical values
}
\description{
Test if a parameter exists in an XML file
}
\examples{
xml_sols <- file.path(get_examples_path( file_type = "xml"),"sols.xml")

xml_doc <- SticsRFiles:::xmldocument(xml_sols)

SticsRFiles:::exists_param(xml_doc, "cfes")
SticsRFiles:::exists_param(xml_doc, c("cfes","mulchbat"))

}
\keyword{internal}
