% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_stics_names.R
\name{expand_stics_names}
\alias{expand_stics_names}
\title{Expand stics names from parameters/variables simple names,
with indices (i.e. par(1), par(2),...regarding to parameter/variable
dimensions)}
\usage{
expand_stics_names(in_csv_file, out_csv_file, header_vec = NULL)
}
\arguments{
\item{in_csv_file}{Input csv file path}

\item{out_csv_file}{Output csv file path}

\item{header_vec}{An optional header vector}
}
\description{
Expand stics names from parameters/variables simple names,
with indices (i.e. par(1), par(2),...regarding to parameter/variable
dimensions)
}
\examples{
\dontrun{

csv_file <- "/path/to/JavaSTICS/folder/config/inputs.csv"
out_csv_file <- "/path/to/out/dir/config/inputs_xpanded.csv"

SticsRFiles:::expand_stics_names(csv_file, out_csv_file)
}

}
\keyword{internal}
