% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ini_doc.R
\name{gen_ini_doc}
\alias{gen_ini_doc}
\title{Generatefrom a template a Stics ini xmlDocument}
\usage{
gen_ini_doc(
  xml_doc = NULL,
  param_table = NULL,
  crop_tag = "crop",
  params_desc = NULL,
  stics_version = "last"
)
}
\arguments{
\item{xml_doc}{an optional xmlDocument object (created from an ini file)}

\item{param_table}{a table (df, tibble) containing parameters to use (optional)}

\item{crop_tag}{a crop identifier for crop parameters
(for example like "crop" used for in a parameter suffix : param_crop1, param_crop2)}

\item{params_desc}{a list decribing crop parameters and others}

\item{stics_version}{the stics files version to use (optional, default to last). Only used if xml_doc = NULL.}
}
\value{
an invisible xmlDocument object or a list of
}
\description{
Generatefrom a template a Stics ini xmlDocument
}
\examples{
\dontrun{
library(readxl)

xl_path <- "inputs_stics_example.xlsx"
download_usm_xl(xl_name = xl_path)
ini_param_df <- read_excel(xl_path, sheet = "Ini")
ini_doc <- SticsRFiles:::gen_ini_doc(param_table = ini_param_df)
}

}
\keyword{internal}
