% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ini_doc.R
\name{gen_ini_doc}
\alias{gen_ini_doc}
\title{Generate  from a template or modify a Stics ini xmlDocument}
\usage{
gen_ini_doc(
  xml_doc = NULL,
  param_table = NULL,
  crop_tag = "crop",
  params_desc = NULL,
  stics_version = "last"
)
}
\arguments{
\item{xml_doc}{an xmlDocument object (created from an ini file)}

\item{param_table}{a table (df, tibble) containing parameters to use}

\item{crop_tag}{a crop identifier for crop parameters ()}

\item{params_desc}{a list decribing crop parameters and others}

\item{stics_version}{the stics files version to use}
}
\value{
an xmlDocument object or a list of
}
\description{
Generate  from a template or modify a Stics ini xmlDocument
}
\keyword{internal}
