% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ini_xml.R
\name{gen_ini_xml}
\alias{gen_ini_xml}
\title{Generate Stics ini xml file(s) from a template or an input file}
\usage{
gen_ini_xml(
  param_table = NULL,
  ini_in_file = NULL,
  out_path = getwd(),
  crop_tag = "Crop",
  stics_version = "latest"
)
}
\arguments{
\item{param_table}{a table (df, tibble) containing parameters to use (see details)}

\item{ini_in_file}{file path to an XML file (optional, if not povided, uses a template from the package corresponding to stics_version)}

\item{out_path}{path to an optional folder where to write the output file(s)}

\item{crop_tag}{identifier for the crop parameters names related to the main crop, or the associated crop if any
(example: Crop is used in the param_table example in the details section below)}

\item{stics_version}{the stics version to use (optional, default to latest). Only used if ini_in_file= NULL, see details.}
}
\value{
an invisible xmlDocument object or a list of
}
\description{
Generate Stics ini xml file(s) from a template or an input file
}
\details{
Please see \code{get_stics_versions_compat()} for the full list of stics versions that can be used for the
argument \code{stics_version}.
\code{param_table} is a \code{data.frame} with the following format:\tabular{lrlrrr}{
   Ini_name \tab nbplantes \tab stade0_Crop1 \tab lai0_Crop1 \tab masec0_Crop1 \tab QNplante0_Crop1 \cr
   USM_2017_T1_ini.xml \tab 1 \tab snu \tab 0 \tab 0 \tab 0 \cr
   Vill09_ini.xml \tab 1 \tab snu \tab 0 \tab 0 \tab 0 \cr
   Vill10_ini.xml \tab 1 \tab snu \tab 0 \tab 0 \tab 0 \cr
   Vill11_ini.xml \tab 1 \tab snu \tab 0 \tab 0 \tab 0 \cr
   Vill12_ini.xml \tab 1 \tab snu \tab 0 \tab 0 \tab 0 \cr
   Vill13_ini.xml \tab 1 \tab snu \tab 0 \tab 0 \tab 0 \cr
   Vill14_ini.xml \tab 1 \tab snu \tab 0 \tab 0 \tab 0 \cr
   Standard_ini.xml \tab 1 \tab snu \tab 0 \tab 0 \tab 0 \cr
}


The first column gives the ini file name (to be generated), all following columns give the parameter value to put
in the file, and each line denotes a separate ini file (for e.g. several USMs).

The first column name must contain the keyword ini or Ini or INI as a prefix to be detected
(as shown in the table extract above).
}
\examples{
\dontrun{
library(readxl)

xl_path <- "inputs_stics_example.xlsx"
download_usm_xl(xl_name = xl_path)
ini_param_df <- read_excel(xl_path, sheet = "Ini")
gen_ini_xml(out_path = "/path/to/dest/dir",
param_table = ini_param_df)
}

}
