% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_obs.R
\name{gen_obs}
\alias{gen_obs}
\title{Write STICS obs data from data.frame}
\usage{
gen_obs(x, path)
}
\arguments{
\item{x}{data.frame to write}

\item{path}{Path to the file to write to}
}
\description{
Write STICS obs data from data.frame
}
\examples{
\dontrun{
gen_obs(obs, "wheat.obs")
}
}
