% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sols_doc.R
\name{gen_sols_doc}
\alias{gen_sols_doc}
\title{Generate from a template a Stics sols xmlDocument}
\usage{
gen_sols_doc(
  xml_doc = NULL,
  sols_nb = NULL,
  sols_param = NULL,
  stics_version = "last"
)
}
\arguments{
\item{xml_doc}{an xmlDocument object (created from a sols.xml file)}

\item{sols_nb}{number of soils to create}

\item{sols_param}{a table (df, tibble) containing parameters to use}

\item{stics_version}{the stics files version to use}
}
\value{
an an invisible xmlDocument object
}
\description{
Generate from a template a Stics sols xmlDocument
}
\examples{
\dontrun{
library(readxl)

xl_path <- "inputs_stics_example.xlsx"
download_usm_xl(xl_name = "inputs_stics_example.xlsx", dest_dir = "/path/to/dest/dir")
sols_param_df <- read_excel(xl_path, sheet = "Soils")
soil_doc <- SticsRFiles:::gen_sols_doc(sols_param = sols_param_df)
}


}
\keyword{internal}
