% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sta_doc.R
\name{gen_sta_doc}
\alias{gen_sta_doc}
\title{Generatefrom a template a Stics sta xmlDocument}
\usage{
gen_sta_doc(xml_doc = NULL, param_table = NULL, stics_version = "last")
}
\arguments{
\item{xml_doc}{an optional xmlDocument object (created from an sta file)}

\item{param_table}{a table (df, tibble) containing parameters to use (optional)}

\item{stics_version}{the stics files version to use (optional, default to last). Only used if xml_doc = NULL.}
}
\value{
an invisible xmlDocument object or a list of
}
\description{
Generatefrom a template a Stics sta xmlDocument
}
\examples{
\dontrun{
library(readxl)

xl_path <- "inputs_stics_example.xlsx"
download_usm_xl(xl_name = xl_path)
sta_param_df <- readxl::read_excel(xl_path, sheet = "Station")
sta_doc <- SticsRFiles:::gen_sta_doc(param_table = sta_param_df)
}

}
\keyword{internal}
