% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sta_doc.R
\name{gen_sta_doc}
\alias{gen_sta_doc}
\title{Generate  from a template or modify a Stics sta xmlDocument}
\usage{
gen_sta_doc(xml_doc = NULL, param_table = NULL, stics_version = "last")
}
\arguments{
\item{xml_doc}{an xmlDocument object (created from an sta file)}

\item{param_table}{a table (df, tibble) containing parameters to use}

\item{stics_version}{the stics files version to use}
}
\value{
an xmlDocument object or a list of
}
\description{
Generate  from a template or modify a Stics sta xmlDocument
}
\examples{
\dontrun{
copy_mailing_example(xl_name = "inputs_stics_example.xlsx", dest_dir = "/path/to/dest/dir")
xl_path <- file.path("/path/to/dest/dir","inputs_stics_example.xlsx")
sta_param_df <- read_excel(xl_path, sheet = "Station")
sta_doc <- SticsRFiles:::gen_sta_doc(param_table = sta_param_df)
}

}
\keyword{internal}
