% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sta_xml.R
\name{gen_sta_xml}
\alias{gen_sta_xml}
\title{Generate Stics sta xml file(s) from a template or an input file}
\usage{
gen_sta_xml(
  param_table = NULL,
  sta_in_file = NULL,
  out_path = getwd(),
  stics_version = "latest"
)
}
\arguments{
\item{param_table}{a table (df, tibble) containing parameters to use (see details)}

\item{sta_in_file}{file path to an XML file (optional, if not povided, uses a template from the package corresponding to stics_version)}

\item{out_path}{path to an optional folder where to write the output file(s)}

\item{stics_version}{the stics version to use (optional, default to latest). Only used if sta_in_file= NULL, see details.}
}
\value{
an invisible xmlDocument object or a list of
}
\description{
Generate Stics sta xml file(s) from a template or an input file
}
\details{
Please see \code{get_stics_versions_compat()} for the full list of stics versions that can be used for the
argument \code{stics_version}.

\code{param_table} is a \code{data.frame} with the following format:\tabular{lrrrrr}{
   Sta_name \tab zr \tab NH3ref \tab latitude \tab patm \tab aclim \cr
   climatex_sta.xml \tab 2.5 \tab 0 \tab 49 \tab 1000 \tab 20 \cr
   climatex2_sta.xml \tab 2.8 \tab 0 \tab 49 \tab 1000 \tab 20 \cr
   climatex3_sta.xml \tab 2.2 \tab 0 \tab 49 \tab 1000 \tab 20 \cr
}


The first column gives the sta file name (to be generated), all following columns give the parameter value to put
in the file, and each line denotes a separate sta file (for e.g. several USMs).

The first column name must contain the keyword sta or Sta or STA as a prefix to be detected
(as shown in the table extract above).
}
\examples{
\dontrun{

xl_path <- download_usm_xl(file = "inputs_stics_example.xlsx")
sta_param_df <- read_params_table(file_path = xl_path, sheet_name = "Station")
gen_sta_xml(out_path = "/path/to/dest/dir", param_table = sta_param_df)

}

}
