% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tec_doc.R
\name{gen_tec_doc}
\alias{gen_tec_doc}
\title{Generate from a template or modify a Stics tec xmlDocument}
\usage{
gen_tec_doc(
  xml_doc = NULL,
  param_table = NULL,
  stics_version = "last",
  dict = NULL,
  ...
)
}
\arguments{
\item{xml_doc}{an xmlDocument object (created from an ini file)}

\item{param_table}{a table (df, tibble) containing parameters to use}

\item{stics_version}{the stics files version to use (optional, default to last). Only used if xml_doc = NULL.}

\item{dict}{List of correspondance between given parameter names and Stics internal names.}
}
\value{
an invisible xmlDocument object or a list of
}
\description{
Generate from a template or modify a Stics tec xmlDocument
}
\examples{
\dontrun{
library(readxl)

download_usm_xl(xl_name = "inputs_stics_example.xlsx", dest_dir = "/path/to/dest/dir")
xl_path <- file.path("/path/to/dest/dir","inputs_stics_example.xlsx")
tec_param_df <- read_excel(xl_path, sheet = "Tec")
tec_doc <- SticsRFiles:::gen_tec_doc(param_table = tec_param_df)
}

}
\keyword{internal}
