% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tec_xml.R
\name{gen_tec_xml}
\alias{gen_tec_xml}
\title{Generate from a template  or an input file a Stics tec xml file}
\usage{
gen_tec_xml(
  tec_out_file = NULL,
  param_table = NULL,
  tec_in_file = NULL,
  out_path = getwd(),
  stics_version = "last",
  dict = NULL
)
}
\arguments{
\item{tec_out_file}{file name of the output tec xml file}

\item{param_table}{a table (df, tibble) containing parameters to use}

\item{tec_in_file}{file path for an input tec xml file}

\item{out_path}{path for the ini output file}

\item{stics_version}{the stics files version to use}

\item{dict}{List of correspondance between given parameter names and internal names.}
}
\value{
an invisible xmlDocument object or a list of
}
\description{
Generate from a template  or an input file a Stics tec xml file
}
\examples{
\dontrun{
copy_mailing_example(xl_name = "inputs_stics_example.xlsx", dest_dir = "/path/to/dest/dir")
xl_path <- file.path("/path/to/dest/dir","inputs_stics_example.xlsx")
tec_param_df <- read_excel(xl_path, sheet = "Tec")
gen_tec_xml(out_path = file.path("/path/to/dest/dir","tec"),
param_table = tec_param_df)
}

}
