% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tec_xml.R
\name{gen_tec_xml}
\alias{gen_tec_xml}
\title{Generate Stics tec xml file(s) from a template or an input file}
\usage{
gen_tec_xml(
  param_table = NULL,
  tec_in_file = NULL,
  out_path = getwd(),
  stics_version = "latest",
  na_values = NA
)
}
\arguments{
\item{param_table}{a table (df, tibble) containing parameters to use (see details)}

\item{tec_in_file}{file path to an XML file (optional, if not povided, uses a template from the package corresponding to stics_version)}

\item{out_path}{path to an optional folder where to write the output file(s)}

\item{stics_version}{the stics version to use (optional, default to latest). Only used if tec_in_file= NULL, see details.}

\item{na_values}{value to use as missing value in param_table (optional, default : NA)}
}
\value{
an invisible xmlDocument object or a list of
}
\description{
Generate Stics tec xml file(s) from a template or an input file
}
\details{
Please see \code{get_stics_versions_compat()} for the full list of stics versions that can be used for the
argument \code{stics_version}.

\code{param_table} is a \code{data.frame} with the following format:\tabular{lrrrrr}{
   Tec_name \tab julres_1 \tab coderes_1 \tab qres_1 \tab Crespc_1 \tab CsurNres_1 \cr
   USM_2017_T1_CI_tec.xml \tab 999 \tab 1 \tab 9 \tab 42 \tab 90 \cr
   BIN_CANPC_05_SEC_220-0-0_34K_CANPC05T3_Q_tec.xml \tab 110 \tab 1 \tab 9 \tab 42 \tab 90 \cr
   BIN_AGT_04_IRR_220-0-0_33K_AGT04T2_Q_tec.xml \tab 73 \tab 1 \tab 9 \tab 42 \tab 90 \cr
   AGA_ARB_13_IRR_220-0-0_37K_ARB13_C_tec.xml \tab 82 \tab 1 \tab 9 \tab 42 \tab 90 \cr
   AGA_ARB_13_SEC_220-0-0_37K_ARB13_C_tec.xml \tab 82 \tab 1 \tab 9 \tab 42 \tab 90 \cr
   FRA_ARB_11_SEC_220-0-0_38K_E_tec.xml \tab 70 \tab 1 \tab 9 \tab 42 \tab 90 \cr
   MAG_ARB_09_SEC_220-0-0_38K_E_tec.xml \tab 81 \tab 1 \tab 9 \tab 42 \tab 90 \cr
   MAG_ARV_12_IRR_220-0-0_36K_ARV12_C_tec.xml \tab 100 \tab 1 \tab 9 \tab 42 \tab 90 \cr
   MAG_ARV_12_SEC_220-0-0_36K_ARV12_C_tec.xml \tab 100 \tab 1 \tab 9 \tab 42 \tab 90 \cr
   FRA_ARB_12_SEC_220-0-0_31K_ARB12_C_tec.xml \tab 92 \tab 1 \tab 9 \tab 42 \tab 90 \cr
   FRA_ARB_13_SEC_220-0-0_37K_ARB13_C_tec.xml \tab 82 \tab 1 \tab 9 \tab 42 \tab 90 \cr
}


The first column gives the tec file name (to be generated), all following columns give the parameter value to put
in the file, and each line denotes a separate tec file (for e.g. several USMs).

The first column name must contain the keyword tec or Tec or TEC as a prefix to be detected
(as shown in the table extract above).
}
\examples{
\dontrun{

xl_path <-  download_usm_xl(xl_name = "inputs_stics_example.xlsx")
tec_param_df <- read_params_table(file_path = xl_path, sheet_name = "Tec")
gen_tec_xml(out_path = "/path/to/dest/dir", param_table = tec_param_df)

}

}
