% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_usms_doc.R
\name{gen_usms_doc}
\alias{gen_usms_doc}
\title{Generate  from a template or modify a Stics usms xmlDocument}
\usage{
gen_usms_doc(
  xml_doc = NULL,
  usms_nb = NULL,
  usms_param = NULL,
  stics_version = "last"
)
}
\arguments{
\item{xml_doc}{an xmlDocument object (created from an ini file)}

\item{usms_nb}{number of usms to create}

\item{usms_param}{a table (df, tibble) containing parameters to use}

\item{stics_version}{the stics files version to use}
}
\value{
an xmlDocument object
}
\description{
Generate  from a template or modify a Stics usms xmlDocument
}
\examples{
\dontrun{
copy_mailing_example(xl_name = "inputs_stics_example.xlsx", dest_dir = "/path/to/dest/dir")
xl_path <- file.path("/path/to/dest/dir","inputs_stics_example.xlsx")
usms_param_df <- read_excel(xl_path, sheet = "USMs")
usms_doc <- SticsRFiles:::gen_usms_doc(usms_param = usms_param_df)
}

}
\keyword{internal}
