% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_usms_xml.R
\name{gen_usms_xml}
\alias{gen_usms_xml}
\title{Generate from a template a Stics usms xml file}
\usage{
gen_usms_xml(
  usms_out_file,
  usms_nb = NULL,
  usms_param = NULL,
  usms_in_file = NULL,
  stics_version = "last"
)
}
\arguments{
\item{usms_out_file}{file name of the output usms xml file}

\item{usms_nb}{number of usms to create}

\item{usms_param}{a table (df, tibble) containing parameters to use}

\item{usms_in_file}{file path for an input usms xml file}

\item{stics_version}{the stics files version to use}
}
\value{
an invisible xmlDocument object
}
\description{
Generate from a template a Stics usms xml file
}
\examples{
\dontrun{
copy_mailing_example(xl_name = "inputs_stics_example.xlsx", dest_dir = "/path/to/dest/dir")
xl_path <- file.path("/path/to/dest/dir","inputs_stics_example.xlsx")
usms_param_df <- read_excel(xl_path, sheet = "USMs")
gen_usms_xml(usms_out_file = file.path("/path/to/dest/dir","usms.xml"),
usms_param = usms_param_df)
}

}
