% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_varmod.R
\name{gen_varmod}
\alias{gen_varmod}
\title{Generating a var.mod type file}
\usage{
gen_varmod(workspace, var_names, file_name = "var.mod")
}
\arguments{
\item{workspace}{Stics or JavaStics workspace path}

\item{var_names}{vector of variables names (see details)}

\item{file_name}{file name to generate (default value: var.mod)}
}
\value{
A logical status TRUE if successfull generation, FALSE otherwise
}
\description{
Generating a daily variable list file from variables names
}
\details{
The variable names can be found using \code{find_var_info()}
}
\examples{
\dontrun{
 gen_varmod(".", c("var1","var2"))
 gen_varmod("/path/to/stics/workspace", c("var1","var2"))
}

}
