% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_txt.R
\name{get_climate_txt}
\alias{get_climate_txt}
\title{Read STICS input meteo file}
\usage{
get_climate_txt(dirpath = getwd(), filename = "climat.txt", preserve = TRUE)
}
\arguments{
\item{dirpath}{Directory path}

\item{filename}{The meteorology file name (default to \code{climat.txt}).}

\item{preserve}{Logical, TRUE for keeping original Stics meteo columns (default)
or FALSE otherwise.}
}
\value{
A data.frame of the input meteorological variables used as input for the
STICS model.
}
\description{
Read the meteorology input for STICS ("climat.txt")
}
\note{
The time-related variables are summarised into one POSIXct column named
code{Date}.
}
\examples{
library(SticsRFiles)
path <- system.file("extdata/txt/V9.0", package = "SticsRFiles")
Meteo <- get_climate_txt(path)

}
\seealso{
\code{\link{get_obs_int}}
}
