% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_txt.R
\name{get_climate_txt}
\alias{get_climate_txt}
\title{Read STICS input meteo file}
\usage{
get_climate_txt(dirpath = getwd(), filename = "climat.txt", preserve = TRUE)
}
\arguments{
\item{dirpath}{Directory path}

\item{filename}{The meteorology file name (default to \code{climat.txt}).}

\item{preserve}{Logical, \verb{TRUE`` for keeping the stics columns related to date (year, month, day, julian), or only keep the date as a }POSIXct\verb{otherwise. Default to}TRUE`.}
}
\value{
A data.frame of the input meteorological variables used as input for the
STICS model.
}
\description{
Read the meteorology input for STICS ("climat.txt")
}
\note{
The time-related variables are summarised into one POSIXct column named
\code{Date}.
}
\examples{
library(SticsRFiles)
path <- get_examples_path( file_type = "txt")
Meteo <- get_climate_txt(path)

}
