% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_results.R
\name{get_daily_results}
\alias{get_daily_results}
\title{Load and format Stics daily output file(s)}
\usage{
get_daily_results(
  workspace,
  usm_name = NULL,
  var_list = NULL,
  doy_list = NULL,
  dates_list = NULL,
  mixed = NULL,
  usms_filename = "usms.xml",
  javastics_path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{workspace}{Stics or JavaStics workspace path containing the \code{mod_s*.sti} files (see details)}

\item{usm_name}{usm name(s) to filter}

\item{var_list}{vector of output variables names to filter (optional, see \code{get_var_info()} to get the names of the variables)}

\item{doy_list}{vector of cumulative DOYs to filter (optional)}

\item{dates_list}{list of dates to filter (optional)}

\item{mixed}{A booean (recycled), or a list of. \code{TRUE}: intercrop, \code{FALSE}: sole crops (default), \code{NULL}: guess from XML file.}

\item{usms_filename}{usms file path (e.g. "usms.xml") if any \code{mixed= NULL}.}

\item{javastics_path}{JavaStics installation path (Optional, needed if the plant files are not in the \code{workspace}
but rather in the JavaStics default workspace)}

\item{verbose}{Logical value (optional), TRUE to display infos on error, FALSE otherwise (default)}
}
\value{
A list, where each element is a tibble of simulation results for the given usm. The list is named
after the usm name.
}
\description{
Reads and format daily output file(s) (mod_s*.sti) for usm(s) with
possible selection on variable names, cumulative DOY, dates list.
}
\details{
The function can guess if the usm(s) is/are mixed or not by reading the XML
input file. To do so, set the \code{mixed} argument to \code{NULL}.
}
\examples{
\dontrun{
path <- get_examples_path( file_type = "sti")
get_daily_results(path,"banana")
}
}
