% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim.R
\name{get_daily_results}
\alias{get_daily_results}
\title{Load and format Stics daily output file(s)}
\usage{
get_daily_results(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because we realised that it's
a special case of the \code{\link[=get_sim]{get_sim()}}
}
\examples{
\dontrun{
get_daily_results(path, "banana")
# ->
get_sim(path, "banana")
}

}
\keyword{internal}
