% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_results.R
\name{get_daily_results}
\alias{get_daily_results}
\title{Loading Stics daily output file(s)}
\usage{
get_daily_results(
  workspace,
  usm_name,
  var_list = NULL,
  doy_list = NULL,
  dates_list = NULL,
  mixed = rep(FALSE, length(usm_name)),
  usms_file = "usms.xml",
  javastics_path = NULL
)
}
\arguments{
\item{workspace}{Stics or JavaStics workspace path containing the \code{mod_s*.sti} files (see details)}

\item{usm_name}{vector of usm(s) names}

\item{var_list}{vector of output variables names (optional, see \code{find_var_info()} to get the names of the variables)}

\item{doy_list}{vector of cumulative DOYs (optional)}

\item{dates_list}{list of dates (optional)}

\item{mixed}{value (recycled) or vector of. \code{TRUE}: intercrop, \code{FALSE}: sole crops (default), \code{NULL}: guess from XML files.}

\item{usms_file}{The name of the usms file (e.g. "usms.xml") in case of \code{NULL} values in \code{mixed}.}

\item{javastics_path}{JavaStics installation path (Optional, needed if the plant files are not in the \code{workspace}
but rather in the JavaStics default workspace)}
}
\value{
A tibble or a list of
}
\description{
Reading daily file(s) (mod_s*.sti) as a tibble with possible selection
on variable names, cumulative DOY, dates list.
}
\details{
The function can guess if the usm(s) are mixed or not by reading the XML
input files. To do so, set the \code{mixed} argument to \code{NULL}.
}
\examples{
path <- system.file(file.path("extdata","sti","V9.0"), package = "SticsRFiles")
get_daily_results(path,"banana")

}
