% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_results.R
\name{get_daily_results}
\alias{get_daily_results}
\title{Loading Stics daily output file(s)}
\usage{
get_daily_results(
  workspace,
  usm_name,
  var_list = NULL,
  doy_list = NULL,
  dates_list = NULL
)
}
\arguments{
\item{workspace}{Stics or JavaStics workspace path}

\item{usm_name}{names of the usm}

\item{var_list}{vector of output variables names (optional)}

\item{doy_list}{vector of cumulative DOYs (optional)}

\item{dates_list}{list of dates (optional)}
}
\value{
A tibble or a list of
}
\description{
Reading daily file(s) as a tibble with possible selection
on variables names, cumulative DOY, dates list
}
\examples{
path <- system.file(file.path("extdata","sti","V9.1"), package = "SticsRFiles")
get_daily_results(path,"banana")

}
