% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_examples_utils.R
\name{get_examples_path}
\alias{get_examples_path}
\title{Getting examples files path attached to a STICS version for a given file type}
\usage{
get_examples_path(file_type, version_name = "last")
}
\arguments{
\item{file_type}{A file type string among files types ("csv", "obs", "sti", "txt", "xml")}

\item{version_name}{An optional version string (default: last version returned by get_stics_versions_compat())}
}
\value{
A directory path for examples files for given file type and STICS version
}
\description{
Getting examples files path attached to a STICS version for a given file type
}
\examples{

\dontrun{

get_examples_path( file_type = "csv" )

[1] "/path/to/user-R-library/SticsRFiles/extdata/csv/V9.1"

get_examples_path( file_type = "csv", version_name = "V8.5")

[1] "/path/to/user-R-library/SticsRFiles/extdata/csv/V8.5"

}
}
