% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs.R
\name{get_obs}
\alias{get_obs}
\title{Read STICS observation files (*.obs)}
\usage{
get_obs(
  dirpath = getwd(),
  obs_filenames = NULL,
  usms = NULL,
  usms_filename = "usms.xml"
)
}
\arguments{
\item{dirpath}{Path to the directory containing the obs path}

\item{obs_filenames}{A vector of observation file name(s). Optional, see details.}

\item{usms}{A vector of usm name(s) corresponding the the obs files. Optional, see details.}

\item{usms_filename}{name of the usm file. Optional, see details.}
}
\value{
A list of STICS-formated observations. Return \code{NULL} if no files were found.
}
\description{
Read STICS observation files for sole crops and store data into a list
}
\details{
If \code{filename} is not specified (or equal to \code{NULL}), the
function read all obs files in the \code{dirpath} folder. If there are no .obs files,
the function returns \code{NULL}
If usms names are not provided they are supposed to have the same name as the obs file.
If usms_filename is not provided, existence of usms are checked in the usms.xml file.
}
\examples{
\dontrun{
path <- system.file(file.path("extdata","obs","V9.0"), package = "SticsRFiles")
Meas <- get_obs(path)
}

}
