% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs.R
\name{get_obs}
\alias{get_obs}
\title{Read STICS observation files (*.obs)}
\usage{
get_obs(
  workspace_path = getwd(),
  usms_list = NULL,
  usms_filename = "usms.xml",
  verbose = TRUE
)
}
\arguments{
\item{workspace_path}{Path of a JavaStics workspace, or a vector of (recursive call).}

\item{usms_list}{Vector of usms to read (optional, used to filter usms)}

\item{usms_filename}{Name of the usm file (optional, default to the standard \code{usms.xml})}

\item{verbose}{Logical value (optional), TRUE to display infos on error, FALSE otherwise (default)}
}
\value{
A list of \code{data.frame}s with observations. The function always returns a list of \code{data.frame}s for the corresponding usm,
even if an \code{.obs} file is missing (in this case it returns a \code{data.frame} with 0 rows and 0 columns.
}
\description{
Read STICS observation files from a JavaStics workspace and store data into a list per usm
}
\details{
The \code{.obs} files names should match the name of the USM, e.g. for a usm called "banana",
the \code{.obs} file should be named \code{banana.obs}. For intercrops, the name should be suffixed by "p" for
the principal and "a" for the associated plant.
If \code{usms_list} is not specified (or equal to \code{NULL}), the
function read the obs files from all usms listed in \code{usms_filename} and present in the \code{workspace_path} folder.
}
\examples{
\dontrun{
path <- system.file(file.path("extdata","obs","V9.0"), package = "SticsRFiles")

# Get observations for all usms, but only banana has observations:
Meas <- get_obs(path)
# Get observations only for banana:
Meas_banana <- get_obs(path, "banana")
}

}
