% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs.R
\name{get_obs}
\alias{get_obs}
\title{Read STICS observation files (*.obs)}
\usage{
get_obs(
  workspace = getwd(),
  usm_name = NULL,
  usms_filename = NULL,
  javastics_path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{workspace}{Path of a JavaStics workspace, or a vector of (recursive call).}

\item{usm_name}{Vector of usms to read (optional, used to filter usms)}

\item{usms_filename}{Name of the usm file (optional)}

\item{javastics_path}{JavaStics installation path (optional, needed if the plant files are not in the \code{workspace}
but rather in the JavaStics default workspace). Only used to get the plants names.}

\item{verbose}{Logical value (optional), TRUE to display infos on error, FALSE otherwise (default)}
}
\value{
A named list of \code{data.frame}s with observations data. Names of the list elements may be either usms names,
or observation file name if the \code{usms_filename} is not specified.
}
\description{
Read STICS observation files from a JavaStics workspace and store data into a list per usm
}
\details{
The \code{.obs} files names should match USMs mames, e.g. for a usm called "banana",
the \code{.obs} file should be named \code{banana.obs}. For intercrops, the name should be suffixed by "p" for
the principal and "a" for the associated plant.
If \code{usm_name} is not specified (or equal to \code{NULL}), the
function reads the observation files from all usms listed in \code{usms_filename} and present in the \code{workspace} folder(s).
Or if \code{usms_filename} is not specified, all the .obs files are loaded from the \code{workspace} folder(s).
}
\examples{
\dontrun{
path <- file.path(get_examples_path(file_type = "obs"),"simple_exemple")

# Get observations for all usms, but only banana has observations:
Meas <- get_obs(path)
# Get observations only for banana:
Meas_banana <- get_obs(path, "banana")

# Get oservations with real plant names when plant folder is not in the workspace:
get_obs(path, "banana", javastics_path= "path/to/javastics")
}

}
