% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs_int.R
\name{get_obs_int}
\alias{get_obs_int}
\title{Read STICS observation file (*.obs)}
\usage{
get_obs_int(workspace, filename, plant_name = NULL, verbose = TRUE)
}
\arguments{
\item{workspace}{A directory where to read the files}

\item{filename}{A vector of observation file names}

\item{plant_name}{A vector of plant names. Length of filename, or recycled if one (optional).}

\item{verbose}{Logical value (optional), TRUE to display read file name, FALSE otherwise (default)}
}
\value{
A \code{data.frame} of all observation files concatenated by row.
}
\description{
Read STICS observation file(s). This function is not exported because
it only read and pre-formate an obs file without any clever tests for intercrops or usms.
}
\examples{
\dontrun{
path <- file.path(get_examples_path( file_type = "obs"))
get_obs_int(path, "banana.obs")
}
}
\keyword{internal}
