% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_options_names.R
\name{get_options_names}
\alias{get_options_names}
\title{Getting options parameter names from an xml parameter file}
\usage{
get_options_names(xml_file_path, option_names = NULL)
}
\arguments{
\item{xml_file_path}{path of xml parameter file}

\item{option_names}{options names vector (optional)}
}
\value{
A vector of strings of options parameter names
}
\description{
Extracting options parameter name from an xml file data, and checking
names if provided in options_names
}
\examples{
\dontrun{

xml_path <- file.path(get_examples_path(file_type = "xml"), "file_plt.xml")

SticsRFiles:::get_options_names(xml_path)

SticsRFiles:::get_options_names(xml_path, c("codemonocot", "codlainet"))
}

}
\keyword{internal}
