% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_bounds.R
\name{get_param_bounds}
\alias{get_param_bounds}
\title{Getting parameters bounds from an xmlDocument object}
\usage{
get_param_bounds(xml_doc, param_name, bounds_name = NULL)
}
\arguments{
\item{xml_doc}{an xmlDocument class object}

\item{param_name}{a parameter name of a vector of parameters names}

\item{bounds_name}{bounds name "min" or "max" (optional, default value c("min","max ))}
}
\value{
A list of parameters bounds values
}
\description{
Extracting parameters min and/or max bounds for a parameter
or a vector of parameters from an xmlDocument class object.
}
\examples{
\dontrun{
library(SticsRFiles)
xml_doc <- xmldocument("path/to/xmlfile")
param_bounds <- get_param_bounds(xml_doc, "param", "min")

param_bounds <- get_param_bounds(xml_doc, c("param1", "param2"), "max")

param_bounds <- get_param_bounds(xml_doc, c("param1", "param2"))
}


}
