% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_info.R
\name{get_param_data_df}
\alias{get_param_data_df}
\title{Getting parameters information using partial search words}
\usage{
get_param_data_df(
  param = NULL,
  file = NULL,
  stics_version = "latest",
  kind = "all",
  exact = FALSE
)
}
\arguments{
\item{param}{Optional name or partial name or a vector of}

\item{file}{Optional, xml file path or a vector of}

\item{stics_version}{An optional version name as listed in
get_stics_versions_compat() return}

\item{kind}{Kind of information to be retrieved for parameters
among "parameter", "formalism" or "all" for both of them}

\item{exact}{Logical, if TRUE, the exact name is searched}
}
\value{
A data.frame containing parameters names and
their origin (file name), and their bounds or the formalism they belong to,
or both of them.
}
\description{
Getting parameters information using partial search words
}
\details{
If not any name vector of names, information are extracted for
all the existing parameters
}
\examples{
\dontrun{

get_param_data_df(param = "albedo")

get_param_data_df(param = "albedo", file = "/path/to/file.xml")

get_param_data_df(param = "albedo", kind = "formalism")

get_param_data_df(param = "albedo", stics_version = "V9.0")

get_param_data_df(param = c("albedo", "latitude", "humcapil"))

get_param_data_df(
  param = c("albedo", "latitude", "humcapil"),
  kind = "formalism"
)

get_param_data_df( file = "/path/to/JavaSTICS/folder/config/inputs.csv")

get_param_data_df( param = "albedo",
   file = "/path/to/JavaSTICS/folder/config/inputs.csv")

}

}
\keyword{internal}
