% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_desc.R
\name{get_param_desc}
\alias{get_param_desc}
\title{Getting parameters description}
\usage{
get_param_desc(
  file_path = NULL,
  stics_version = "latest",
  name = NULL,
  kind = FALSE
)
}
\arguments{
\item{file_path}{csv file path}

\item{stics_version}{The STICS version. See \code{get_stics_versions_compat()}
to get all compatible versions. Default
to "latest", a special code to get the latest version.}

\item{name}{a name vector for selecting loaded content using
name column matching}

\item{kind}{a name vector for selecting loaded content using
kind column matching}
}
\description{
Getting parameters description
}
\keyword{internal}
