% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_formalism.R
\name{get_param_formalism}
\alias{get_param_formalism}
\title{Get the "formalisme" node "@nom" attribute value of a
(or a list of) parameter(s) in an xmlDocument}
\usage{
get_param_formalism(xml_document, param_name)
}
\arguments{
\item{xml_document}{an xmlDocument object (created from any xml file)}

\item{param_name}{a formalism name or a vector of}
}
\value{
a formalism name or a list of
}
\description{
Get the "formalisme" node "@nom" attribute value of a
(or a list of) parameter(s) in an xmlDocument
}
\examples{
\dontrun{

xml_sta <- system.file(paste0("extdata/xml/examples/V9.0/file_sta.xml"),
package = "SticsRFiles")

sta_doc <- SticsRFiles:::xmldocument(xml_sta)

par_form <- SticsRFiles:::get_param_formalism(sta_doc,"zr")

par_form_list <- SticsRFiles:::get_param_formalism(sta_doc,c("zr","altistation"))

}


}
