% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_info.R
\name{get_param_info}
\alias{get_param_info}
\title{Finding parameters information using partial search words}
\usage{
get_param_info(name = NULL, version = NULL, kind = "all", exact = FALSE)
}
\arguments{
\item{name}{Optional name or partial name or a vector of}

\item{version}{Optional, Stics version.
Only the 2 last are referenced: V9.0, V9.1 (default value)}

\item{kind}{Kind of information to be retrieved for parameters
among "parameter", "formalism" or "all" for both of them}

\item{exact}{Logical, if TRUE, the exact name is searched}
}
\value{
A data.frame containing parameters names and
their origin (file name), and their bounds or the formalism they belong to,
or both of them.
}
\description{
Finding parameters information using partial search words
}
\details{
If not any name vector of names, information are extracted for
all the existing parameters
}
\examples{
\dontrun{

get_param_info(name = "albedo")

get_param_info(name = "albedo", kind = "formalism)

get_param_info(name = "albedo", version = "V9.0")

get_param_info(name = c("albedo", "latitude", "humcapil"))

get_param_info(name = c("albedo", "latitude", "humcapil"),
kind = "formalism)

}


}
