% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_names.R
\name{get_param_names}
\alias{get_param_names}
\title{Get a list of Stics xml parameters names an xmlDocument or  XML node}
\usage{
get_param_names(xml_object, param_list = c(), full_list = FALSE)
}
\arguments{
\item{xml_object}{an xml XMLInternalElementNode or xmlDocument object}

\item{param_list}{param names vector, only used for recursive calls}

\item{full_list}{TRUE for getting all names, FALSE otherwise (default)
for unique names list}
}
\value{
a character vector of parameters names
}
\description{
Get a list of Stics xml parameters names an xmlDocument or  XML node
}
\examples{
\dontrun{
xml_path = system.file("extdata/xml/examples/V9.0/sols.xml", package = "SticsRFiles")
sols_doc <- SticsRFiles:::xmldocument(xml_path)
SticsRFiles:::get_param_names(sols_doc)
}

}
\keyword{internal}
