% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_txt.R
\name{get_param_txt}
\alias{get_param_txt}
\title{Read STICS input parameters from text files}
\usage{
get_param_txt(dirpath = getwd(), param = NULL, ...)
}
\arguments{
\item{dirpath}{USM directory path}

\item{param}{Parameter name. Optional, if not provided, the function
return an object with all parameters}

\item{...}{Helper to pass arguments from \code{\link{get_param_txt}} to the
other functions}

\item{filepath}{File path}

\item{several_fert}{Is there several fertilization in the USM ?}

\item{several_thin}{Is there several thinning in the USM ?}

\item{is_pasture}{Is the plant a pasture ?}

\item{variety}{Integer. The plant variety to get the parameter from.}
}
\value{
A list of all parameters:
\item{ini}{Initialization parameters}
\item{general}{General parameters}
\item{tec}{Technical parameters}
\item{plant}{Plant parameters}
\item{soil}{Soil parameters}
\item{station}{Station parameters}
\item{output}{The output variables the user require from STICS}
The function can return several sub-lists for each \code{plant} and \code{tec}
if mixed crops, numbered by usage
}
\description{
Read STICS model input parameters from a pre-existing STICS input
file. Generally used after calling building a usm with \code{JavaStics}.
}
\note{
Users generally only use \code{get_param_txt} that identify parameters for
other functions and call them.
}
\examples{
\dontrun{
# Read the interrow distance parameter:

library(SticsRFiles)
path = system.file("extdata/txt/V9.1", package = "SticsRFiles")
get_param_txt(path, param='interrang')

}

}
\seealso{
\code{\link{set_param_txt}}.
}
