% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_txt.R
\name{get_param_txt}
\alias{get_param_txt}
\alias{get_ini_txt}
\alias{get_general_txt}
\alias{get_tmp_txt}
\alias{get_plant_txt}
\alias{get_tec_txt}
\alias{get_soil_txt}
\alias{get_station_txt}
\alias{get_usm_txt}
\alias{get_out_var_txt}
\title{Read STICS input parameters}
\usage{
get_param_txt(dirpath = getwd(), param = NULL, ...)

get_ini_txt(filepath = "ficini.txt")

get_general_txt(filepath = "tempopar.sti")

get_tmp_txt(filepath = "tempoparv6.sti")

get_plant_txt(filepath = "ficplt1.txt", variety = NULL, max_variety = 30)

get_tec_txt(
  filepath = "fictec1.txt",
  several_fert = T,
  several_thin = T,
  is_pasture = F
)

get_soil_txt(filepath = "param.sol")

get_station_txt(filepath = "station.txt")

get_usm_txt(filepath = "new_travail.usm")

get_out_var_txt(filepath = "var.mod")
}
\arguments{
\item{dirpath}{Directory path}

\item{param}{Parameter name. Optional, if not provided, the function
return an object with all parameters}

\item{...}{Helper to pass arguments from \code{\link{get_param_txt}} to the
other functions}

\item{filepath}{File path}

\item{variety}{Integer. The plant variety to get the parameter from.}

\item{max_variety}{Maximum number of variety authorized (this is only for STICS
compatibility)}

\item{several_fert}{Is there several fertilization in the USM ?}

\item{several_thin}{Is there several thinning in the USM ?}

\item{is_pasture}{Is the plant a pasture ?}
}
\value{
A list of all parameters:
\item{ini}{Initialization parameters}
\item{general}{General parameters}
\item{tec}{Technical parameters}
\item{plant}{Plant parameters}
\item{soil}{Soil parameters}
\item{station}{Station parameters}
\item{output}{The output variables the user require from STICS}
The function can return several sub-lists for each \code{plant} and \code{tec}
if mixed crops, numbered by usage
}
\description{
Read STICS model input parameters from a pre-existing STICS input
file. Generally used after calling \code{\link{set_usm_txt}}.
}
\note{
Users generally only use \code{get_param_txt} that identify parameters for
other functions and call them.
}
\examples{
\dontrun{
# Read the interrow distance parameter:

library(SticsRFiles)
path = system.file("extdata/xml/examples/V9.0", package = "SticsRFiles")
get_param_txt(param='interrang')

}

}
\seealso{
\code{\link{set_param_txt}}.
}
