% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_type.R
\name{get_param_type}
\alias{get_param_type}
\title{Get a Stics xml parameter type and xpath from a request in an xmlDocument}
\usage{
get_param_type(xml_doc, param_name, parent_name = NULL,
  parent_sel_attr = NULL, id = NULL)
}
\arguments{
\item{xml_doc}{an xmlDocument object (created from an ini file)}

\item{param_name}{parameter name}

\item{parent_name}{xml parent node name}

\item{parent_sel_attr}{parent attribute value for selecting}

\item{id}{node identifier(s) (if multiple nodes) for selecting}
}
\value{
a list with fields $type, $xpath and $length
}
\description{
Get a Stics xml parameter type and xpath from a request in an xmlDocument
}
