% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_value.R
\name{get_param_value}
\alias{get_param_value}
\title{Get a set of a (or a list of) Stics xml parameter(s) values from a request}
\usage{
get_param_value(
  xml_doc,
  param_name,
  parent_name = NULL,
  parent_sel_attr = NULL,
  ids = NULL,
  show_xpath = FALSE
)
}
\arguments{
\item{xml_doc}{an xmlDocument object}

\item{param_name}{parameter name or a vector of names}

\item{parent_name}{parent node name or attribute name (optional)}

\item{parent_sel_attr}{parent attribute value (optional)}

\item{ids}{elements indices (optional)}

\item{show_xpath}{Print the xpath}
}
\value{
a numeric vector values of parameter or a list of
}
\description{
Extracting parameter value from an xml document object
}
\examples{
\dontrun{
xml_path = system.file("extdata/xml/examples/V9.0/sols.xml", package = "SticsRFiles")
sols_doc <- SticsRFiles:::xmldocument(xml_path)
SticsRFiles:::get_param_value(sols_doc, "argi")
SticsRFiles:::get_param_value(sols_doc, c("argi", "norg"))

SticsRFiles:::get_param_value(sols_doc, "argi",
parent_name = "sol", parent_sel_attr = "solcanne"))

SticsRFiles:::get_param_value(sols_doc, c("argi", "norg"),
parent_name = "sol", parent_sel_attr = c("solcanne", "solbanane"))

SticsRFiles:::get_param_value(list(sols_doc, sols_doc), c("argi","norg"),
parent_name = "sol", parent_sel_attr = c("solcanne","solbanane"))

}

}
\keyword{internal}
