% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_xml.R
\name{get_param_xml}
\alias{get_param_xml}
\title{Getting parameter values from xml files}
\usage{
get_param_xml(xml_files, param_names, select = NULL, value = NULL, ...)
}
\arguments{
\item{xml_files}{an xml file, or a vector/list of}

\item{param_names}{parameter names list (i.e.: option parameter name, parameter name,
other kinds in ini and some tec parameters not taken into account for the moment)}

\item{select}{node name or attribute name name to use for selection  (optional)}

\item{value}{value used for select (optional)}

\item{...}{To pass some other arguments}
}
\value{
A list of parameter list length, with parameters values for each xml file
}
\description{
Extracting parameter values for a list of xmlDocument and
of parameters
}
\examples{
\dontrun{
xml_path = system.file("extdata/xml/examples/V9.1/sols.xml", package = "SticsRFiles")
get_param_xml(xml_path, "argi")
get_param_xml(xml_path, c("argi", "norg"))

get_param_xml(xml_path, "argi",
select = "sol", value = "solcanne")

get_param_xml(xml_path, c("argi", "norg"),
select = "sol", value = c("solcanne", "solbanane"))

get_param_xml(list(xml_path, xml_path), c("argi","norg"),
select = "sol", value = c("solcanne","solbanane"))

}

}
