% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_xml.R
\name{get_param_xml}
\alias{get_param_xml}
\title{Getting parameter values from xml files}
\usage{
get_param_xml(
  xml_files,
  param_names_list,
  parent_name = NULL,
  parent_sel_attr = NULL,
  show_xpath = FALSE
)
}
\arguments{
\item{xml_files}{an xml file, or a vector/list of}

\item{param_names_list}{parameter names list (i.e.: option parameter name, parameter name,
other kinds in ini and some tec parameters not taken into account for the moment)}

\item{parent_name}{parent node name or attribute name (optional)}

\item{parent_sel_attr}{parent attribute value (optional)}

\item{show_xpath}{Logical for displaying or not xpath value}
}
\value{
A list of parameter list length, with parameters values for each xml file
}
\description{
Extracting parameter values for a list of xmlDocument and
of parameters
}
\examples{
\dontrun{
xml_path = system.file("extdata/xml/examples/V9.0/sols.xml", package = "SticsRFiles")
get_param_xml(xml_path, "argi")
get_param_xml(xml_path, c("argi", "norg"))

get_param_xml(xml_path, "argi",
parent_name = "sol", parent_sel_attr = "solcanne")

get_param_xml(xml_path, c("argi", "norg"),
parent_name = "sol", parent_sel_attr = c("solcanne", "solbanane"))

get_param_xml(list(xml_path, xml_path), c("argi","norg"),
parent_name = "sol", parent_sel_attr = c("solcanne","solbanane"))

}

}
