% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_xml.R
\name{get_param_xml}
\alias{get_param_xml}
\title{Getting parameter values from xml files}
\usage{
get_param_xml(
  xml_files,
  param_names_list,
  parent_name = NULL,
  parent_sel_attr = NULL,
  show_xpath = FALSE
)
}
\arguments{
\item{xml_files}{an xml file, or a vector/list of}

\item{param_names_list}{parameter names list (i.e.: option parameter name, parameter name,
other kinds in ini and some tec parameters not taken into account for the moment)}

\item{parent_name}{parent node name or attribute name (optional)}

\item{parent_sel_attr}{parent attribute value (optional)}

\item{show_xpath}{Logical for displaying or not xpath value}
}
\value{
A list of parameter list length, with parameters values for each xml file
}
\description{
Extracting parameter values for a list of xmlDocument and
of parameters
}
\examples{
\dontrun{
get_param_xml(path/to/xml/file,"codetemp")
get_param_xml(c(path/to/xml/file1,path/to/xml/file2),"codetemp")
get_param_xml(path/to/xml/file,c("codetemp","codegdh"))
get_param_xml(c(path/to/xml/file1,path/to/xml/file2),c("codetemp","codegdh"))
}

}
