% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_xml.R
\name{get_param_xml}
\alias{get_param_xml}
\title{Getting parameter values from xml files}
\usage{
get_param_xml(xml_file, param_name = NULL, select = NULL, value = NULL, ...)
}
\arguments{
\item{xml_file}{An xml file, or a vector/list of}

\item{param_name}{parameter names vector (i.e.: parameter name or option code, optional)}

\item{select}{node name or attribute name to use for selection (optional, default to no selection)}

\item{value}{value used for select (optional)}

\item{...}{Pass further arguments to \code{get_param_value()}}
}
\value{
A list of parameter values for each xml_file (a list of list)
}
\description{
Extracting parameter values for a list of xml files and parameters
}
\examples{
xml_path = file.path(get_examples_path( file_type = "xml"),"sols.xml")
get_param_xml(xml_path, "argi")
get_param_xml(xml_path, c("argi", "norg"))

get_param_xml(xml_path, "argi",
select = "sol", value = "solcanne")

get_param_xml(xml_path, c("argi", "norg"),
select = "sol", value = c("solcanne", "solbanane"))

get_param_xml(list(xml_path, xml_path), c("argi","norg"),
select = "sol", value = c("solcanne","solbanane"))

}
