% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params_from_doc.R
\name{get_params_from_doc}
\alias{get_params_from_doc}
\title{Get a list of Stics xml parameters names in an xml_document}
\usage{
get_params_from_doc(xml_doc_object, type_name = NULL, unique_val = TRUE)
}
\arguments{
\item{xml_doc_object}{an xml_document object (created from an xml file)}

\item{type_name}{type name, one of "option", "param", "colonne"}

\item{unique_val}{logical, TRUE to get unique names list, FALSE otherwise}
}
\value{
a character vector of parameters names
}
\description{
Get a list of Stics xml parameters names in an xml_document
}
\keyword{internal}
