% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params_from_table.R
\name{get_params_from_table}
\alias{get_params_from_table}
\title{Get a list of Stics xml parameters names and values from a table (data.frame, tibble)}
\usage{
get_params_from_table(
  params_table,
  param_names = NULL,
  xml_doc = NULL,
  lines_id = NULL,
  stopping = FALSE,
  dict = NULL
)
}
\arguments{
\item{params_table}{a table (df, tibble) containing parameters to use}

\item{param_names}{parameters names to select}

\item{xml_doc}{an xmlDocument object (of any xml type)}

\item{lines_id}{table lines identifiers to select}

\item{stopping}{logical value for stopping if any unknown parameters (if TRUE)}

\item{dict}{List of names correspondence between short names (tags) and real parameters names}
}
\value{
a named list (with param names as list names) of data.frame/tibble
}
\description{
Get a list of Stics xml parameters names and values from a table (data.frame, tibble)
}
