% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params_names.R
\name{get_params_names}
\alias{get_params_names}
\title{Get a list of Stics xml parameters names an xmlDocument node}
\usage{
get_params_names(xml_node, param_list = c())
}
\arguments{
\item{xml_node}{an xml XMLInternalElementNode}

\item{param_list}{param names vector, used for recursive calls}
}
\value{
a character vector of parameters names
}
\description{
Get a list of Stics xml parameters names an xmlDocument node
}
