% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params_names.R
\name{get_params_names}
\alias{get_params_names}
\title{Get a list of Stics xml parameters names an xmlDocument or  XML node}
\usage{
get_params_names(xml_object, param_list = c(), full_list = FALSE)
}
\arguments{
\item{xml_object}{an xml XMLInternalElementNode or xmlDocument object}

\item{param_list}{param names vector, used for recursive calls}

\item{full_list}{TRUE or getting all names, FALSE otherwise (default)
for unique names list}
}
\value{
a character vector of parameters names
}
\description{
Get a list of Stics xml parameters names an xmlDocument or  XML node
}
