% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plant_name.R
\name{get_plant_files}
\alias{get_plant_files}
\title{Getting xml plant files by usm from an usms.xml file}
\usage{
get_plant_files(usms_file, usms = NULL)
}
\arguments{
\item{usms_file}{usms.xml file path}

\item{usms}{usms names to get, if NULL all teh usms are taken into account}
}
\value{
A named list with usm names containing plant files names
}
\description{
Getting xml plant files by usm from an usms.xml file
}
\examples{
\dontrun{
get_plant_files("path/to/usms.xml")

get_plant_files("path/to/usms.xml", c("usm1", "usm2")
}

}
\keyword{internal}
