% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plants_nb.R
\name{get_plants_nb}
\alias{get_plants_nb}
\title{Getting plants number per usm, all or selected from a given list}
\usage{
get_plants_nb(usm_xml_path, usms_list = c())
}
\arguments{
\item{usm_xml_path}{Path of usms.xml file}

\item{usms_list}{Usms selection list inside usms from usms.xml file (optional)}
}
\value{
A numeric vector of plants number per usm
}
\description{
Extracting plant number from usms.xml file data
}
\examples{
\dontrun{
xml_usms <- system.file(paste0("extdata/xml/examples/V9.0/usms.xml"),
package = "SticsRFiles")
get_plants_nb(xml_usms)
get_plants_nb(xml_usms,"wheat")
get_plants_nb(xml_usms,c("wheat","intercrop_pea_barley"))
}

}
