% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plants_nb.R
\name{get_plants_nb}
\alias{get_plants_nb}
\title{Getting plants number per usm, all or selected from a given list}
\usage{
get_plants_nb(usm_xml_path, usms_list = c())
}
\arguments{
\item{usm_xml_path}{Path of usms.xml file}

\item{usms_list}{Usms selection list inside usms from usms.xml file (optional)}
}
\description{
Extracting plant number from usms.xml file data
}
\examples{
\dontrun{
get_plants_nb("path/to/USM")
get_plants_nb("path/to/USM","ble")
get_plants_nb("path/to/USM",c("ble","mais"))
}

}
\keyword{internal}
