% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_script.R
\name{get_script}
\alias{get_script}
\title{Get script(s) example(s) for generating xml Stics files from tables (data.frame, tibble)}
\usage{
get_script(in_name = NULL, dest_dir = getwd())
}
\arguments{
\item{in_name}{name of the script or partial match, or "all" for getting all scripts}

\item{dest_dir}{directory path where to copy the script}
}
\value{
logical value TRUE if the file copy is successful, FALSE otherwise
}
\description{
Get script(s) example(s) for generating xml Stics files from tables (data.frame, tibble)
}
