% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim.R
\name{get_sim}
\alias{get_sim}
\title{Load and format Stics daily output file(s)}
\usage{
get_sim(
  workspace,
  usm_name = NULL,
  var_list = NULL,
  dates_list = NULL,
  usms_filepath = NULL,
  javastics_path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{workspace}{Path of a JavaStics workspace with the \code{mod_s*.sti} files, or a vector of (recursive call).}

\item{usm_name}{Vector of usms to read (optional, used to filter usms)}

\item{var_list}{(optional) vector of output variables names to filter (see \code{get_var_info()} to get variables names)}

\item{dates_list}{(optional) list of dates to filter (POSIX date)}

\item{usms_filepath}{(optional) usms file path (\emph{e.g.} "usms.xml"). Used to get the real plant name in the Plant column.}

\item{javastics_path}{(optional) JavaStics installation path Needed if the plant files are not in the \code{workspace}
but rather in the JavaStics default workspace. Only used to get the plants names.}

\item{verbose}{(optional) Logical value, \code{TRUE} to display infos on error, \code{FALSE} otherwise (default)}
}
\value{
A list, where each element is a \code{data.frame}s of simulation results for the given usm. The list is named
after the USM name. Intercrops are returned in a single \code{data.frame}, and are identified
using either the "Plant" or "Dominance" columns.
}
\description{
Reads and format daily output file(s) (mod_s*.sti) for usm(s) with
possible selection on variable names, cumulative DOY and dates
}
\details{
If \code{usm_name} is not specified (or equal to \code{NULL}), the
function reads the files from all usms in the \code{workspace}(s).
If \code{usms_filepath} is not specified, the plants are named "plant_1" by default (+ "plant_2" for
intercrops).
}
\examples{
\dontrun{
path <- get_examples_path(file_type = "sti")
get_sim(path,"banana")
}

}
