% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soils_list.R
\name{get_soils_list}
\alias{get_soils_list}
\title{Get the soil names for an usms.xml file}
\usage{
get_soils_list(file_path, name = NULL)
}
\arguments{
\item{file_path}{The file path, either an usm file (e.g. "usms.xml"), or a soil file (e.g. "sols.xml")}

\item{name}{A vector of soil name(s) or partial name(s), optional}
}
\value{
A vector of soil names
}
\description{
Extracts the soil names from a "usms.xml" file, or from a soil file
}
\details{
The file given as the \code{file_path} is either a "usms" file type to get all the soils used in a particular USM,
or a soil file type ("sols") to get all soil types available in a soil file.
}
\examples{
path = get_examples_path( file_type = "xml" )

# Read from a usms file (soils used in a USM):
soil_list <- get_soils_list(file_path = file.path(path, "usms.xml"))

# Read from a soil file (all soil types available in a soil file)
soil_list <- get_soils_list(file_path = file.path(path, "sols.xml"))

soil_list <- get_soils_list(file_path = file.path(path, "usms.xml"), name = c("solcanne", "sole"))

}
