% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stics_versions_compat.R
\name{get_stics_versions_compat}
\alias{get_stics_versions_compat}
\title{Get the compatible stics versions}
\usage{
get_stics_versions_compat()
}
\value{
A named list with the STICS versions compatible with this package ($versions_list),
and the last version in use ($last_version).
}
\description{
Get the versions of stics that are fully compatible with this package.
}
\examples{
get_stics_versions_compat()


}
