% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_txt.R
\name{get_txt_generic}
\alias{get_txt_generic}
\title{Read parameter values from file}
\usage{
get_txt_generic(file, names = TRUE)
}
\arguments{
\item{file}{Path (including name) of the file to read}

\item{names}{Boolean, read the parameter names ?}

\item{filepath}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{filepath} is no
longer supported, use \code{file} instead.}
}
\value{
A named (if names=TRUE) list of parameter values
}
\description{
Generic function to read STICS parameter files
}
\examples{
\dontrun{
path <- file.path(get_examples_path(file_type = "txt", stics_version = "V8.5"), "station.txt")
get_txt_generic(path)
}

}
\keyword{internal}
