% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_txt.R
\name{get_txt_generic}
\alias{get_txt_generic}
\title{Read parameter values from file}
\usage{
get_txt_generic(filepath, names = TRUE)
}
\arguments{
\item{filepath}{The path to the parameter file}

\item{names}{Boolean, read the parameter names ?}
}
\value{
A named (if names=TRUE) list of parameter values
}
\description{
Generic function to read STICS parameter files
}
\examples{
\dontrun{
path <- file.path(get_examples_path(file_type = "txt", version_name = "V8.5"), "station.txt")
get_txt_generic(path)
}

}
\keyword{internal}
