% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_txt_param_value.R
\name{get_txt_param_value}
\alias{get_txt_param_value}
\title{Getting parameters values from Stics text files}
\usage{
get_txt_param_value(
  file_path,
  param_names = NULL,
  file_type = "name_value",
  names_dict = NULL
)
}
\arguments{
\item{file_path}{file path}

\item{param_names}{character vector of parameters names}

\item{file_type}{key word designing a file structure type (optional)}

\item{names_dict}{A list containining for a given field name the corresponding
Stics parameter name (optional)}
}
\value{
a named list of parameters values
}
\description{
Extracting parameter value from a text input file for a given list
of parameters names or all values
}
