% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usms_files.R
\name{get_usms_files}
\alias{get_usms_files}
\title{Getting existing xml files path list per usm from an usms.xml file}
\usage{
get_usms_files(
  workspace_path,
  usms_list = NULL,
  file_name = "usms.xml",
  file_type = NULL,
  javastics_path = NULL
)
}
\arguments{
\item{workspace_path}{JavaStics workspace path}

\item{usms_list}{Vector of usms names (Optional)}

\item{file_name}{Usms XML file name (Optional)}

\item{file_type}{Type of file to get (see Types definition)}

\item{javastics_path}{JavaStics installation path (Optional)}
}
\value{
A named list with existing files path in each usm element
}
\description{
Getting existing xml files path list per usm from an usms.xml file
}
\examples{

\dontrun{

get_usms_files(workspace_path = "/path/to/workspace",
javastics_path = "/path/to/javastics")

get_usms_files(workspace_path = "/path/to/workspace",
javastics_path = "/path/to/javastics", usm_list = c("usm1", "usm3"))

get_usms_files(workspace_path = "/path/to/workspace",
file_type = c("finit", "ftec" ))


}

}
