% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usms_list.R
\name{get_usms_list}
\alias{get_usms_list}
\title{Getting usms names list}
\usage{
get_usms_list(usms_path, xml_name = "usms.xml")
}
\arguments{
\item{usms_path}{usms dir path or usm file path}

\item{xml_name}{usms xml file name (optional, default = "usms.xml")}
}
\value{
A list with usms names
}
\description{
Extracting a usm names list from an usms.xml file
}
\examples{
usms_list <- get_usms_list("/home/plecharpent/Work/JavaSTICS-v131-stics-v841/example")
}
