% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values_from_table.R
\name{get_values_from_table}
\alias{get_values_from_table}
\title{Get a list of Stics xml parameters values from a table (data.frame, tibble)}
\usage{
get_values_from_table(params_table, param_names = NULL,
  lines_id = NULL)
}
\arguments{
\item{params_table}{a table (df, tibble) containing parameters to use}

\item{param_names}{parameters names}

\item{lines_id}{lines ids for selecting}
}
\value{
a named list (with param names as list names) of data.frame/tibble
}
\description{
Get a list of Stics xml parameters values from a table (data.frame, tibble)
}
