% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_out_var.R
\name{get_var_info}
\alias{get_var_info}
\title{Find output variable names for STICS}
\usage{
get_var_info(var = NULL, keyword = NULL, version = "last")
}
\arguments{
\item{var}{Character vector with a (partial) STICS output variable name}

\item{keyword}{Search by keyword instead of the variable name
(search in the name and the description fields)}

\item{version}{The stics version. See \code{get_stics_versions_compat()} to get all compatible versions. Default
to "last", a special code to get the last version.}
}
\description{
Helper function that return the name used as input
for the STICS model with a partial match.
}
\details{
The function understand \code{\link[base]{regex}} as input.
}
\examples{
\dontrun{
# Find by variable name (fuzzy search):
SticsRFiles::get_var_info("lai")

# Find by keyword (fuzzy search in variable name and description):
SticsRFiles::get_var_info(keyword= "lai")

# Find for a particular version:
SticsRFiles::get_var_info("lai", version= "V9.0")
}

}
\seealso{
\code{\link{all_out_var}}
}
