% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_mod.R
\name{get_var_mod}
\alias{get_var_mod}
\title{Get desired STICS outputs}
\usage{
get_var_mod(workspace, file_name = "var.mod")
}
\arguments{
\item{workspace}{Stics or JavaStics workspace path}

\item{file_name}{file name to generate (default value: "var.mod")}
}
\value{
The variables that will be returned by STICS
}
\description{
Get the variables STICS will return ("var.mod" file)
}
\examples{
get_var_mod(system.file("extdata/txt/V8.5", package = "SticsRFiles"))

}
\seealso{
\code{gen_varmod}
}
