% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_varmod.R
\name{get_var_mod}
\alias{get_var_mod}
\title{Get desired STICS outputs}
\usage{
get_var_mod(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated to homogenised writing process
}
\examples{
 \dontrun{
get_var_mod(workspace,usms_file)
# ->
get_varmod(workspace,usms_file)
}

}
\keyword{internal}
