% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_mod.R
\name{get_varmod}
\alias{get_varmod}
\title{Get desired STICS outputs}
\usage{
get_varmod(workspace, file_name = "var.mod")
}
\arguments{
\item{workspace}{Path of the directory containing the Stics var.mod file to modify}

\item{file_name}{file name to generate (default value: "var.mod")}
}
\value{
The variables that will be returned by STICS
}
\description{
Get the variables STICS will return ("var.mod" file)
}
\examples{
get_var_mod(get_examples_path( file_type = "txt" ))

}
\seealso{
\code{gen_varmod}
}
