% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stics_versions_compat.R
\name{get_version_num}
\alias{get_version_num}
\title{Getting version number from the version string}
\usage{
get_version_num(version)
}
\arguments{
\item{version}{string (i.e. "VX.Y", or "vX.Y")}
}
\value{
version number
}
\description{
Getting version number from the version string
}
\examples{
\dontrun{
SticsRFiles:::get_version_num()
}
}
\keyword{internal}
