% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stics_versions_compat.R
\name{get_version_string}
\alias{get_version_string}
\title{Getting version string from the version number}
\usage{
get_version_string(stics_version)
}
\arguments{
\item{stics_version}{numeric (i.e. X.Y)}
}
\value{
version string
}
\description{
Getting version string from the version number
}
\examples{
\dontrun{
SticsRFiles:::get_version_string()
}
}
\keyword{internal}
