% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stics_versions_compat.R
\name{get_versions_info}
\alias{get_versions_info}
\title{Getting versions data (versions strings and examples files directories list)}
\usage{
get_versions_info(version_name = NULL)
}
\arguments{
\item{version_name}{Optional version string (i.e. "VX.Y")}
}
\value{
A data.frame with versions data
}
\description{
Getting versions data (versions strings and examples files directories list)
}
\examples{
\dontrun{
get_versions_info()

versions compat  csv  obs  sti  txt           xml       xml_tmpl   xl
1     V8.5   V8.5 V8.5 BASE BASE V8.5 examples/V8.5 templates/V8.5 BASE
2     V9.0   V9.0 V9.0 BASE BASE V9.0 examples/V9.0 templates/V9.0 BASE
3     V9.1   V9.1 V9.1 BASE BASE V9.1 examples/V9.1 templates/V9.1 BASE

get_versions_info( version_name = "V8.5")

versions compat  csv  obs  sti  txt           xml       xml_tmpl   xl
1     V8.5   V8.5 V8.5 BASE BASE V8.5 examples/V8.5 templates/V8.5 BASE

}
}
\keyword{internal}
