% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xml_base_doc.R
\name{get_xml_base_doc}
\alias{get_xml_base_doc}
\title{Get an xmlDocument from a Stics xml file template}
\usage{
get_xml_base_doc(xml_type = NULL, stics_version = "latest")
}
\arguments{
\item{xml_type}{xml file type (see types returned when calling get_xml_base_doc())}

\item{stics_version}{the stics files version to use}
}
\value{
an xmlDocument object
}
\description{
Get an xmlDocument from a Stics xml file template
}
\examples{
\dontrun{
# Getting xml Stics files types list (i.e. keywords)
SticsRFiles:::get_xml_base_doc()

# Getting a soil document with one soil definition
SticsRFiles:::get_xml_base_doc("sols")

# Stics version can be provided, V9.1
# corresponds to stics_version = "latest"
# View available Stics files version
SticsRFiles:::get_xml_stics_version()
# Giving Stics version
SticsRFiles:::get_xml_base_doc("sols", stics_version = "V9.1")

}

}
\keyword{internal}
