% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xml_doc_example.R
\name{get_xml_doc_example}
\alias{get_xml_doc_example}
\title{Get an xmlDocument from a Stics xml file example}
\usage{
get_xml_doc_example(xml_name = NULL, stics_version = "latest")
}
\arguments{
\item{xml_name}{xml file name (see file names by calling get_xml_doc_example())}

\item{stics_version}{the stics files version to use}
}
\value{
An xmlDocument object
}
\description{
Get an xmlDocument from a Stics xml file example
}
\examples{
\dontrun{
# Retrieving xml examples files to get xmlDocument from
SticsRFiles:::get_xml_doc_example()

# Loading an usms.xml file
usm_doc <- SticsRFiles:::get_xml_doc_example("usms.xml")
}

}
\keyword{internal}
