% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xml_files_param_df.R
\name{get_xml_files_param_df}
\alias{get_xml_files_param_df}
\title{Extracting a data.frame of parameters values for an xml file or a set of
(whatever the xml file kind)}
\usage{
get_xml_files_param_df(file_path, param_names = NULL, wide_shape = FALSE)
}
\arguments{
\item{file_path}{A file path or a vector of STICS XML files}

\item{param_names}{vector of parameters names (optional)}

\item{wide_shape}{Optional logical for keeping the long data.frame
format FALSE, default) or converting it to a wider format one (TRUE), with
parameters as columns}
}
\value{
A single long data.frame with name, type, param, id and value
columns, value is of type character.
or a wide a data.frame with name, type, and all other columns
correspond to parameter names, with indices as suffix for multiple values
(as in _tec.xml files for water supply, N supply, ...)
}
\description{
Extracting a data.frame of parameters values for an xml file or a set of
(whatever the xml file kind)
}
\details{
An extract of returned data.frames with the 2 formats from an
usms.xml file as input
\itemize{
\item Default data.frame long format
}\tabular{lllrl}{
   name \tab type \tab param \tab id \tab value \cr
   SugarCane \tab usms \tab datedebut \tab NA \tab 286 \cr
   SugarCane \tab usms \tab datefin \tab NA \tab 650 \cr
   SugarCane \tab usms \tab finit \tab NA \tab canne_ini.xml \cr
   SugarCane \tab usms \tab nomsol \tab NA \tab solcanne \cr
   SugarCane \tab usms \tab fstation \tab NA \tab climcanj_sta.xml \cr
   ... \tab ... \tab ... \tab ... \tab ... \cr
   ... \tab ... \tab ... \tab ... \tab ... \cr
   param_gen.xml \tab fichierpar \tab Qmulchdec \tab 17 \tab 0 \cr
   param_gen.xml \tab fichierpar \tab Qmulchdec \tab 18 \tab 0 \cr
   param_gen.xml \tab fichierpar \tab Qmulchdec \tab 19 \tab 0 \cr
   param_gen.xml \tab fichierpar \tab Qmulchdec \tab 20 \tab 0 \cr
   param_gen.xml \tab fichierpar \tab Qmulchdec \tab 21 \tab 0 \cr
}


name: a file name or an usm or soil name
type: type of file
param: param name
id: NA for scalar, integer id for vectors parameters
value: parameter value
\itemize{
\item wide data.frame format
}\tabular{llrrll}{
   name \tab type \tab datedebut \tab datefin \tab finit \tab nomsol \cr
   SugarCane \tab usms \tab 286 \tab 650 \tab canne_ini.xml \tab solcanne \cr
   potato \tab usms \tab 91 \tab 250 \tab patate_ini.xml \tab solpatate \cr
   banana \tab usms \tab 30 \tab 300 \tab banane_ini.xml \tab solbanane \cr
   sorghum \tab usms \tab 112 \tab 360 \tab sorgho_ini.xml \tab solsorgho \cr
   barley \tab usms \tab 88 \tab 196 \tab orge_ini.xml \tab solorge \cr
}


name: a file name or an usm or soil name
type: type of file
all other columns names correspond to parameter names, with indices as suffix
for multiple values (as in _tec.xml files)
}
\examples{
\dontrun{
dir_path <- get_examples_path("xml")

get_xml_files_param_df(
  file_path = file.path(dir_path, "sols.xml")
)

files_list <- file.path(dir_path, c("sols.xml", "usms.xml", "param_gen.xml"))
# long format (default)
get_xml_files_param_df(file_path = files_list)

# wide format
get_xml_files_param_df(file_path = files_list, wide_shape = TRUE)
}

}
