% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xml_stics_versions_tags.R
\name{get_xml_stics_versions_tags}
\alias{get_xml_stics_versions_tags}
\title{Get all stics XML versions}
\usage{
get_xml_stics_versions_tags(last = FALSE)
}
\arguments{
\item{last}{Boolean. Is only the last version needed ?}
}
\value{
All the STICS XML file versions available
}
\description{
Get all stics XML versions
}
