% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_stics_doc.R
\name{is_stics_doc}
\alias{is_stics_doc}
\alias{is_stics_ini}
\alias{is_stics_usms}
\alias{is_stics_sols}
\alias{is_stics_tec}
\alias{is_stics_sta}
\alias{is_stics_plt}
\alias{is_stics_par}
\alias{is_stics_newpar}
\title{Is it a STICS xml document}
\usage{
is_stics_doc(xml_doc, doc_type = NULL, doc_types = NULL)

is_stics_ini(xml_doc)

is_stics_usms(xml_doc)

is_stics_sols(xml_doc)

is_stics_tec(xml_doc)

is_stics_sta(xml_doc)

is_stics_plt(xml_doc)

is_stics_par(xml_doc)

is_stics_newpar(xml_doc)
}
\arguments{
\item{xml_doc}{An xml document}

\item{doc_type}{The type of xml document}

\item{doc_types}{The different possible document types (optional)}
}
\value{
A logical value giving if xml_doc is a STICS xml_document object
(TRUE, with document type as attribute "type"), or not (FALSE)
}
\description{
Test if an XML document is a STICS document.
}
\examples{
\dontrun{
xml_path <- file.path(get_examples_path(file_type = "xml"), "sols.xml")
sols_doc <- SticsRFiles:::xmldocument(xml_path)
SticsRFiles:::is_stics_doc(sols_doc)
}

}
\keyword{internal}
