% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_out_var.R
\name{is_stics_var}
\alias{is_stics_var}
\title{Search if a STICS variable exist}
\usage{
is_stics_var(var, stics_version = "latest", version = lifecycle::deprecated())
}
\arguments{
\item{var}{A vector of variable names}

\item{stics_version}{Name of the STICS version. Optional, can be used
to search parameters information relative to a specific STICS version.
By default the latest version returned by \code{get_stics_versions_compat()}
is used.}

\item{version}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{version} is no
longer supported, use \code{stics_version} instead.}
}
\value{
A boolean vector: \code{TRUE} if the variable exist, \code{FALSE} if it doesn't
}
\description{
Tells if one or more variable names are valid STICS
output variables.
}
\examples{
\dontrun{
is_stics_var(c("lai(n)", "masec(n)", "truc"))
}
}
\seealso{
\code{get_var_info()} for interactive use.
}
