% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs.R
\name{parse_mixed_obs}
\alias{parse_mixed_obs}
\title{Get mixed obs}
\usage{
parse_mixed_obs(obs_names)
}
\value{

}
\description{
Get mixed observation files by name
}
\examples{

parse_mixed_obs(list("banana.obs", "IC_banana_sorghuma.obs", "IC_banana_sorghump.obs"))

}
\keyword{internal}
