% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_params_table.R
\name{read_params_table}
\alias{read_params_table}
\title{Getting parameters data from tables files (Excel sheet, csv)}
\usage{
read_params_table(file_path, sheet_name = NULL)
}
\arguments{
\item{file_path}{Excel or csv file path}

\item{sheet_name}{Name of an Excel sheet (useless for csv files)}
}
\value{
A tibble of parameters
}
\description{
Getting parameters data from tables files (Excel sheet, csv)
}
\examples{

usm_xl_file <-  download_usm_xl(xl_name = "inputs_stics_example.xlsx")
xl_param <- read_excel(usm_xl_file, sheet = "USMs")

usm_csv_file <- download_usm_csv(csv_name = "inputs_stics_example_USMs.csv")
read_params_table(file_path = usm_csv_file)

}
