% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_node_from_doc.R
\name{remove_node_from_doc}
\alias{remove_node_from_doc}
\title{Remove node}
\usage{
remove_node_from_doc(
  xml_doc,
  param_name,
  parent_path = NULL,
  remove_parent = FALSE,
  nodes_ids = NULL
)
}
\arguments{
\item{xml_doc}{The XML document}

\item{param_name}{The parameter name}

\item{parent_path}{The path to the parent}

\item{remove_parent}{Boolean. Is the parent to be removed ?}

\item{nodes_ids}{The requested node IDs}
}
\description{
Remove a node from an XML document
}
