% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_param_txt.R
\name{set_file_txt}
\alias{set_file_txt}
\title{Internal function to set some STICS input file parameters}
\usage{
set_file_txt(
  file,
  param,
  value,
  append,
  plant_id = NULL,
  variety = NULL,
  layer = NULL,
  stics_version = "latest"
)
}
\arguments{
\item{file}{Path to the parameter file}

\item{param}{Parameter name}

\item{value}{New parameter value}

\item{append}{Boolean. Append input to existing file}

\item{plant_id}{The plant identifier (main crop: 1 ; associated crop: 2).}

\item{variety}{The plant variety to set the parameter value,
either the variety
name (\code{codevar} in the plant file) or the index
(\code{variete} in the technical file).}

\item{layer}{The soil layer if any (only concerns soil-related parameters)}

\item{stics_version}{An optional version name as listed in
get_stics_versions_compat() return}
}
\description{
Replace or set an input parameter from a pre-existing
STICS input file. This function is called by some of the
generic \code{set_*} functions under the hood.
}
\details{
The function uses \code{base::sys.call()} to know from which function
of the \code{set_*} family it is called, so it won't work properly
if called by the user directly. This is why this function
is internal.
}
\note{
This function is not used for \code{\link{set_soil_txt}}.
}
\seealso{
\code{\link{set_param_txt}}.
}
\keyword{internal}
