% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_param_txt.R
\name{set_param_txt}
\alias{set_param_txt}
\alias{set_usm_txt}
\alias{set_station_txt}
\alias{set_ini_txt}
\alias{set_general_txt}
\alias{set_tmp_txt}
\alias{set_plant_txt}
\alias{set_tec_txt}
\alias{set_soil_txt}
\title{Set (replace) STICS input file parameters}
\usage{
set_param_txt(
  dirpath = getwd(),
  param,
  value,
  add = F,
  plant = 1,
  variety = NULL
)

set_usm_txt(filepath = "new_travail.usm", param, value, add = F)

set_station_txt(filepath = "station.txt", param, value, add = F)

set_ini_txt(filepath = "ficini.txt", param, value, add = F)

set_general_txt(filepath = "tempopar.sti", param, value, add = F)

set_tmp_txt(filepath = "tempoparv6.sti", param, value, add = F)

set_plant_txt(filepath = "ficplt1.txt", param, value, add = F, variety = NULL)

set_tec_txt(filepath = "fictec1.txt", param, value, add = F)

set_soil_txt(filepath = "param.sol", param, value)
}
\arguments{
\item{dirpath}{USM directory path}

\item{param}{Parameter name}

\item{value}{New parameter value}

\item{add}{Boolean. Append input to existing file (add to the list)}

\item{plant}{Plant index. Optional, only for plant or technical parameters}

\item{variety}{Integer. The plant variety to set the parameter value.}

\item{filepath}{Path to the parameter file}
}
\description{
Replace or set an input parameter from a pre-existing STICS input
file.
}
\details{
The \code{plant} parameter can be either equal to \code{1}, \code{2} for
the associated plant in the case of intercrop, or \code{c(1,2)} for both
Principal and associated plants.
\code{\link{find_var_info}} is a helper function that returns all possible
output variables.
}
\note{
\code{set_out_var_txt} is not used by \code{set_param_txt}. To replace the output
variables required from STICS, please directly call \code{set_out_var_txt}.
}
\examples{
\dontrun{
# Replace the interrow distance parameter to 0.01:

set_param_txt(dirpath = "stics_usm/usm_1", param = "interrang", value = 0.01)

# Change the value of durvieF for the current variety:
set_param_txt(dirpath = "inst/extdata/txt/V8.5", param = "durvieF", value = 245)

# Change the value of durvieF for another variety:
set_param_txt(dirpath = "inst/extdata/txt/V8.5", param = "durvieF", variety = "Nefer", value = 178)

}

}
