% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_param_txt.R
\name{set_param_txt}
\alias{set_param_txt}
\alias{set_usm_txt}
\alias{set_station_txt}
\alias{set_ini_txt}
\alias{set_general_txt}
\alias{set_tmp_txt}
\alias{set_plant_txt}
\alias{set_tec_txt}
\alias{set_soil_txt}
\title{Set (replace) STICS input file parameters}
\usage{
set_param_txt(
  workspace = getwd(),
  param,
  value,
  add = FALSE,
  plant_id = 1,
  variety = NULL,
  layer = NULL,
  dirpath = lifecycle::deprecated(),
  plant = lifecycle::deprecated()
)

set_usm_txt(
  file = "new_travail.usm",
  param,
  value,
  add = F,
  filepath = lifecycle::deprecated()
)

set_station_txt(
  file = "station.txt",
  param,
  value,
  add = F,
  filepath = lifecycle::deprecated()
)

set_ini_txt(
  file = "ficini.txt",
  param,
  value,
  add = F,
  filepath = lifecycle::deprecated()
)

set_general_txt(
  file = "tempopar.sti",
  param,
  value,
  add = F,
  filepath = lifecycle::deprecated()
)

set_tmp_txt(
  file = "tempoparv6.sti",
  param,
  value,
  add = F,
  filepath = lifecycle::deprecated()
)

set_plant_txt(
  file = "ficplt1.txt",
  param,
  value,
  add = F,
  variety = NULL,
  filepath = lifecycle::deprecated()
)

set_tec_txt(
  file = "fictec1.txt",
  param,
  value,
  add = F,
  filepath = lifecycle::deprecated()
)

set_soil_txt(
  file = "param.sol",
  param,
  value,
  layer = NULL,
  filepath = lifecycle::deprecated()
)
}
\arguments{
\item{workspace}{Path of the workspace containing the Stics (txt) input files.}

\item{param}{Parameter name}

\item{value}{New parameter value}

\item{add}{Boolean. Append input to existing file (add to the list)}

\item{plant_id}{The plant identifier (main crop: 1 ; associated crop: 2).
Only used for plant or technical parameters.}

\item{variety}{The plant variety to set the parameter value, either the name of the variety
(\code{codevar} parameter in the plant file) or the index (\code{variete} parameter in the technical file). (optional, see details)}

\item{layer}{The soil layer if any (only concerns soil-related parameters)}

\item{dirpath}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{dirpath} is no
longer supported, use \code{workspace} instead.}

\item{plant}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{plant} is no
longer supported, use \code{plant_id} instead.}

\item{file}{Path to the parameter file}

\item{filepath}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{filepath} is no
longer supported, use \code{file} instead.}
}
\description{
Replace or set an input parameter from a pre-existing STICS input
file.
}
\details{
The \code{plant} parameter can be either equal to \code{1}, \code{2} for
the associated plant in the case of intercrop, or \code{c(1,2)} for both
Principal and associated plants.
\code{\link{get_var_info}} is a helper function that returns all possible
output variables.
If the \code{variety} is not given and if \code{param} is a varietal parameter,
the function will modify the value of \code{param} for the simulated variety, as given
in the technical file.
}
\note{
\code{gen_varmod} is not used by \code{set_param_txt}. To replace the output
variables required from STICS, please directly call \code{gen_varmod}.
}
\examples{
\dontrun{

# Getting example data path
path = get_examples_path( file_type = "txt" )

# Replace the interrow distance parameter to 0.01:
set_param_txt(workspace = path, param = "interrang", value = 0.01)

# Change the value of durvieF for the current variety:
set_param_txt(workspace = path, param = "durvieF", value = 245)

# Change the value of durvieF for another variety:
set_param_txt(workspace = path, param = "durvieF", variety = "Nefer", value = 178)

# Change the value of infil for a given layer:
set_param_txt(workspace = path, param = "infil", layer = 2, value = 60)

# If the parameter is found in several files, use the set_* functions direclty, e.g.
# cailloux is found in the general file ("codetycailloux") and the soil file. If we want to
# change its value in the soil file, we use set_soil_txt():
set_soil_txt(file = file.path(path, "param.sol"), param = "cailloux", layer = 2, value = 1)
}

}
