% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_param_value.R
\name{set_param_value}
\alias{set_param_value}
\title{Setting parameter value for different kinds of parameters}
\usage{
set_param_value(xml_doc_object, param_name, param_value,
  parent_name = NULL, parent_sel_attr = NULL, ids = NULL,
  show_xpath = FALSE)
}
\arguments{
\item{xml_doc_object}{an xmlDocument object}

\item{param_name}{parameter name}

\item{param_value}{parameter value}

\item{parent_name}{parent node name or attribute name (optional)}

\item{parent_sel_attr}{parent attribute value (optional)}

\item{ids}{elements indices (optional)}
}
\description{
Setting parameter value in a xmlDocument object
}
\examples{


}
