% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_param_value.R
\name{set_param_value}
\alias{set_param_value}
\title{Setting parameter value for different kinds of parameters}
\usage{
set_param_value(
  xml_doc,
  param_name,
  param_value,
  parent_name = NULL,
  parent_sel_attr = NULL,
  ...
)
}
\arguments{
\item{xml_doc}{an xmlDocument object}

\item{param_name}{parameter name}

\item{param_value}{vector of parameter values, or a list of}

\item{parent_name}{parent node name or attribute name (optional)}

\item{parent_sel_attr}{parent attribute value (optional)}

\item{ids}{elements indices (optional)}

\item{show_xpath}{Print the xpath}
}
\value{
A logical vector with successes status return
}
\description{
Setting parameter value in a xmlDocument object
}
\examples{
\dontrun{
xml_path = file.path(get_examples_path( file_type = "xml"),"sols.xml")
sols_doc <- SticsRFiles:::xmldocument(xml_path)
SticsRFiles:::get_param_value(sols_doc, "argi")


# setting all argi parameters with the same value
SticsRFiles:::set_param_value(sols_doc, "argi", 15)
SticsRFiles:::get_param_value(sols_doc, "argi")


# setting specific values for some soils
SticsRFiles:::set_param_value(sols_doc, "argi", c(30,35),
parent_name = "sol", parent_sel_attr = c("solcanne","solbanane"))
SticsRFiles:::get_param_value(sols_doc, "argi")


}

}
\keyword{internal}
