% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_param_xml.R
\name{set_param_xml}
\alias{set_param_xml}
\title{Setting parameter values into xml files}
\usage{
set_param_xml(
  xml_file,
  param_names,
  param_values,
  out_path = NULL,
  out_file = NULL,
  parent_name = NULL,
  parent_sel_attr = NULL,
  ids = NULL,
  show_xpath = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{xml_file}{an xml file path}

\item{param_names}{parameter names vector
(i.e.: option parameter name, parameter name,
other kinds in ini and some tec parameters not taken into account for the moment)}

\item{param_values}{parameter(s) values or a list of}

\item{out_path}{a directory path}

\item{out_file}{an xml file name}

\item{parent_name}{parent node name or attribute name (optional)}

\item{parent_sel_attr}{parent attribute value (optional)}

\item{ids}{elements indices (optional)}

\item{show_xpath}{Logical for displaying or not xpath value}

\item{overwrite}{Logical TRUE for overwriting the output file, FALSE otherwise (default)}
}
\value{
A logical value TRUE for operation success, FALSE otherwise
}
\description{
Setting parameter values for a parameter or a vector of and
with a parameters values vector
}
\examples{

\dontrun{
xml_path = system.file("extdata/xml/examples/V9.1/sols.xml", package = "SticsRFiles")
file.copy(xml_path, getwd())
get_param_xml("sols.xml", "argi")

# Setting all "argi" values to 50
set_param_xml("sols.xml", "argi", 50)

get_param_xml("sols.xml", "argi")

# Setting a specific value to "argi" for "solcanne" soil
set_param_xml("sols.xml", "argi",56,
parent_name = "sol", parent_sel_attr = "solcanne")

get_param_xml("sols.xml", "argi",
parent_name = "sol", parent_sel_attr = "solcanne")


# Setting a specific values to 2 parameters "argi" and "norg" for "solcanne" soil
set_param_xml("sols.xml", c("argi", "norg"),list(100,150),
parent_name = "sol", parent_sel_attr = "solcanne")

get_param_xml("sols.xml", c("argi", "norg"),
parent_name = "sol", parent_sel_attr = "solcanne")

}

}
