% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_param_xml.R
\name{set_param_xml}
\alias{set_param_xml}
\title{Setting parameter values into xml files}
\usage{
set_param_xml(
  xml_file,
  param_name,
  param_value,
  out_path = NULL,
  select = NULL,
  value = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{xml_file}{an xml file path}

\item{param_name}{parameter names vector, i.e.: parameter name or option code}

\item{param_value}{A vector or a list of parameter(s) values (see details).}

\item{out_path}{an xml file path (optional, default: xml_file)}

\item{select}{node name or attribute name to use for selection (optional, default to no selection)}

\item{value}{value used for select (optional)}

\item{overwrite}{Logical TRUE for overwriting the output file, FALSE otherwise (default)}

\item{...}{Pass further arguments to \code{set_param_value()}.}
}
\value{
A logical value TRUE for operation success, FALSE otherwise
}
\description{
Setting parameter values for a parameter or a vector of and
with a parameters values vector
}
\details{
It is possible to give several values for a parameter by passing a vector of values. For example
for two parameters with two values each: param_value= list(c(1,2), c(2.3,4.5))
}
\examples{

\dontrun{

# Soil file

file.copy(file.path(get_examples_path( file_type = "xml"),"sols.xml"), getwd())

# For scalar parameters per soil

# Setting all soils "argi" values to 50
set_param_xml("sols.xml", "argi", 50, overwrite = TRUE)

get_param_xml("sols.xml", "argi")

# Setting a specific value to "argi" for "solcanne" soil
set_param_xml("sols.xml", "argi",56,
select = "sol", value = "solcanne", overwrite = TRUE)

get_param_xml("sols.xml", "argi",
select = "sol", value = "solcanne")


# Setting a specific values to 2 parameters "argi" and "norg" for "solcanne" soil
set_param_xml("sols.xml", c("argi", "norg"),list(100,150),
select = "sol", value = "solcanne", overwrite = TRUE)

get_param_xml("sols.xml", c("argi", "norg"),
select = "sol", value = "solcanne")


# For vector parameters per soil (5 values, one per soil layer)
set_param_xml("sols.xml", c("epc", "HCCF"),select = "sol", value = c("solcanne", "solbanane"),
param_value = list(c(20:24,10:14),c(50:54,40:44)), overwrite = TRUE)

get_param_xml("sols.xml", c("epc", "HCCF"),select = "sol", value = c("solcanne", "solbanane"))


# Crop management file
file.copy(file.path(get_examples_path( file_type = "xml"),"file_tec.xml"), getwd())

# Modifying irrigations parameters
set_param_xml("file_tec.xml", c("julapI_or_sum_upvt", "amount"), param_value = list(200:215,20:35),
 overwrite = TRUE)

get_param_xml("file_tec.xml", c("julapI_or_sum_upvt", "amount"))

}

}
