% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_usms_param_xml.R
\name{set_usms_param_xml}
\alias{set_usms_param_xml}
\title{Setting a usm param value(s) in an usms xmlDocument}
\usage{
set_usms_param_xml(xml_doc_object, usms_param = NULL, overwrite = FALSE)
}
\arguments{
\item{xml_doc_object}{an xmlDocument object (created from an usms file)}

\item{usms_param}{usms parameters (data.frame)}

\item{overwrite}{replace existing usms (TRUE) or not,
updating existing ones (FALSE)}
}
\description{
Setting a usm param value(s) in an usms xmlDocument
}
\examples{
\dontrun{
xml_path <- file.path(get_examples_path(file_type = "xml"), "usms.xml")
usms_doc <- SticsRFiles:::xmldocument(xml_path)

xl_path <- file.path(get_examples_path(file_type = "xml"),
                     "inputs_stics_example.xlsx")
usms_df <- read_excel(xl_path, sheet = "USMs")

# For updating an existing xml doc (using existing usms names)
# Creating a fake existing_doc
existing_doc <- SticsRFiles:::gen_xml_doc("usms", nodes_nb = 3)
SticsRFiles:::set_param_value(existing_doc,
  param_name = "usm",
  param_value = usms_df$usm_nom[c(3, 1, 5)]
)

SticsRFiles:::set_usms_param_xml(existing_doc, usms_df)


# For a new xml doc
# In that case: usms_df must contain all the usms parameters !
usms_nb <- dim(usms_df)[1]
new_doc <- SticsRFiles:::gen_xml_doc("usms", nodes_nb = usms_nb)

SticsRFiles:::set_usms_param_xml(new_doc, usms_df, overwrite = TRUE)
}

}
\keyword{internal}
