% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_environment.R
\name{sticsenv_ls}
\alias{sticsenv_ls}
\title{List stics env}
\usage{
sticsenv_ls(name = NULL, env_name = sticsenv_name(), detail = FALSE)
}
\arguments{
\item{name}{Optionally, the name of an object in the stics environment}

\item{env_name}{The name of the stics environment}

\item{detail}{Return all details ? (use \code{ls.str()})}
}
\value{
A vector of all object names in the environment
}
\description{
List all objects in a stics environment
}
\keyword{internal}
