% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stics_environment.R
\name{sticsenv_name}
\alias{sticsenv_name}
\title{STICS env name}
\usage{
sticsenv_name()
}
\value{
The default STICS env name
}
\description{
Get the default STICS environment name
}
\keyword{internal}
