% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_v9_v10.R
\name{upgrade_ini_xml_9_10}
\alias{upgrade_ini_xml_9_10}
\title{Upgrade _ini.xml file(s) from STICS version 9 to 10}
\usage{
upgrade_ini_xml_9_10(file, out_dir, param_gen_file, overwrite = FALSE)
}
\arguments{
\item{file}{Path of an initialization (*_ini.xml) file or a vector of}

\item{out_dir}{Output directory path of the generated files}

\item{param_gen_file}{Path of the param_gen.xml file corresponding
to the file version}

\item{overwrite}{logical (optional),
TRUE for overwriting file if it exists, FALSE otherwise}
}
\value{
None
}
\description{
Upgrade _ini.xml file(s) from STICS version 9 to 10
}
\details{
See SticsRFiles::get_stics_versions_compat() for listing versions
}
\examples{

dir_path <- get_examples_path(file_type = "xml", stics_version = "V9.2")

upgrade_ini_xml_9_10(
  file = file.path(dir_path, "file_ini.xml"),
  out_dir = tempdir(),
  param_gen_file = file.path(dir_path, "param_gen.xml")
)

}
