% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_v10_v11.R
\name{upgrade_param_gen_xml_10_11}
\alias{upgrade_param_gen_xml_10_11}
\title{Upgrade xml param_gen file from STICS version 10 to 11}
\usage{
upgrade_param_gen_xml_10_11(
  file,
  out_dir,
  hauteur_threshold = NULL,
  par_to_net = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{file}{xml param_gen file path or a vector of}

\item{out_dir}{Output directory path}

\item{hauteur_threshold}{new parameter for V11}

\item{par_to_net}{new parameter for V11}

\item{overwrite}{Logical TRUE for overwriting files,
FALSE otherwise (default)}
}
\value{
None
}
\description{
Upgrade xml param_gen file from STICS version 10 to 11
}
