% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_workspace_xml.R
\name{upgrade_workspace_xml}
\alias{upgrade_workspace_xml}
\title{Upgrading XML files of a JavaStics workspace directory
to a newer Stics version format}
\usage{
upgrade_workspace_xml(
  workspace,
  javastics,
  out_dir,
  stics_version = "V9.2",
  target_version = "V10.0",
  plant = FALSE,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{workspace}{Path of a JavaStics workspace}

\item{javastics}{Path of JavaStics}

\item{out_dir}{Output directory of the generated files}

\item{stics_version}{Name of the Stics version (VX.Y format)}

\item{target_version}{Name of the Stics version to upgrade files to  (VX.Y format)}

\item{plant}{logical (optional), TRUE for upgrading plant files if a "plant"
sub-directory of workspace exists, FALSE otherwise}

\item{overwrite}{logical (optional),
TRUE for overwriting files if they exist, FALSE otherwise}

\item{...}{Additional input arguments}
}
\description{
Upgrading XML files of a JavaStics workspace directory
to a newer Stics version format
}
\details{
See SticsRFiles::get_stics_versions_compat() for listing versions
}
\examples{
\dontrun{
upgrade_workspace_xml(workspace = "/path/to/javastics/workspace",
                      javastics = "/path/to/javastics/directory",
                      out_dir = "/path/to/an/output/directory")
}
}
