% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_v10_v11.R
\name{upgrade_workspace_xml_10_11}
\alias{upgrade_workspace_xml_10_11}
\title{Upgrading a Javastics xml workspace}
\usage{
upgrade_workspace_xml_10_11(
  workspace,
  javastics = NULL,
  out_dir,
  from_version = "V10.0",
  target_version = "V11.0",
  overwrite = FALSE
)
}
\arguments{
\item{workspace}{JavaStics xml workspace path}

\item{javastics}{JavaStics folder path (Optional)}

\item{out_dir}{Output directry path}

\item{from_version}{Starting STICS version (character or numeric)}

\item{target_version}{Target STICS version (character or numeric)}

\item{overwrite}{Logical TRUE for overwriting files,
FALSE otherwise (default)}
}
\value{
None
}
\description{
Upgrading a Javastics xml workspace
}
