% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_to_col_names.R
\name{var_to_col_names}
\alias{var_to_col_names}
\title{Convert Stics variables names, or generated column names to
Stics specific columns names}
\usage{
var_to_col_names(var_list = c())
}
\arguments{
\item{var_list}{Variables names}
}
\value{
List of formatted column names
}
\description{
Change the `varname(n)` or `varname.n.` to `varname_n`
other names are unchanged (i.e.: varname, varname_n, ...)
}
