% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_to_col_names.R
\name{var_to_stics_name}
\alias{var_to_stics_name}
\title{Variable name to stics variable name}
\usage{
var_to_stics_name(var, version = "last")
}
\arguments{
\item{var}{Vector of variable names}

\item{version}{Stics version.}
}
\value{
A vector of variable names compatible with stics (usually used for \code{gen_varmod()})
}
\description{
Convert a variable name into a stics-compatible variable name. E.g.
\code{lai_n} into \code{lai(n)}.
}
\examples{
\dontrun{
var_to_stics_name(c("lai_n","masec_n"))
}

}
\seealso{
\code{gen_varmod()}
}
\keyword{internal}
